@@ -15,7 +15,13 @@ add_subdirectory(${SOCLE_COMMON_DIR} socle_common_lib)
 add_subdirectory(${SMITHD_DIR})
 
 if(UNIX)
-    SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g  -Wall -std=c++11")
+    IF(NOT CMAKE_BUILD_TYPE)
+        SET(CMAKE_BUILD_TYPE Debug)
+    ENDIF(NOT CMAKE_BUILD_TYPE)
+
+    SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -std=c++17 -g")
+    SET(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -g")
+    SET(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} -O3 -s -DBUILD_RELEASE")
 
     # detect Alpine - and disable backtrace_* function use
     if(EXISTS "/etc/alpine-release")
