@@ -40,106 +40,70 @@
 #include <vector>
 
 #include <cfgapi.hpp>
-#include <logger.hpp>
+#include <log/logger.hpp>
 #include <policy.hpp>
 #include <mitmproxy.hpp>
 #include <mitmhost.hpp>
 #include <cmdserver.hpp>
 
 #include <socle.hpp>
-#include <smithproxy.hpp>
+#include <main.hpp>
 
-#include <shmtable.hpp>
-
-loglevel  args_debug_flag = NON;
-
-std::string cfg_tcp_listen_port = "50080";
-std::string cfg_ssl_listen_port = "50443";
-std::string cfg_dtls_port = "50443";
-std::string cfg_udp_port = "50080";
-std::string cfg_socks_port = "1080";
-
-std::string cfg_tcp_listen_port_base = "50080";
-std::string cfg_ssl_listen_port_base = "50443";
-std::string cfg_dtls_port_base = "50443";
-std::string cfg_udp_port_base = "50080";
-std::string cfg_socks_port_base = "1080";
-
-std::string config_file;
-bool config_file_check_only = false;
-
-std::string cfg_messages_dir = "/etc/smithproxy/msg/en/";
+#include <authfactory.hpp>
 
 
-int cfg_tcp_workers   = 0;
-int cfg_ssl_workers   = 0;
-int cfg_dtls_workers  = 0;
-int cfg_udp_workers   = 0;
-int cfg_socks_workers = 0;
+CfgFactory::CfgFactory(): log(get_log()), args_debug_flag(NON), syslog_level(INF) {
 
-std::string cfg_config_file;
-std::string cfg_tenant_index;
-std::string cfg_tenant_name;
+    listen_tcp_port = "50080";
+    listen_tls_port = "50443";
+    listen_dtls_port = "50443";
+    listen_udp_port = "50080";
+    listen_socks_port = "1080";
 
-std::string cfg_syslog_server   = "";
-int         cfg_syslog_port     = 514;
-int         cfg_syslog_facility =  23; //local7
-loglevel    cfg_syslog_level    = INF;
-int         cfg_syslog_family   = 4;
+    listen_tcp_port_base = "50080";
+    listen_tls_port_base = "50443";
+    listen_dtls_port_base = "50443";
+    listen_udp_port_base = "50080";
+    listen_socks_port_base = "1080";
 
+    config_file_check_only = false;
 
-std::string cfg_log_target_base;
-std::string cfg_sslkeylog_target_base;
-
-std::string cfg_log_target;
-std::string cfg_sslkeylog_target;
-bool cfg_log_console;
+    dir_msg_templates = "/etc/smithproxy/msg/en/";
 
 
-Config cfgapi;
-time_t system_started;
+    num_workers_tcp = 0;
+    num_workers_tls = 0;
+    num_workers_dtls = 0;
+    num_workers_udp = 0;
+    num_workers_socks = 0;
 
-std::map<std::string,AddressObject*> cfgapi_obj_address;
-std::map<std::string,range> cfgapi_obj_port;
-std::map<std::string,int> cfgapi_obj_proto;
-std::vector<PolicyRule*> cfgapi_obj_policy;
-std::map<std::string,ProfileDetection*> cfgapi_obj_profile_detection;
-std::map<std::string,ProfileContent*> cfgapi_obj_profile_content;
-std::map<std::string,ProfileTls*> cfgapi_obj_profile_tls;
-std::map<std::string,ProfileAuth*> cfgapi_obj_profile_auth;
-std::map<std::string,ProfileAlgDns*> cfgapi_obj_profile_alg_dns;
-
-std::vector<int> cfgapi_obj_udp_quick_ports;
-std::vector<std::string> cfgapi_obj_nameservers;
-
-// multi-tenancy support
-std::string cfgapi_tenant_name = "default";
-unsigned int cfgapi_tenant_index = 0;
-
-//portal variables
-std::string cfg_auth_address;
-std::string cfg_auth_http;
-std::string cfg_auth_https;
-std::string cfg_auth_sslkey;
-std::string cfg_auth_sslcert;
-std::string cfgapi_tenant_magic_ip;
+    syslog_server = "";
+    syslog_port = 514;
+    syslog_facility = 23; //local7
+    syslog_family = 4;
 
 
-std::string cfg_traflog_dir = "/var/local/smithproxy/data";
-std::string cfg_traflog_file_pref = "";
-std::string cfg_traflog_file_suff = "smcap";
+    // multi-tenancy support
+    tenant_name = "default";
+    tenant_index = 0;
 
 
+    traflog_dir = "/var/local/smithproxy/data";
+    //traflog_file_prefix = "";
+    traflog_file_suffix = "smcap";
 
-std::recursive_mutex cfgapi_write_lock;
+    log_console = false;
+    ts_sys_started = std::time(nullptr);
+
+}
 
 struct cfgapi_table_ cfgapi_table;
 
-bool cfgapi_init(const char* fnm) {
+bool CfgFactory::cfgapi_init(const char* fnm) {
     
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    DIAS_("Reading config file");
+    _dia("Reading config file");
     
     // Read the file. If there is an error, report it and exit.
     try {
@@ -147,85 +111,85 @@ bool cfgapi_init(const char* fnm) {
     }
     catch(const FileIOException &fioex)
     {
-        ERR_("I/O error while reading config file: %s",fnm);
+        _err("I/O error while reading config file: %s",fnm);
         return false;   
     }
     catch(const ParseException &pex)
     {
-        ERR_("Parse error in %s at %s:%d - %s", fnm, pex.getFile(), pex.getLine(), pex.getError());
+        _err("Parse error in %s at %s:%d - %s", fnm, pex.getFile(), pex.getLine(), pex.getError());
         return false;
     }
     
-    system_started = ::time(nullptr);
+    ts_sys_started = ::time(nullptr);
     
     return true;
 }
 
-AddressObject* cfgapi_lookup_address(const char* name) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+AddressObject* CfgFactory::lookup_address (const char *name) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    if(cfgapi_obj_address.find(name) != cfgapi_obj_address.end()) {
-        return cfgapi_obj_address[name];
+    if(db_address.find(name) != db_address.end()) {
+        return db_address[name];
     }
     
     return nullptr;
 }
 
-range cfgapi_lookup_port(const char* name) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+range CfgFactory::lookup_port (const char *name) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    if(cfgapi_obj_port.find(name) != cfgapi_obj_port.end()) {
-        return cfgapi_obj_port[name];
+    if(db_port.find(name) != db_port.end()) {
+        return db_port[name];
     }    
     
     return NULLRANGE;
 }
 
-int cfgapi_lookup_proto(const char* name) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::lookup_proto (const char *name) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    if(cfgapi_obj_proto.find(name) != cfgapi_obj_proto.end()) {
-        return cfgapi_obj_proto[name];
+    if(db_proto.find(name) != db_proto.end()) {
+        return db_proto[name];
     }    
     
     return 0;
 }
 
-ProfileContent* cfgapi_lookup_profile_content(const char* name) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+ProfileContent* CfgFactory::lookup_prof_content (const char *name) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    if(cfgapi_obj_profile_content.find(name) != cfgapi_obj_profile_content.end()) {
-        return cfgapi_obj_profile_content[name];
+    if(db_prof_content.find(name) != db_prof_content.end()) {
+        return db_prof_content[name];
     }    
     
     return nullptr;
 }
 
-ProfileDetection* cfgapi_lookup_profile_detection(const char* name) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+ProfileDetection* CfgFactory::lookup_prof_detection (const char *name) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    if(cfgapi_obj_profile_detection.find(name) != cfgapi_obj_profile_detection.end()) {
-        return cfgapi_obj_profile_detection[name];
+    if(db_prof_detection.find(name) != db_prof_detection.end()) {
+        return db_prof_detection[name];
     }    
     
     return nullptr;
 }
 
-ProfileTls* cfgapi_lookup_profile_tls(const char* name) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+ProfileTls* CfgFactory::lookup_prof_tls (const char *name) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    if(cfgapi_obj_profile_tls.find(name) != cfgapi_obj_profile_tls.end()) {
-        return cfgapi_obj_profile_tls[name];
+    if(db_prof_tls.find(name) != db_prof_tls.end()) {
+        return db_prof_tls[name];
     }    
     
     return nullptr;
 }
 
-ProfileAlgDns* cfgapi_lookup_profile_alg_dns(const char* name) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+ProfileAlgDns* CfgFactory::lookup_prof_alg_dns (const char *name) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    if(cfgapi_obj_profile_alg_dns.find(name) != cfgapi_obj_profile_alg_dns.end()) {
-        return cfgapi_obj_profile_alg_dns[name];
+    if(db_prof_alg_dns.find(name) != db_prof_alg_dns.end()) {
+        return db_prof_alg_dns[name];
     }    
     
     return nullptr;
@@ -233,72 +197,72 @@ ProfileAlgDns* cfgapi_lookup_profile_alg_dns(const char* name) {
 }
 
 
-ProfileAuth* cfgapi_lookup_profile_auth(const char* name) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+ProfileAuth* CfgFactory::lookup_prof_auth (const char *name) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    if(cfgapi_obj_profile_auth.find(name) != cfgapi_obj_profile_auth.end()) {
-        return cfgapi_obj_profile_auth[name];
+    if(db_prof_auth.find(name) != db_prof_auth.end()) {
+        return db_prof_auth[name];
     }    
     
     return nullptr;
 }
 
-bool cfgapi_load_settings() {
+bool CfgFactory::load_settings () {
 
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+    std::lock_guard<std::recursive_mutex> l(lock_);
 
     if(! cfgapi.getRoot().exists("settings"))
         return false;
 
     if(cfgapi.getRoot()["settings"].exists("nameservers")) {
 
-        if(!cfgapi_obj_nameservers.empty()) {
-            DEBS_("cfgapi_load_settings: clearing existing entries in: nameservers");
-            cfgapi_obj_nameservers.clear();
+        if(!db_nameservers.empty()) {
+            _deb("load_settings: clearing existing entries in: nameservers");
+            db_nameservers.clear();
         }
 
-        int num = cfgapi.getRoot()["settings"]["nameservers"].getLength();
-        for(int i = 0; i < num; ++i) {
+        const int num = cfgapi.getRoot()["settings"]["nameservers"].getLength();
+        for(int i = 0; i < num; i++) {
             std::string ns = cfgapi.getRoot()["settings"]["nameservers"][i];
-            cfgapi_obj_nameservers.push_back(ns);
+            db_nameservers.push_back(ns);
         }
     }
 
-    cfgapi.getRoot()["settings"].lookupValue("certs_path",SSLCertStore::certs_path);
-    cfgapi.getRoot()["settings"].lookupValue("certs_ca_key_password",SSLCertStore::password);
-    cfgapi.getRoot()["settings"].lookupValue("certs_ca_path",SSLCertStore::def_cl_capath);
+    cfgapi.getRoot()["settings"].lookupValue("certs_path",SSLFactory::default_cert_path());
+    cfgapi.getRoot()["settings"].lookupValue("certs_ca_key_password",SSLFactory::default_cert_password());
+    cfgapi.getRoot()["settings"].lookupValue("certs_ca_path",SSLFactory::default_client_ca_path());
 
-    cfgapi.getRoot()["settings"].lookupValue("plaintext_port",cfg_tcp_listen_port_base); cfg_tcp_listen_port = cfg_tcp_listen_port_base;
-    cfgapi.getRoot()["settings"].lookupValue("plaintext_workers",cfg_tcp_workers);
-    cfgapi.getRoot()["settings"].lookupValue("ssl_port",cfg_ssl_listen_port_base); cfg_ssl_listen_port = cfg_ssl_listen_port_base;
-    cfgapi.getRoot()["settings"].lookupValue("ssl_workers",cfg_ssl_workers);
+    cfgapi.getRoot()["settings"].lookupValue("plaintext_port",listen_tcp_port_base); listen_tcp_port = listen_tcp_port_base;
+    cfgapi.getRoot()["settings"].lookupValue("plaintext_workers",num_workers_tcp);
+    cfgapi.getRoot()["settings"].lookupValue("ssl_port",listen_tls_port_base); listen_tls_port = listen_tls_port_base;
+    cfgapi.getRoot()["settings"].lookupValue("ssl_workers",num_workers_tls);
     cfgapi.getRoot()["settings"].lookupValue("ssl_autodetect",MitmMasterProxy::ssl_autodetect);
     cfgapi.getRoot()["settings"].lookupValue("ssl_autodetect_harder",MitmMasterProxy::ssl_autodetect_harder);
-    cfgapi.getRoot()["settings"].lookupValue("ssl_ocsp_status_ttl",SSLCertStore::ssl_ocsp_status_ttl);
-    cfgapi.getRoot()["settings"].lookupValue("ssl_crl_status_ttl",SSLCertStore::ssl_crl_status_ttl);
+    cfgapi.getRoot()["settings"].lookupValue("ssl_ocsp_status_ttl",SSLFactory::ssl_ocsp_status_ttl);
+    cfgapi.getRoot()["settings"].lookupValue("ssl_crl_status_ttl",SSLFactory::ssl_crl_status_ttl);
 
-    cfgapi.getRoot()["settings"].lookupValue("udp_port",cfg_udp_port_base); cfg_udp_port = cfg_udp_port_base;
-    cfgapi.getRoot()["settings"].lookupValue("udp_workers",cfg_udp_workers);
+    cfgapi.getRoot()["settings"].lookupValue("udp_port",listen_udp_port_base); listen_udp_port = listen_udp_port_base;
+    cfgapi.getRoot()["settings"].lookupValue("udp_workers",num_workers_udp);
 
-    cfgapi.getRoot()["settings"].lookupValue("dtls_port",cfg_dtls_port_base);  cfg_dtls_port = cfg_dtls_port_base;
-    cfgapi.getRoot()["settings"].lookupValue("dtls_workers",cfg_dtls_workers);
+    cfgapi.getRoot()["settings"].lookupValue("dtls_port",listen_dtls_port_base);  listen_dtls_port = listen_dtls_port_base;
+    cfgapi.getRoot()["settings"].lookupValue("dtls_workers",num_workers_dtls);
 
     if(cfgapi.getRoot()["settings"].exists("udp_quick_ports")) {
 
-        if(!cfgapi_obj_udp_quick_ports.empty()) {
-            DEBS_("cfgapi_load_settings: clearing existing entries in: udp_quick_ports");
-            cfgapi_obj_udp_quick_ports.clear();
+        if(!db_udp_quick_ports.empty()) {
+            _deb("load_settings: clearing existing entries in: udp_quick_ports");
+            db_udp_quick_ports.clear();
         }
 
         int num = cfgapi.getRoot()["settings"]["udp_quick_ports"].getLength();
         for(int i = 0; i < num; ++i) {
             int port = cfgapi.getRoot()["settings"]["udp_quick_ports"][i];
-            cfgapi_obj_udp_quick_ports.push_back(port);
+            db_udp_quick_ports.push_back(port);
         }
     }
 
-    cfgapi.getRoot()["settings"].lookupValue("socks_port",cfg_socks_port_base); cfg_socks_port = cfg_socks_port_base;
-    cfgapi.getRoot()["settings"].lookupValue("socks_workers",cfg_socks_workers);
+    cfgapi.getRoot()["settings"].lookupValue("socks_port",listen_socks_port_base); listen_socks_port = listen_socks_port_base;
+    cfgapi.getRoot()["settings"].lookupValue("socks_workers",num_workers_socks);
 
     if(cfgapi.getRoot().exists("settings")) {
         if(cfgapi.getRoot()["settings"].exists("socks")) {
@@ -306,15 +270,15 @@ bool cfgapi_load_settings() {
         }
     }
 
-    cfgapi.getRoot()["settings"].lookupValue("log_level",cfgapi_table.logging.level.level_);
+    cfgapi.getRoot()["settings"].lookupValue("log_level",cfgapi_table.logging.level.level_ref());
 
-    cfgapi.getRoot()["settings"].lookupValue("syslog_server",cfg_syslog_server);
-    cfgapi.getRoot()["settings"].lookupValue("syslog_port",cfg_syslog_port);
-    cfgapi.getRoot()["settings"].lookupValue("syslog_facility",cfg_syslog_facility);
-    cfgapi.getRoot()["settings"].lookupValue("syslog_level",cfg_syslog_level.level_);
-    cfgapi.getRoot()["settings"].lookupValue("syslog_family",cfg_syslog_family);
+    cfgapi.getRoot()["settings"].lookupValue("syslog_server",syslog_server);
+    cfgapi.getRoot()["settings"].lookupValue("syslog_port",syslog_port);
+    cfgapi.getRoot()["settings"].lookupValue("syslog_facility",syslog_facility);
+    cfgapi.getRoot()["settings"].lookupValue("syslog_level",syslog_level.level_ref());
+    cfgapi.getRoot()["settings"].lookupValue("syslog_family",syslog_family);
 
-    cfgapi.getRoot()["settings"].lookupValue("messages_dir",cfg_messages_dir);
+    cfgapi.getRoot()["settings"].lookupValue("messages_dir",dir_msg_templates);
 
     cfgapi.getRoot()["settings"]["cli"].lookupValue("port",cli_port_base); cli_port = cli_port_base;
     cfgapi.getRoot()["settings"]["cli"].lookupValue("enable_password",cli_enable_password);
@@ -322,12 +286,12 @@ bool cfgapi_load_settings() {
 
     if(cfgapi.getRoot().exists("settings")) {
         if(cfgapi.getRoot()["settings"].exists("auth_portal")) {
-            cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("address", cfg_auth_address);
-            cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("http_port", cfg_auth_http);
-            cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("https_port", cfg_auth_https);
-            cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("ssl_key", cfg_auth_sslkey);
-            cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("ssl_cert", cfg_auth_sslcert);
-            cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("magic_ip", cfgapi_tenant_magic_ip);
+            cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("address", auth_address);
+            cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("http_port", auth_http);
+            cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("https_port", auth_https);
+            cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("ssl_key", auth_sslkey);
+            cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("ssl_cert", auth_sslcert);
+            cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("magic_ip", tenant_magic_ip);
         }
     }
 
@@ -335,17 +299,17 @@ bool cfgapi_load_settings() {
 }
 
 
-int cfgapi_load_obj_address() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::load_db_address () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     int num = 0;
     
-    DIAS_("cfgapi_load_addresses: start");
+    _dia("cfgapi_load_addresses: start");
     
     if(cfgapi.getRoot().exists("address_objects")) {
 
         num = cfgapi.getRoot()["address_objects"].getLength();
-        DIA_("cfgapi_load_addresses: found %d objects",num);
+        _dia("cfgapi_load_addresses: found %d objects",num);
         
         Setting& curr_set = cfgapi.getRoot()["address_objects"];
 
@@ -357,35 +321,37 @@ int cfgapi_load_obj_address() {
             Setting& cur_object = curr_set[i];
 
             if (  ! cur_object.getName() ) {
-                DIA_("cfgapi_load_address: unnamed object index %d: not ok", i);
+                _dia("cfgapi_load_address: unnamed object index %d: not ok", i);
                 continue;
             }
             
             name = cur_object.getName();
 
-            DEB_("cfgapi_load_addresses: processing '%s'",name.c_str());
+            _deb("cfgapi_load_addresses: processing '%s'",name.c_str());
             
             if( cur_object.lookupValue("type",type)) {
                 switch(type) {
                     case 0: // CIDR notation
                         if (cur_object.lookupValue("cidr",address)) {
                             CIDR* c = cidr_from_str(address.c_str());
-                            cfgapi_obj_address[name] = new CidrAddress(c);
-                            cfgapi_obj_address[name]->prof_name = name;
-                            DIA_("cfgapi_load_addresses: cidr '%s': ok",name.c_str());
+                            db_address[name] = new CidrAddress(c);
+                            db_address[name]->prof_name = name;
+                            _dia("cfgapi_load_addresses: cidr '%s': ok",name.c_str());
                         }
                     break;
                     case 1: // FQDN notation
                         if (cur_object.lookupValue("fqdn",address))  {
                             FqdnAddress* f = new FqdnAddress(address);
-                            cfgapi_obj_address[name] = f;
-                            cfgapi_obj_address[name]->prof_name = name;
-                            DIA_("cfgapi_load_addresses: fqdn '%s': ok",name.c_str());
+                            db_address[name] = f;
+                            db_address[name]->prof_name = name;
+                            _dia("cfgapi_load_addresses: fqdn '%s': ok",name.c_str());
                         }
                     break;
+                    default:
+                        _dia("cfgapi_load_addresses: fqdn '%s': unknown type value(ignoring)",name.c_str());
                 }
             } else {
-                DIA_("cfgapi_load_addresses: '%s': not ok",name.c_str());
+                _dia("cfgapi_load_addresses: '%s': not ok",name.c_str());
             }
         }
     }
@@ -393,18 +359,18 @@ int cfgapi_load_obj_address() {
     return num;
 }
 
-int cfgapi_load_obj_port() {
+int CfgFactory::load_db_port () {
     
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     int num = 0;
-    
-    DIAS_("cfgapi_load_ports: start");
+
+    _dia("cfgapi_load_ports: start");
     
     if(cfgapi.getRoot().exists("port_objects")) {
 
         num = cfgapi.getRoot()["port_objects"].getLength();
-        DIA_("cfgapi_load_ports: found %d objects",num);
+        _dia("cfgapi_load_ports: found %d objects",num);
         
         Setting& curr_set = cfgapi.getRoot()["port_objects"];
 
@@ -416,26 +382,26 @@ int cfgapi_load_obj_port() {
             Setting& cur_object = curr_set[i];
 
             if (  ! cur_object.getName() ) {
-                DIA_("cfgapi_load_ports: unnamed object index %d: not ok", i);
+                _dia("cfgapi_load_ports: unnamed object index %d: not ok", i);
                 continue;
             }
 
             name = cur_object.getName();
 
-            DEB_("cfgapi_load_ports: processing '%s'",name.c_str());
+            _deb("cfgapi_load_ports: processing '%s'",name.c_str());
             
             if( cur_object.lookupValue("start",a) &&
                 cur_object.lookupValue("end",b)   ) {
                 
                 if(a <= b) {
-                    cfgapi_obj_port[name] = range(a,b);
+                    db_port[name] = range(a,b);
                 } else {
-                    cfgapi_obj_port[name] = range(b,a);
+                    db_port[name] = range(b,a);
                 }
-                
-                DIA_("cfgapi_load_ports: '%s': ok",name.c_str());
+
+                _dia("cfgapi_load_ports: '%s': ok",name.c_str());
             } else {
-                DIA_("cfgapi_load_ports: '%s': not ok",name.c_str());
+                _dia("cfgapi_load_ports: '%s': not ok",name.c_str());
             }
         }
     }
@@ -443,18 +409,18 @@ int cfgapi_load_obj_port() {
     return num;
 }
 
-int cfgapi_load_obj_proto() {
+int CfgFactory::load_db_proto () {
     
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     int num = 0;
-    
-    DIAS_("cfgapi_load_proto: start");
+
+    _dia("cfgapi_load_proto: start");
     
     if(cfgapi.getRoot().exists("proto_objects")) {
 
         num = cfgapi.getRoot()["proto_objects"].getLength();
-        DIA_("cfgapi_load_proto: found %d objects",num);
+        _dia("cfgapi_load_proto: found %d objects",num);
         
         Setting& curr_set = cfgapi.getRoot()["proto_objects"];
 
@@ -465,21 +431,21 @@ int cfgapi_load_obj_proto() {
             Setting& cur_object = curr_set[i];
 
             if (  ! cur_object.getName() ) {
-                DIA_("cfgapi_load_proto: unnamed object index %d: not ok", i);
+                _dia("cfgapi_load_proto: unnamed object index %d: not ok", i);
                 continue;
             }
             
             name = cur_object.getName();
 
-            DEB_("cfgapi_load_proto: processing '%s'",name.c_str());
+            _deb("cfgapi_load_proto: processing '%s'",name.c_str());
             
             if( cur_object.lookupValue("id",a) ) {
                 
-                cfgapi_obj_proto[name] = a;
-                
-                DIA_("cfgapi_load_proto: '%s': ok",name.c_str());
+                db_proto[name] = a;
+
+                _dia("cfgapi_load_proto: '%s': ok",name.c_str());
             } else {
-                DIA_("cfgapi_load_proto: '%s': not ok",name.c_str());
+                _dia("cfgapi_load_proto: '%s': not ok",name.c_str());
             }
         }
     }
@@ -488,18 +454,18 @@ int cfgapi_load_obj_proto() {
 }
 
 
-int cfgapi_load_obj_policy() {
+int CfgFactory::load_db_policy () {
     
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     int num = 0;
-    
-    DIAS_("cfgapi_load_policy: start");
+
+    _dia("cfgapi_load_policy: start");
     
     if(cfgapi.getRoot().exists("policy")) {
 
         num = cfgapi.getRoot()["policy"].getLength();
-        DIA_("cfgapi_load_policy: found %d objects",num);
+        _dia("cfgapi_load_policy: found %d objects",num);
         
         Setting& curr_set = cfgapi.getRoot()["policy"];
 
@@ -517,52 +483,52 @@ int cfgapi_load_obj_policy() {
             std::string nat;
             
             bool error = false;
+
+            _dia("cfgapi_load_policy: processing #%d",i);
             
-            DEB_("cfgapi_load_policy: processing #%d",i);
-            
-            PolicyRule* rule = new PolicyRule();
+            auto* rule = new PolicyRule();
 
             if(cur_object.lookupValue("proto",proto)) {
-                int r = cfgapi_lookup_proto(proto.c_str());
+                int r = lookup_proto(proto.c_str());
                 if(r != 0) {
                     rule->proto_name = proto;
                     rule->proto = r;
                     rule->proto_default = false;
-                    DIA_("cfgapi_load_policy[#%d]: proto object: %s",i,proto.c_str());
+                    _dia("cfgapi_load_policy[#%d]: proto object: %s",i,proto.c_str());
                 } else {
-                    DIA_("cfgapi_load_policy[#%d]: proto object not found: %s",i,proto.c_str());
+                    _dia("cfgapi_load_policy[#%d]: proto object not found: %s",i,proto.c_str());
                     error = true;
                 }
             }
             
             const Setting& sett_src = cur_object["src"];
             if(sett_src.isScalar()) {
-                DIA_("cfgapi_load_policy[#%d]: scalar src address object",i);
+                _dia("cfgapi_load_policy[#%d]: scalar src address object",i);
                 if(cur_object.lookupValue("src",src)) {
                     
-                    AddressObject* r = cfgapi_lookup_address(src.c_str());
+                    AddressObject* r = lookup_address(src.c_str());
                     if(r != nullptr) {
                         rule->src.push_back(r);
                         rule->src_default = false;
-                        DIA_("cfgapi_load_policy[#%d]: src address object: %s",i,src.c_str());
+                        _dia("cfgapi_load_policy[#%d]: src address object: %s",i,src.c_str());
                     } else {
-                        DIA_("cfgapi_load_policy[#%d]: src address object not found: %s",i,src.c_str());
+                        _dia("cfgapi_load_policy[#%d]: src address object not found: %s",i,src.c_str());
                         error = true;
                     }
                 }
             } else {
                 int sett_src_count = sett_src.getLength();
-                DIA_("cfgapi_load_policy[#%d]: src address list",i);
+                _dia("cfgapi_load_policy[#%d]: src address list",i);
                 for(int y = 0; y < sett_src_count; y++) {
                     const char* obj_name = sett_src[y];
                     
-                    AddressObject* r = cfgapi_lookup_address(obj_name);
+                    AddressObject* r = lookup_address(obj_name);
                     if(r != nullptr) {
                         rule->src.push_back(r);
                         rule->src_default = false;
-                        DIA_("cfgapi_load_policy[#%d]: src address object: %s",i,obj_name);
+                        _dia("cfgapi_load_policy[#%d]: src address object: %s",i,obj_name);
                     } else {
-                        DIA_("cfgapi_load_policy[#%d]: src address object not found: %s",i,obj_name);
+                        _dia("cfgapi_load_policy[#%d]: src address object not found: %s",i,obj_name);
                         error = true;
                     }
 
@@ -572,31 +538,31 @@ int cfgapi_load_obj_policy() {
             const Setting& sett_sport = cur_object["sport"];
             if(sett_sport.isScalar()) {
                 if(cur_object.lookupValue("sport",sport)) {
-                    range r = cfgapi_lookup_port(sport.c_str());
+                    range r = lookup_port(sport.c_str());
                     if(r != NULLRANGE) {
                         rule->src_ports.push_back(r);
                         rule->src_ports_names.push_back(sport);
                         rule->src_ports_default = false;
-                        DIA_("cfgapi_load_policy[#%d]: src_port object: %s",i,sport.c_str());
+                        _dia("cfgapi_load_policy[#%d]: src_port object: %s",i,sport.c_str());
                     } else {
-                        DIA_("cfgapi_load_policy[#%d]: src_port object not found: %s",i,sport.c_str());
+                        _dia("cfgapi_load_policy[#%d]: src_port object not found: %s",i,sport.c_str());
                         error = true;
                     }
                 }
             } else {
                 int sett_sport_count = sett_sport.getLength();
-                DIA_("cfgapi_load_policy[#%d]: sport list",i);
+                _dia("cfgapi_load_policy[#%d]: sport list",i);
                 for(int y = 0; y < sett_sport_count; y++) {
                     const char* obj_name = sett_sport[y];
                     
-                    range r = cfgapi_lookup_port(obj_name);
+                    range r = lookup_port(obj_name);
                     if(r != NULLRANGE) {
                         rule->src_ports.push_back(r);
-                        rule->src_ports_names.push_back(obj_name);
+                        rule->src_ports_names.emplace_back(obj_name);
                         rule->src_ports_default = false;
-                        DIA_("cfgapi_load_policy[#%d]: src_port object: %s",i,obj_name);
+                        _dia("cfgapi_load_policy[#%d]: src_port object: %s",i,obj_name);
                     } else {
-                        DIA_("cfgapi_load_policy[#%d]: src_port object not found: %s",i,obj_name);
+                        _dia("cfgapi_load_policy[#%d]: src_port object not found: %s",i,obj_name);
                         error = true;
                     }
                 }
@@ -605,29 +571,29 @@ int cfgapi_load_obj_policy() {
             const Setting& sett_dst = cur_object["dst"];
             if(sett_dst.isScalar()) {
                 if(cur_object.lookupValue("dst",dst)) {
-                    AddressObject* r = cfgapi_lookup_address(dst.c_str());
+                    AddressObject* r = lookup_address(dst.c_str());
                     if(r != nullptr) {
                         rule->dst.push_back(r);
                         rule->dst_default = false;
-                        DIA_("cfgapi_load_policy[#%d]: dst address object: %s",i,dst.c_str());
+                        _dia("cfgapi_load_policy[#%d]: dst address object: %s",i,dst.c_str());
                     } else {
-                        DIA_("cfgapi_load_policy[#%d]: dst address object not found: %s",i,dst.c_str());
+                        _dia("cfgapi_load_policy[#%d]: dst address object not found: %s",i,dst.c_str());
                         error = true;
                     }                
                 }
             } else {
                 int sett_dst_count = sett_dst.getLength();
-                DIA_("cfgapi_load_policy[#%d]: dst list",i);
+                _dia("cfgapi_load_policy[#%d]: dst list",i);
                 for(int y = 0; y < sett_dst_count; y++) {
                     const char* obj_name = sett_dst[y];
 
-                    AddressObject* r = cfgapi_lookup_address(obj_name);
+                    AddressObject* r = lookup_address(obj_name);
                     if(r != nullptr) {
                         rule->dst.push_back(r);
                         rule->dst_default = false;
-                        DIA_("cfgapi_load_policy[#%d]: dst address object: %s",i,obj_name);
+                        _dia("cfgapi_load_policy[#%d]: dst address object: %s",i,obj_name);
                     } else {
-                        DIA_("cfgapi_load_policy[#%d]: dst address object not found: %s",i,obj_name);
+                        _dia("cfgapi_load_policy[#%d]: dst address object not found: %s",i,obj_name);
                         error = true;
                     }                
                 }
@@ -637,49 +603,49 @@ int cfgapi_load_obj_policy() {
             const Setting& sett_dport = cur_object["dport"];
             if(sett_dport.isScalar()) { 
                 if(cur_object.lookupValue("dport",dport)) {
-                    range r = cfgapi_lookup_port(dport.c_str());
+                    range r = lookup_port(dport.c_str());
                     if(r != NULLRANGE) {
                         rule->dst_ports.push_back(r);
                         rule->dst_ports_names.push_back(dport);
                         rule->dst_ports_default = false;
-                        DIA_("cfgapi_load_policy[#%d]: dst_port object: %s",i,dport.c_str());
+                        _dia("cfgapi_load_policy[#%d]: dst_port object: %s",i,dport.c_str());
                     } else {
-                        DIA_("cfgapi_load_policy[#%d]: dst_port object not found: %s",i,dport.c_str());
+                        _dia("cfgapi_load_policy[#%d]: dst_port object not found: %s",i,dport.c_str());
                         error = true;
                     }
                 }
             } else {
                 int sett_dport_count = sett_dport.getLength();
-                DIA_("cfgapi_load_policy[#%d]: dst_port object list",i);
+                _dia("cfgapi_load_policy[#%d]: dst_port object list",i);
                 for(int y = 0; y < sett_dport_count; y++) {
                     const char* obj_name = sett_dport[y];
                     
-                    range r = cfgapi_lookup_port(obj_name);
+                    range r = lookup_port(obj_name);
                     if(r != NULLRANGE) {
                         rule->dst_ports.push_back(r);
-                        rule->dst_ports_names.push_back(obj_name);
+                        rule->dst_ports_names.emplace_back(obj_name);
                         rule->dst_ports_default = false;
-                        DIA_("cfgapi_load_policy[#%d]: dst_port object: %s",i,obj_name);
+                        _dia("cfgapi_load_policy[#%d]: dst_port object: %s",i,obj_name);
                     } else {
-                        DIA_("cfgapi_load_policy[#%d]: dst_port object not found: %s",i,obj_name);
+                        _dia("cfgapi_load_policy[#%d]: dst_port object not found: %s",i,obj_name);
                         error = true;
                     }                    
                 }
             }
             
             if(cur_object.lookupValue("action",action)) {
-                int r_a = 1;
+                int r_a = POLICY_ACTION_PASS;
                 if(action == "deny") {
-                    DIA_("cfgapi_load_policy[#%d]: action: deny",i);
+                    _dia("cfgapi_load_policy[#%d]: action: deny",i);
                     r_a = POLICY_ACTION_DENY;
                     rule->action_name = action;
 
                 } else if (action == "accept"){
-                    DIA_("cfgapi_load_policy[#%d]: action: accept",i);
+                    _dia("cfgapi_load_policy[#%d]: action: accept",i);
                     r_a = POLICY_ACTION_PASS;
                     rule->action_name = action;
                 } else {
-                    DIA_("cfgapi_load_policy[#%d]: action: unknown action '%s'",i,action.c_str());
+                    _dia("cfgapi_load_policy[#%d]: action: unknown action '%s'",i,action.c_str());
                     r_a  = POLICY_ACTION_DENY;
                     error = true;
                 }
@@ -694,16 +660,16 @@ int cfgapi_load_obj_policy() {
                 int nat_a = POLICY_NAT_NONE;
                 
                 if(nat == "none") {
-                    DIA_("cfgapi_load_policy[#%d]: nat: none",i);
+                    _dia("cfgapi_load_policy[#%d]: nat: none",i);
                     nat_a = POLICY_NAT_NONE;
                     rule->nat_name = nat;
 
                 } else if (nat == "auto"){
-                    DIA_("cfgapi_load_policy[#%d]: nat: auto",i);
+                    _dia("cfgapi_load_policy[#%d]: nat: auto",i);
                     nat_a = POLICY_NAT_AUTO;
                     rule->nat_name = nat;
                 } else {
-                    DIA_("cfgapi_load_policy[#%d]: nat: unknown nat method '%s'",i,nat.c_str());
+                    _dia("cfgapi_load_policy[#%d]: nat: unknown nat method '%s'",i,nat.c_str());
                     nat_a  = POLICY_NAT_NONE;
                     rule->nat_name = "none";
                     error = true;
@@ -726,63 +692,63 @@ int cfgapi_load_obj_policy() {
                 std::string name_alg_dns;
                 
                 if(cur_object.lookupValue("detection_profile",name_detection)) {
-                    ProfileDetection* prf  = cfgapi_lookup_profile_detection(name_detection.c_str());
+                    ProfileDetection* prf  = lookup_prof_detection(name_detection.c_str());
                     if(prf != nullptr) {
-                        DIA_("cfgapi_load_policy[#%d]: detect profile %s",i,name_detection.c_str());
+                        _dia("cfgapi_load_policy[#%d]: detect profile %s",i,name_detection.c_str());
                         rule->profile_detection = prf;
                     } else {
-                        ERR_("cfgapi_load_policy[#%d]: detect profile %s cannot be loaded",i,name_detection.c_str());
+                        _err("cfgapi_load_policy[#%d]: detect profile %s cannot be loaded",i,name_detection.c_str());
                         error = true;
                     }
                 }
                 
                 if(cur_object.lookupValue("content_profile",name_content)) {
-                    ProfileContent* prf  = cfgapi_lookup_profile_content(name_content.c_str());
+                    ProfileContent* prf  = lookup_prof_content(name_content.c_str());
                     if(prf != nullptr) {
-                        DIA_("cfgapi_load_policy[#%d]: content profile %s",i,name_content.c_str());
+                        _dia("cfgapi_load_policy[#%d]: content profile %s",i,name_content.c_str());
                         rule->profile_content = prf;
                     } else {
-                        ERR_("cfgapi_load_policy[#%d]: content profile %s cannot be loaded",i,name_content.c_str());
+                        _err("cfgapi_load_policy[#%d]: content profile %s cannot be loaded",i,name_content.c_str());
                         error = true;
                     }
                 }                
                 if(cur_object.lookupValue("tls_profile",name_tls)) {
-                    ProfileTls* tls  = cfgapi_lookup_profile_tls(name_tls.c_str());
+                    ProfileTls* tls  = lookup_prof_tls(name_tls.c_str());
                     if(tls != nullptr) {
-                        DIA_("cfgapi_load_policy[#%d]: tls profile %s",i,name_tls.c_str());
+                        _dia("cfgapi_load_policy[#%d]: tls profile %s",i,name_tls.c_str());
                         rule->profile_tls= tls;
                     } else {
-                        ERR_("cfgapi_load_policy[#%d]: tls profile %s cannot be loaded",i,name_tls.c_str());
+                        _err("cfgapi_load_policy[#%d]: tls profile %s cannot be loaded",i,name_tls.c_str());
                         error = true;
                     }
                 }         
                 if(cur_object.lookupValue("auth_profile",name_auth)) {
-                    ProfileAuth* auth  = cfgapi_lookup_profile_auth(name_auth.c_str());
+                    ProfileAuth* auth  = lookup_prof_auth(name_auth.c_str());
                     if(auth != nullptr) {
-                        DIA_("cfgapi_load_policy[#%d]: auth profile %s",i,name_auth.c_str());
+                        _dia("cfgapi_load_policy[#%d]: auth profile %s",i,name_auth.c_str());
                         rule->profile_auth= auth;
                     } else {
-                        ERR_("cfgapi_load_policy[#%d]: auth profile %s cannot be loaded",i,name_auth.c_str());
+                        _err("cfgapi_load_policy[#%d]: auth profile %s cannot be loaded",i,name_auth.c_str());
                         error = true;
                     }
                 }
                 if(cur_object.lookupValue("alg_dns_profile",name_alg_dns)) {
-                    ProfileAlgDns* dns  = cfgapi_lookup_profile_alg_dns(name_alg_dns.c_str());
+                    ProfileAlgDns* dns  = lookup_prof_alg_dns(name_alg_dns.c_str());
                     if(dns != nullptr) {
-                        DIA_("cfgapi_load_policy[#%d]: DNS alg profile %s",i,name_alg_dns.c_str());
+                        _dia("cfgapi_load_policy[#%d]: DNS alg profile %s",i,name_alg_dns.c_str());
                         rule->profile_alg_dns = dns;
                     } else {
-                        ERR_("cfgapi_load_policy[#%d]: DNS alg %s cannot be loaded",i,name_alg_dns.c_str());
+                        _err("cfgapi_load_policy[#%d]: DNS alg %s cannot be loaded",i,name_alg_dns.c_str());
                         error = true;
                     }
                 }                    
             }
             
             if(!error){
-                DIA_("cfgapi_load_policy[#%d]: ok",i);
-                cfgapi_obj_policy.push_back(rule);
+                _dia("cfgapi_load_policy[#%d]: ok",i);
+                db_policy.push_back(rule);
             } else {
-                ERR_("cfgapi_load_policy[#%d]: not ok (will not process traffic)",i);
+                _err("cfgapi_load_policy[#%d]: not ok (will not process traffic)",i);
             }
         }
     }
@@ -790,177 +756,180 @@ int cfgapi_load_obj_policy() {
     return num;
 }
 
-int cfgapi_obj_policy_match(baseProxy* proxy) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::policy_match (baseProxy *proxy) {
+
+    auto log = logan_lite("policy.rule");
+
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     int x = 0;
-    for( std::vector<PolicyRule*>::iterator i = cfgapi_obj_policy.begin(); i != cfgapi_obj_policy.end(); ++i) {
-        PolicyRule* rule = (*i);
+    for( auto rule: db_policy) {
+
         bool r = rule->match(proxy);
         
         if(r) {
-            DIA_("cfgapi_obj_policy_match: matched #%d",x);
+            _deb("policy_match: matched #%d",x);
             return x;
         }
         
         x++;
     }
-    
-    DIAS_("cfgapi_obj_policy_match: implicit deny");
+
+    _not("policy_match: implicit deny");
     return -1;
 }
 
-int cfgapi_obj_policy_match(std::vector<baseHostCX*>& left, std::vector<baseHostCX*>& right) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::policy_match (std::vector<baseHostCX *> &left, std::vector<baseHostCX *> &right) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     int x = 0;
-    for( std::vector<PolicyRule*>::iterator i = cfgapi_obj_policy.begin(); i != cfgapi_obj_policy.end(); ++i) {
-        PolicyRule* rule = (*i);
+    for( auto rule: db_policy) {
+
         bool r = rule->match(left,right);
         
         if(r) {
-            DIA_("cfgapi_obj_policy_match_lr: matched #%d",x);
+            _dia("cfgapi_obj_policy_match_lr: matched #%d",x);
             return x;
         }
         
         x++;
     }
-    
-    DIAS_("cfgapi_obj_policy_match_lr: implicit deny");
+
+    _dia("cfgapi_obj_policy_match_lr: implicit deny");
     return -1;
 }    
 
-int cfgapi_obj_policy_action(int index) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::policy_action (int index) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     if(index < 0) {
         return -1;
     }
     
-    if(index < (signed int)cfgapi_obj_policy.size()) {
-        return cfgapi_obj_policy.at(index)->action;
+    if(index < (signed int)db_policy.size()) {
+        return db_policy.at(index)->action;
     } else {
-        DIA_("cfg_obj_policy_action[#%d]: out of bounds, deny",index);
+        _dia("cfg_obj_policy_action[#%d]: out of bounds, deny",index);
         return POLICY_ACTION_DENY;
     }
 }
 
-ProfileContent* cfgapi_obj_policy_profile_content(int index) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+ProfileContent* CfgFactory::policy_prof_content (int index) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     if(index < 0) {
         return nullptr;
     }
     
-    if(index < (signed int)cfgapi_obj_policy.size()) {
-        return cfgapi_obj_policy.at(index)->profile_content;
+    if(index < (signed int)db_policy.size()) {
+        return db_policy.at(index)->profile_content;
     } else {
-        DIA_("cfgapi_obj_policy_profile_content[#%d]: out of bounds, nullptr",index);
+        _dia("policy_prof_content[#%d]: out of bounds, nullptr",index);
         return nullptr;
     }
 }
 
-ProfileDetection* cfgapi_obj_policy_profile_detection(int index) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+ProfileDetection* CfgFactory::policy_prof_detection (int index) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     if(index < 0) {
         return nullptr;
     }
     
-    if(index < (signed int)cfgapi_obj_policy.size()) {
-        return cfgapi_obj_policy.at(index)->profile_detection;
+    if(index < (signed int)db_policy.size()) {
+        return db_policy.at(index)->profile_detection;
     } else {
-        DIA_("cfgapi_obj_policy_profile_detection[#%d]: out of bounds, nullptr",index);
+        _dia("policy_prof_detection[#%d]: out of bounds, nullptr",index);
         return nullptr;
     }
 }
 
-ProfileTls* cfgapi_obj_policy_profile_tls(int index) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+ProfileTls* CfgFactory::policy_prof_tls (int index) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     if(index < 0) {
         return nullptr;
     }
     
-    if(index < (signed int)cfgapi_obj_policy.size()) {
-        return cfgapi_obj_policy.at(index)->profile_tls;
+    if(index < (signed int)db_policy.size()) {
+        return db_policy.at(index)->profile_tls;
     } else {
-        DIA_("cfgapi_obj_policy_profile_tls[#%d]: out of bounds, nullptr",index);
+        _dia("policy_prof_tls[#%d]: out of bounds, nullptr",index);
         return nullptr;
     }
 }
 
 
-ProfileAlgDns* cfgapi_obj_policy_profile_alg_dns(int index) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+ProfileAlgDns* CfgFactory::policy_prof_alg_dns (int index) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     if(index < 0) {
         return nullptr;
     }
     
-    if(index < (signed int)cfgapi_obj_policy.size()) {
-        return cfgapi_obj_policy.at(index)->profile_alg_dns;
+    if(index < (signed int)db_policy.size()) {
+        return db_policy.at(index)->profile_alg_dns;
     } else {
-        DIA_("cfgapi_obj_policy_profile_alg_dns[#%d]: out of bounds, nullptr",index);
+        _dia("policy_prof_alg_dns[#%d]: out of bounds, nullptr",index);
         return nullptr;
     }
 }
 
 
-ProfileAuth* cfgapi_obj_policy_profile_auth(int index) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+ProfileAuth* CfgFactory::policy_prof_auth (int index) {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     if(index < 0) {
         return nullptr;
     }
     
-    if(index < (signed int)cfgapi_obj_policy.size()) {
-        return cfgapi_obj_policy.at(index)->profile_auth;
+    if(index < (signed int)db_policy.size()) {
+        return db_policy.at(index)->profile_auth;
     } else {
-        DIA_("cfgapi_obj_policy_profile_auth[#%d]: out of bounds, nullptr",index);
+        _dia("policy_prof_auth[#%d]: out of bounds, nullptr",index);
         return nullptr;
     }
 }
 
 
 
-int cfgapi_load_obj_profile_detection() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::load_db_prof_detection () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     int num = 0;
-    
-    DIAS_("cfgapi_load_obj_profile_detect: start");
+
+    _dia("cfgapi_load_obj_profile_detect: start");
     
     if(cfgapi.getRoot().exists("detection_profiles")) {
 
         num = cfgapi.getRoot()["detection_profiles"].getLength();
-        DIA_("cfgapi_load_obj_profile_detect: found %d objects",num);
+        _dia("cfgapi_load_obj_profile_detect: found %d objects",num);
         
         Setting& curr_set = cfgapi.getRoot()["detection_profiles"];
         
         for( int i = 0; i < num; i++) {
             std::string name;
-            ProfileDetection* a = new ProfileDetection;
+            auto* a = new ProfileDetection;
             
             Setting& cur_object = curr_set[i];
 
             if (  ! cur_object.getName() ) {
-                DIA_("cfgapi_load_obj_profile_detect: unnamed object index %d: not ok", i);
+                _dia("cfgapi_load_obj_profile_detect: unnamed object index %d: not ok", i);
                 continue;
             }
 
             name = cur_object.getName();
-           
-            DIA_("cfgapi_load_obj_profile_detect: processing '%s'",name.c_str());
+
+            _dia("cfgapi_load_obj_profile_detect: processing '%s'",name.c_str());
             
             if( cur_object.lookupValue("mode",a->mode) ) {
                 
                 a->prof_name = name;
-                cfgapi_obj_profile_detection[name] = a;
-                
-                DIA_("cfgapi_load_obj_profile_detect: '%s': ok",name.c_str());
+                db_prof_detection[name] = a;
+
+                _dia("cfgapi_load_obj_profile_detect: '%s': ok",name.c_str());
             } else {
-                DIA_("cfgapi_load_obj_profile_detect: '%s': not ok",name.c_str());
+                _dia("cfgapi_load_obj_profile_detect: '%s': not ok",name.c_str());
             }
         }
     }
@@ -969,43 +938,43 @@ int cfgapi_load_obj_profile_detection() {
 }
 
 
-int cfgapi_load_obj_profile_content() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::load_db_prof_content () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     int num = 0;
-    
-    DIAS_("cfgapi_load_obj_profile_content: start");
+
+    _dia("load_db_prof_content: start");
     
     if(cfgapi.getRoot().exists("content_profiles")) {
 
         num = cfgapi.getRoot()["content_profiles"].getLength();
-        DIA_("cfgapi_load_obj_profile_content: found %d objects",num);
+        _dia("load_db_prof_content: found %d objects",num);
         
         Setting& curr_set = cfgapi.getRoot()["content_profiles"];
 
         for( int i = 0; i < num; i++) {
             std::string name;
-            ProfileContent* a = new ProfileContent;
+            auto* a = new ProfileContent;
             
             Setting& cur_object = curr_set[i];
 
             if (  ! cur_object.getName() ) {
-                DIA_("cfgapi_load_obj_profile_content: unnamed object index %d: not ok", i);
+                _dia("load_db_prof_content: unnamed object index %d: not ok", i);
                 continue;
             }
 
             name = cur_object.getName();
 
-            DEB_("cfgapi_load_obj_profile_content: processing '%s'",name.c_str());
+            _dia("load_db_prof_content: processing '%s'",name.c_str());
             
             if( cur_object.lookupValue("write_payload",a->write_payload) ) {
                 
                 a->prof_name = name;
-                cfgapi_obj_profile_content[name] = a;
+                db_prof_content[name] = a;
                 
                 if(cur_object.exists("content_rules")) {
                     int jnum = cur_object["content_rules"].getLength();
-                    DIA_("replace rules in profile '%s', size %d",name.c_str(),jnum);
+                    _dia("replace rules in profile '%s', size %d",name.c_str(),jnum);
                     for (int j = 0; j < jnum; j++) {
                         Setting& cur_replace_rule = cur_object["content_rules"][j];
 
@@ -1026,8 +995,8 @@ int cfgapi_load_obj_profile_content() {
                         cur_replace_rule.lookupValue("fill_length",fill_length);
                         cur_replace_rule.lookupValue("replace_each_nth",replace_each_nth);
                         
-                        if(m.size() > 0 && action_defined) {
-                            DIA_("    [%d] match '%s' and replace with '%s'",j,m.c_str(),r.c_str());
+                        if( (! m.empty() ) && action_defined) {
+                            _dia("    [%d] match '%s' and replace with '%s'",j,m.c_str(),r.c_str());
                             ProfileContentRule p;
                             p.match = m;
                             p.replace = r;
@@ -1037,15 +1006,15 @@ int cfgapi_load_obj_profile_content() {
                             a->content_rules.push_back(p);
                             
                         } else {
-                            ERR_("    [%d] unfinished replace policy",j);
+                            _dia("    [%d] unfinished replace policy",j);
                         }
                     }
                 }
-                
-                
-                DIA_("cfgapi_load_obj_profile_content: '%s': ok",name.c_str());
+
+
+                _dia("load_db_prof_content: '%s': ok",name.c_str());
             } else {
-                DIA_("cfgapi_load_obj_profile_content: '%s': not ok",name.c_str());
+                _dia("load_db_prof_content: '%s': not ok",name.c_str());
             }
         }
     }
@@ -1053,34 +1022,34 @@ int cfgapi_load_obj_profile_content() {
     return num;
 }
 
-int cfgapi_load_obj_profile_tls() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::load_db_prof_tls () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     int num = 0;
-    
-    DIAS_("cfgapi_load_obj_profile_tls: start");
+
+    _dia("load_db_prof_tls: start");
     
     if(cfgapi.getRoot().exists("tls_profiles")) {
 
         num = cfgapi.getRoot()["tls_profiles"].getLength();
-        DIA_("cfgapi_load_obj_profile_tls: found %d objects",num);
+        _dia("load_db_prof_tls: found %d objects",num);
         
         Setting& curr_set = cfgapi.getRoot()["tls_profiles"];
 
         for( int i = 0; i < num; i++) {
             std::string name;
-            ProfileTls* a = new ProfileTls;
+            auto* a = new ProfileTls;
             
             Setting& cur_object = curr_set[i];
 
             if (  ! cur_object.getName() ) {
-                DIA_("cfgapi_load_obj_profile_tls: unnamed object index %d: not ok", i);
+                _dia("load_db_prof_tls: unnamed object index %d: not ok", i);
                 continue;
             }
 
             name = cur_object.getName();
 
-            DEB_("cfgapi_load_obj_profile_tls: processing '%s'",name.c_str());
+            _dia("load_db_prof_tls: processing '%s'",name.c_str());
             
             if( cur_object.lookupValue("inspect",a->inspect) ) {
                 
@@ -1132,11 +1101,11 @@ int cfgapi_load_obj_profile_tls() {
                 }
                 cur_object.lookupValue("sslkeylog",a->sslkeylog);
                 
-                cfgapi_obj_profile_tls[name] = a;
-                
-                DIA_("cfgapi_load_obj_profile_tls: '%s': ok",name.c_str());
+                db_prof_tls[name] = a;
+
+                _dia("load_db_prof_tls: '%s': ok",name.c_str());
             } else {
-                DIA_("cfgapi_load_obj_profile_tls: '%s': not ok",name.c_str());
+                _dia("load_db_prof_tls: '%s': not ok",name.c_str());
             }
         }
     }
@@ -1144,38 +1113,40 @@ int cfgapi_load_obj_profile_tls() {
     return num;
 }
 
-int cfgapi_load_obj_profile_alg_dns() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::load_db_prof_alg_dns () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
 
     int num = 0;
-    DIAS_("cfgapi_load_obj_alg_dns_profile: start");
+    _dia("cfgapi_load_obj_alg_dns_profile: start");
     if(cfgapi.getRoot().exists("alg_dns_profiles")) {
         num = cfgapi.getRoot()["alg_dns_profiles"].getLength();
-        DIA_("cfgapi_load_obj_alg_dns_profile: found %d objects",num);
+        _dia("cfgapi_load_obj_alg_dns_profile: found %d objects",num);
         
         Setting& curr_set = cfgapi.getRoot()["alg_dns_profiles"];
 
         for( int i = 0; i < num; i++) {
             std::string name;
-            ProfileAlgDns* a = new ProfileAlgDns;
+            auto* a = new ProfileAlgDns;
             
             Setting& cur_object = curr_set[i];
 
             if (  ! cur_object.getName() ) {
-                DIA_("cfgapi_load_obj_alg_dns_profile: unnamed object index %d: not ok", i);
+                _dia("cfgapi_load_obj_alg_dns_profile: unnamed object index %d: not ok", i);
+
+                delete a; // coverity: 1407948
                 continue;
             }
             
             name = cur_object.getName();
 
-            DEB_("cfgapi_load_obj_alg_dns_profile: processing '%s'",name.c_str());
+            _dia("cfgapi_load_obj_alg_dns_profile: processing '%s'",name.c_str());
             
             a->prof_name = name;
             cur_object.lookupValue("match_request_id",a->match_request_id);
             cur_object.lookupValue("randomize_id",a->randomize_id);
             cur_object.lookupValue("cached_responses",a->cached_responses);
             
-            cfgapi_obj_profile_alg_dns[name] = a;
+            db_prof_alg_dns[name] = a;
         }
     }
     
@@ -1183,57 +1154,60 @@ int cfgapi_load_obj_profile_alg_dns() {
 }
 
 
-int cfgapi_load_obj_profile_auth() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::load_db_prof_auth () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
     int num = 0;
-    
-    DIAS_("cfgapi_load_obj_profile_auth: start");
-    
-    DIAS_("cfgapi_load_obj_profile_auth: portal settings");
-    cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("address",cfgapi_identity_portal_address);
-    cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("address6",cfgapi_identity_portal_address6);
-    cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("http_port",cfgapi_identity_portal_port_http);
-    cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("https_port",cfgapi_identity_portal_port_https);    
-    
-    DIAS_("cfgapi_load_obj_profile_auth: profiles");
+
+    _dia("load_db_prof_auth: start");
+
+    _dia("load_db_prof_auth: portal settings");
+    cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("address", AuthFactory::get().portal_address);
+    cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("address6", AuthFactory::get().portal_address6);
+    cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("http_port", AuthFactory::get().portal_port_http);
+    cfgapi.getRoot()["settings"]["auth_portal"].lookupValue("https_port", AuthFactory::get().portal_port_https);
+
+    _dia("load_db_prof_auth: profiles");
     if(cfgapi.getRoot().exists("auth_profiles")) {
 
         num = cfgapi.getRoot()["auth_profiles"].getLength();
-        DIA_("cfgapi_load_obj_profile_auth: found %d objects",num);
+        _dia("load_db_prof_auth: found %d objects",num);
         
         Setting& curr_set = cfgapi.getRoot()["auth_profiles"];
 
         for( int i = 0; i < num; i++) {
             std::string name;
-            ProfileAuth* a = new ProfileAuth;
+            auto* a = new ProfileAuth;
             
             Setting& cur_object = curr_set[i];
 
             if (  ! cur_object.getName() ) {
-                DIA_("cfgapi_load_obj_profile_auth: unnamed object index %d: not ok", i);
+                _dia("load_db_prof_auth: unnamed object index %d: not ok", i);
+                delete a; // coverity: 1408003
                 continue;
             }
             
             name = cur_object.getName();
 
-            DEB_("cfgapi_load_obj_profile_auth: processing '%s'",name.c_str());
+            _deb("load_db_prof_auth: processing '%s'",name.c_str());
             
             a->prof_name = name;
             cur_object.lookupValue("authenticate",a->authenticate);
             cur_object.lookupValue("resolve",a->resolve);
             
             if(cur_object.exists("identities")) {
-                DIAS_("cfgapi_load_obj_profile_auth: profiles: subpolicies exists");
+                _dia("load_db_prof_auth: profiles: subpolicies exists");
                 int sub_pol_num = cur_object["identities"].getLength();
-                DIA_("cfgapi_load_obj_profile_auth: profiles: %d subpolicies detected",sub_pol_num);
+                _dia("load_db_prof_auth: profiles: %d subpolicies detected",sub_pol_num);
                 for (int j = 0; j < sub_pol_num; j++) {
                     Setting& cur_subpol = cur_object["identities"][j];
                     
-                    ProfileSubAuth* n_subpol = new ProfileSubAuth();
+                    auto* n_subpol = new ProfileSubAuth();
 
                     if (  ! cur_subpol.getName() ) {
-                        DIA_("cfgapi_load_obj_profile_auth: profiles: unnamed object index %d: not ok", j);
+                        _dia("load_db_prof_auth: profiles: unnamed object index %d: not ok", j);
+
+                        delete n_subpol; // coverity: 1407960
                         continue;
                     }
 
@@ -1246,54 +1220,54 @@ int cfgapi_load_obj_profile_auth() {
                     std::string name_alg_dns;
                     
                     if(cur_subpol.lookupValue("detection_profile",name_detection)) {
-                        ProfileDetection* prf  = cfgapi_lookup_profile_detection(name_detection.c_str());
+                        ProfileDetection* prf  = lookup_prof_detection(name_detection.c_str());
                         if(prf != nullptr) {
-                            DIA_("cfgapi_load_obj_profile_auth[sub-profile:%s]: detect profile %s",n_subpol->name.c_str(),name_detection.c_str());
+                            _dia("load_db_prof_auth[sub-profile:%s]: detect profile %s",n_subpol->name.c_str(),name_detection.c_str());
                             n_subpol->profile_detection = prf;
                         } else {
-                            ERR_("cfgapi_load_obj_profile_auth[sub-profile:%s]: detect profile %s cannot be loaded",n_subpol->name.c_str(),name_detection.c_str());
+                            _err("load_db_prof_auth[sub-profile:%s]: detect profile %s cannot be loaded",n_subpol->name.c_str(),name_detection.c_str());
                         }
                     }
                     
                     if(cur_subpol.lookupValue("content_profile",name_content)) {
-                        ProfileContent* prf  = cfgapi_lookup_profile_content(name_content.c_str());
+                        ProfileContent* prf  = lookup_prof_content(name_content.c_str());
                         if(prf != nullptr) {
-                            DIA_("cfgapi_load_obj_profile_auth[sub-profile:%s]: content profile %s",n_subpol->name.c_str(),name_content.c_str());
+                            _dia("load_db_prof_auth[sub-profile:%s]: content profile %s",n_subpol->name.c_str(),name_content.c_str());
                             n_subpol->profile_content = prf;
                         } else {
-                            ERR_("cfgapi_load_obj_profile_auth[sub-profile:%s]: content profile %s cannot be loaded",n_subpol->name.c_str(),name_content.c_str());
+                            _err("load_db_prof_auth[sub-profile:%s]: content profile %s cannot be loaded",n_subpol->name.c_str(),name_content.c_str());
                         }
                     }                
                     if(cur_subpol.lookupValue("tls_profile",name_tls)) {
-                        ProfileTls* tls  = cfgapi_lookup_profile_tls(name_tls.c_str());
+                        ProfileTls* tls  = lookup_prof_tls(name_tls.c_str());
                         if(tls != nullptr) {
-                            DIA_("cfgapi_load_obj_profile_auth[sub-profile:%s]: tls profile %s",n_subpol->name.c_str(),name_tls.c_str());
+                            _dia("load_db_prof_auth[sub-profile:%s]: tls profile %s",n_subpol->name.c_str(),name_tls.c_str());
                             n_subpol->profile_tls= tls;
                         } else {
-                            ERR_("cfgapi_load_obj_profile_auth[sub-profile:%s]: tls profile %s cannot be loaded",n_subpol->name.c_str(),name_tls.c_str());
+                            _err("load_db_prof_auth[sub-profile:%s]: tls profile %s cannot be loaded",n_subpol->name.c_str(),name_tls.c_str());
                         }
                     }         
 
                     // we don't need auth profile in auth sub-profile
                     
                     if(cur_subpol.lookupValue("alg_dns_profile",name_alg_dns)) {
-                        ProfileAlgDns* dns  = cfgapi_lookup_profile_alg_dns(name_alg_dns.c_str());
+                        ProfileAlgDns* dns  = lookup_prof_alg_dns(name_alg_dns.c_str());
                         if(dns != nullptr) {
-                            DIA_("cfgapi_load_obj_profile_auth[sub-profile:%s]: DNS alg profile %s",n_subpol->name.c_str(),name_alg_dns.c_str());
+                            _dia("load_db_prof_auth[sub-profile:%s]: DNS alg profile %s",n_subpol->name.c_str(),name_alg_dns.c_str());
                             n_subpol->profile_alg_dns = dns;
                         } else {
-                            ERR_("cfgapi_load_obj_profile_auth[sub-profile:%s]: DNS alg %s cannot be loaded",n_subpol->name.c_str(),name_alg_dns.c_str());
+                            _err("load_db_prof_auth[sub-profile:%s]: DNS alg %s cannot be loaded",n_subpol->name.c_str(),name_alg_dns.c_str());
                         }
                     }                    
 
                     
                     a->sub_policies.push_back(n_subpol);
-                    DIA_("cfgapi_load_obj_profile_auth: profiles: %d:%s",j,n_subpol->name.c_str());
+                    _dia("load_db_prof_auth: profiles: %d:%s",j,n_subpol->name.c_str());
                 }
             }
-            cfgapi_obj_profile_auth[name] = a;
-            
-            DIA_("cfgapi_load_obj_profile_auth: '%s': ok",name.c_str());
+            db_prof_auth[name] = a;
+
+            _dia("load_db_prof_auth: '%s': ok",name.c_str());
         }
     }
     
@@ -1303,171 +1277,151 @@ int cfgapi_load_obj_profile_auth() {
 
 
 
-int cfgapi_cleanup_obj_address() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::cleanup_db_address () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    int r = cfgapi_obj_address.size();
-    
-    for (std::map<std::string,AddressObject*>::iterator i = cfgapi_obj_address.begin(); i != cfgapi_obj_address.end(); ++i)
+    int r = db_address.size();
+
+    for (auto& a: db_address)
     {
-        std::pair<std::string,AddressObject*> a = (*i);
         AddressObject* c = a.second;
-        if (c != nullptr) delete c;
-
-        a.second = nullptr;
+        delete c;
     }
     
-    cfgapi_obj_address.clear();
+    db_address.clear();
     
-    DEB_("cfgapi_cleanup_obj_address: %d objects freed",r);
+    _deb("cleanup_db_address: %d objects freed",r);
     return r;
 }
 
-int cfgapi_cleanup_obj_policy() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::cleanup_db_policy () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    int r = cfgapi_obj_policy.size();
-    for(std::vector<PolicyRule*>::iterator i = cfgapi_obj_policy.begin(); i < cfgapi_obj_policy.end(); ++i) {
-        PolicyRule* ptr = (*i);
-        if (ptr != nullptr) delete ptr;
-        (*i) = nullptr;
+    int r = db_policy.size();
+
+    for(auto* ptr: db_policy) {
+        delete ptr;
     }
+    db_policy.clear();
     
-    cfgapi_obj_policy.clear();
-    
-    DEB_("cfgapi_cleanup_obj_policy: %d objects freed",r);
+    _deb("cleanup_db_policy: %d objects freed",r);
     return r;
 }
 
-int cfgapi_cleanup_obj_port() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::cleanup_db_port () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    int r = cfgapi_obj_port.size();
-    cfgapi_obj_port.clear();
-    
-    return r;
-}
-
-int cfgapi_cleanup_obj_proto() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
-    
-    int r = cfgapi_obj_proto.size();
-    cfgapi_obj_proto.clear();
+    int r = db_port.size();
+    db_port.clear();
     
     return r;
 }
 
-
-int cfgapi_cleanup_obj_profile_content() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::cleanup_db_proto () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    int r = cfgapi_obj_profile_content.size();
-    for(std::map<std::string, ProfileContent*>::iterator i = cfgapi_obj_profile_content.begin(); i != cfgapi_obj_profile_content.end(); ++i) {
-        std::pair<std::string,ProfileContent*> t = (*i);
+    int r = db_proto.size();
+    db_proto.clear();
+    
+    return r;
+}
+
+
+int CfgFactory::cleanup_db_prof_content () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
+    
+    int r = db_prof_content.size();
+    for(auto& t: db_prof_content) {
         ProfileContent* c = t.second;
-        if (c != nullptr) delete c;
+        delete c;
     }
-    cfgapi_obj_profile_content.clear();
+    db_prof_content.clear();
     
     return r;
 }
-int cfgapi_cleanup_obj_profile_detection() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::cleanup_db_prof_detection () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    int r = cfgapi_obj_profile_detection.size();
-    for(std::map<std::string, ProfileDetection*>::iterator i = cfgapi_obj_profile_detection.begin(); i != cfgapi_obj_profile_detection.end(); ++i) {
-        std::pair<std::string,ProfileDetection*> t = (*i);
+    int r = db_prof_detection.size();
+    for(auto& t: db_prof_detection) {
         ProfileDetection* c = t.second;
-        if (c != nullptr) delete c;
+        delete c;
     }
-    cfgapi_obj_profile_detection.clear();
+    db_prof_detection.clear();
     
     return r;
 }
-int cfgapi_cleanup_obj_profile_tls() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::cleanup_db_prof_tls () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    int r = cfgapi_obj_profile_tls.size();
-    for(std::map<std::string, ProfileTls*>::iterator i = cfgapi_obj_profile_tls.begin(); i != cfgapi_obj_profile_tls.end(); ++i) {
-        std::pair<std::string,ProfileTls*> t = (*i);
+    int r = db_prof_tls.size();
+    for(auto& t: db_prof_tls) {
         ProfileTls* c = t.second;
-        if (c != nullptr) delete c;
+        delete c;
     }
-    cfgapi_obj_profile_tls.clear();
+    db_prof_tls.clear();
     
     return r;
 }
 
-int cfgapi_cleanup_obj_profile_alg_dns() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::cleanup_db_prof_alg_dns () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    int r = cfgapi_obj_profile_alg_dns.size();
-    for(std::map<std::string, ProfileAlgDns*>::iterator i = cfgapi_obj_profile_alg_dns.begin(); i != cfgapi_obj_profile_alg_dns.end(); ++i) {
-        std::pair<std::string,ProfileAlgDns*> t = (*i);
+    int r = db_prof_alg_dns.size();
+    for(auto& t: db_prof_alg_dns) {
         ProfileAlgDns* c = t.second;
-        if (c != nullptr) delete c;
+        delete c;
     }
-    cfgapi_obj_profile_alg_dns.clear();
+    db_prof_alg_dns.clear();
     
     return r;
 }
 
 
 
-int cfgapi_cleanup_obj_profile_auth() {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::cleanup_db_prof_auth () {
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    int r = cfgapi_obj_profile_auth.size();
-    for(std::map<std::string, ProfileAuth*>::iterator i = cfgapi_obj_profile_auth.begin(); i != cfgapi_obj_profile_auth.end(); ++i) {
-        std::pair<std::string,ProfileAuth*> t = (*i);
+    int r = db_prof_auth.size();
+    for(auto& t: db_prof_auth) {
         ProfileAuth* c = t.second;
         
         for(auto j: c->sub_policies) {
             delete j;
         }
         
-        if (c != nullptr) delete c;
+        delete c;
     }
-    cfgapi_obj_profile_auth.clear();
+    db_prof_auth.clear();
     
     return r;
 }
 
 
-bool cfgapi_obj_profile_content_apply(baseHostCX* originator, baseProxy* new_proxy, ProfileContent* pc) {
-    
-    MitmProxy* mitm_proxy = static_cast<MitmProxy*>(new_proxy); 
-    AppHostCX* mitm_originator = static_cast<AppHostCX*>(originator);
-    
+bool CfgFactory::prof_content_apply (baseHostCX *originator, baseProxy *new_proxy, ProfileContent *pc) {
+
+    auto log = logan_lite("policy.rule");
+
+    auto* mitm_proxy = dynamic_cast<MitmProxy*>(new_proxy);
+
     bool ret = true;
-    
     bool cfg_wrt;
 
-    const char* pc_name = "none";
-    const char* pc_global_no = "global_no";
-    const char* pc_global_yes = "global_yes";
-    const char* pc_global = pc_global_no;
-    
     if(mitm_proxy != nullptr) {
         if(pc != nullptr) {
-            pc_name = pc->prof_name.c_str();
-            DIA_("cfgapi_obj_policy_apply: policy content profile[%s]: write payload: %d", pc_name, pc->write_payload);
+            const char* pc_name = pc->prof_name.c_str();
+            _dia("policy_apply: policy content profile[%s]: write payload: %d", pc_name, pc->write_payload);
             mitm_proxy->write_payload(pc->write_payload);
     
-            if(pc->content_rules.size() > 0) {
-                DIA_("cfgapi_obj_policy_apply: policy content profile[%s]: applying content rules, size %d", pc_name, pc->content_rules.size());
+            if( ! pc->content_rules.empty() ) {
+                _dia("policy_apply: policy content profile[%s]: applying content rules, size %d", pc_name, pc->content_rules.size());
                 mitm_proxy->init_content_replace();
                 mitm_proxy->content_replace(pc->content_rules);
             }
         }
         else if(cfgapi.getRoot()["settings"].lookupValue("default_write_payload",cfg_wrt)) {
-            DIA_("cfgapi_obj_policy_apply: global content profile: %d", cfg_wrt);
+            _dia("policy_apply: global content profile: %d", cfg_wrt);
             mitm_proxy->write_payload(cfg_wrt);
-            if(cfg_wrt) {
-                pc_global = pc_global_yes;
-            }
-            
-            pc_name = pc_global;
         }
         
         if(mitm_proxy->write_payload()) {
@@ -1475,7 +1429,7 @@ bool cfgapi_obj_profile_content_apply(baseHostCX* originator, baseProxy* new_pro
             mitm_proxy->tlog()->left_write("Connection start\n");
         }
     } else {
-        WARS_("cfgapi_obj_policy_apply: cannot apply content profile: cast to MitmProxy failed.");
+        _war("policy_apply: cannot apply content profile: cast to MitmProxy failed.");
         ret = false;
     } 
     
@@ -1483,11 +1437,11 @@ bool cfgapi_obj_profile_content_apply(baseHostCX* originator, baseProxy* new_pro
 }
 
 
-bool cfgapi_obj_profile_detect_apply(baseHostCX* originator, baseProxy* new_proxy, ProfileDetection* pd) {
+bool CfgFactory::prof_detect_apply (baseHostCX *originator, baseProxy *new_proxy, ProfileDetection *pd) {
 
-    MitmProxy* mitm_proxy = static_cast<MitmProxy*>(new_proxy); 
-    AppHostCX* mitm_originator = static_cast<AppHostCX*>(originator);
-    
+    auto* mitm_originator = dynamic_cast<AppHostCX*>(originator);
+    auto log = logan_lite("policy.rule");
+
     const char* pd_name = "none";
     bool ret = true;
     
@@ -1496,42 +1450,53 @@ bool cfgapi_obj_profile_detect_apply(baseHostCX* originator, baseProxy* new_prox
         mitm_originator->mode(AppHostCX::MODE_NONE);
         if(pd != nullptr)  {
             pd_name = pd->prof_name.c_str();
-            DIA_("cfgapi_obj_policy_apply[%s]: policy detection profile: mode: %d", pd_name, pd->mode);
-            mitm_originator->mode(pd->mode);
+            _dia("policy_apply[%s]: policy detection profile: mode: %d", pd_name, pd->mode);
+            mitm_originator->mode(static_cast<AppHostCX::mode_t>(pd->mode));
         }
     } else {
-        WARS_("cfgapi_obj_policy_apply: cannot apply detection profile: cast to AppHostCX failed.");
+        _war("policy_apply: cannot apply detection profile: cast to AppHostCX failed.");
         ret = false;
     }    
     
     return ret;
 }
 
-bool cfgapi_obj_profile_tls_apply(baseHostCX* originator, baseProxy* new_proxy, ProfileTls* ps) {
-    
-    MitmProxy* mitm_proxy = static_cast<MitmProxy*>(new_proxy); 
-    AppHostCX* mitm_originator = static_cast<AppHostCX*>(originator);
+bool CfgFactory::prof_tls_apply (baseHostCX *originator, baseProxy *new_proxy, ProfileTls *ps) {
+
+    auto log = logan_lite("policy.rule");
+
+    auto* mitm_proxy = dynamic_cast<MitmProxy*>(new_proxy);
+    auto* mitm_originator = dynamic_cast<AppHostCX*>(originator);
     
     bool tls_applied = false;
-    
+
+    if(! mitm_proxy) {
+        _err("prof_tls_apply: proxy is null");
+        return false;
+    }
+
+    if(! mitm_originator) {
+        _err("prof_tls_apply: originator cx is null");
+        return false;
+    }
+
     if(ps != nullptr) {
         // we should also apply tls profile to originating side! Verification is not in effect, but BYPASS is!
-        if (cfgapi_obj_policy_apply_tls(ps,mitm_originator->com())) {
+        if (policy_apply_tls(ps, mitm_originator->com())) {
             
-            for( cx_iterator i = mitm_proxy->rs().begin(); i != mitm_proxy->rs().end() ; ++i ) {
-                baseHostCX* cx = (*i);
+            for( auto* cx: mitm_proxy->rs()) {
                 baseCom* xcom = cx->com();
                 
-                tls_applied = cfgapi_obj_policy_apply_tls(ps,xcom);
+                tls_applied = policy_apply_tls(ps, xcom);
                 if(!tls_applied) {
-                    ERR_("%s: cannot apply TLS profile to target connection %s",new_proxy->c_name(), cx->c_name());
+                    _err("%s: cannot apply TLS profile to target connection %s",new_proxy->c_name(), cx->c_name());
                 } else {
                     
                     //applying bypass based on DNS cache
                     
-                    SSLCom* sslcom = dynamic_cast<SSLCom*>(xcom);
+                    auto* sslcom = dynamic_cast<SSLCom*>(xcom);
                     if(sslcom && ps->sni_filter_bypass.valid()) {
-                        if(ps->sni_filter_bypass.ptr()->size() > 0 && ps->sni_filter_use_dns_cache) {
+                        if( ( ! ps->sni_filter_bypass.ptr()->empty() ) && ps->sni_filter_use_dns_cache) {
                         
                             bool interrupt = false;
                             for(std::string& filter_element: *ps->sni_filter_bypass) {
@@ -1540,42 +1505,41 @@ bool cfgapi_obj_profile_tls_apply(baseHostCX* originator, baseProxy* new_proxy, 
                                 
                                 if(f.match(c)) {
                                     if(sslcom->bypass_me_and_peer()) {
-                                        INF_("Connection %s bypassed: IP in DNS cache matching TLS bypass list (%s).",originator->full_name('L').c_str(),filter_element.c_str());
+                                        _inf("Connection %s bypassed: IP in DNS cache matching TLS bypass list (%s).",originator->full_name('L').c_str(),filter_element.c_str());
                                         interrupt = true;
                                     } else {
-                                        WAR_("Connection %s: cannot be bypassed.",originator->full_name('L').c_str());
+                                        _war("Connection %s: cannot be bypassed.",originator->full_name('L').c_str());
                                     }
                                 } else if (ps->sni_filter_use_dns_domain_tree) {
-                                    domain_cache.lock();
-                                    //INF_("FQDN doesn't match SNI element, looking for sub-domains of %s", filter_element.c_str());
-                                    //FIXME: we assume flter is 2nd level domain... 
+                                    std::scoped_lock<std::recursive_mutex> dd_(DNS::get_domain_lock());
+
+                                    //_inf("FQDN doesn't match SNI element, looking for sub-domains of %s", filter_element.c_str());
+                                    //FIXME: we assume filter is 2nd level domain...
                                     
-                                    auto subdomain_cache = domain_cache.get(filter_element);
+                                    auto subdomain_cache = DNS::get_domain_cache().get(filter_element);
                                     if(subdomain_cache != nullptr) {
-                                        for(auto subdomain: subdomain_cache->cache()) {
+                                        for(auto const& subdomain: subdomain_cache->cache()) {
                                             
                                             std::vector<std::string> prefix_n_domainname = string_split(subdomain.first,':');
                                             if(prefix_n_domainname.size() < 2) continue; // don't continue if we can't strip A: or AAAA:
                                             
-                                            FqdnAddress f(prefix_n_domainname.at(1)+"."+filter_element);
-                                            DEB_("Connection %s: subdomain check: test if %s matches %s",originator->full_name('L').c_str(),f.to_string().c_str(),xcom->owner_cx()->host().c_str());
+                                            FqdnAddress ff(prefix_n_domainname.at(1)+"."+filter_element);
+                                            _deb("Connection %s: subdomain check: test if %s matches %s",originator->full_name('L').c_str(),ff.to_string().c_str(),xcom->owner_cx()->host().c_str());
                                             
-                                            if(f.match(c)) {
+                                            if(ff.match(c)) {
                                                 if(sslcom->bypass_me_and_peer()) {
-                                                    INF_("Connection %s bypassed: IP in DNS subdomain cache matching TLS bypass list (%s).",originator->full_name('L').c_str(),filter_element.c_str());
+                                                    _inf("Connection %s bypassed: IP in DNS sub-domain cache matching TLS bypass list (%s).",originator->full_name('L').c_str(),filter_element.c_str());
                                                 } else {
-                                                    WAR_("Connection %s: cannot be bypassed.",originator->full_name('L').c_str());                                                    
+                                                    _war("Connection %s: cannot be bypassed.",originator->full_name('L').c_str());
                                                 }
                                                 interrupt = true; //exit also from main loop
                                                 break;
                                             }
                                         }
                                     }
-                                    
-                                    domain_cache.unlock();
                                 }
                                 
-                                delete c;
+                                cidr_free(c);
                                 
                                 if(interrupt) 
                                     break;
@@ -1591,17 +1555,19 @@ bool cfgapi_obj_profile_tls_apply(baseHostCX* originator, baseProxy* new_proxy, 
     return tls_applied;
 }
 
-bool cfgapi_obj_alg_dns_apply(baseHostCX* originator, baseProxy* new_proxy, ProfileAlgDns* p_alg_dns) {
-    
-    AppHostCX* mitm_originator = static_cast<AppHostCX*>(originator);    
-    MitmHostCX* mh = dynamic_cast<MitmHostCX*>(mitm_originator);
+bool CfgFactory::prof_alg_dns_apply (baseHostCX *originator, baseProxy *new_proxy, ProfileAlgDns *p_alg_dns) {
+
+    auto log = logan_lite("policy.rule");
+
+    auto* mitm_originator = dynamic_cast<AppHostCX*>(originator);
+    auto* mh = dynamic_cast<MitmHostCX*>(mitm_originator);
 
     bool ret = false;
     
     if(mh != nullptr) {
 
         if(p_alg_dns != nullptr) {
-            DNS_Inspector* n = new DNS_Inspector();
+            auto* n = new DNS_Inspector();
             if(n->l4_prefilter(mh)) {
                 n->opt_match_id = p_alg_dns->match_request_id;
                 n->opt_randomize_id = p_alg_dns->randomize_id;
@@ -1615,66 +1581,72 @@ bool cfgapi_obj_alg_dns_apply(baseHostCX* originator, baseProxy* new_proxy, Prof
         }
         
     } else {
-        NOT_("Connection %s cannot be inspected by ALGs",originator->full_name('L').c_str());
+        _not("Connection %s cannot be inspected by ALGs",originator->full_name('L').c_str());
     }    
     
     return ret;
 }
 
-int cfgapi_obj_policy_apply(baseHostCX* originator, baseProxy* new_proxy) {
-    std::lock_guard<std::recursive_mutex> l(cfgapi_write_lock);
+int CfgFactory::policy_apply (baseHostCX *originator, baseProxy *proxy) {
+
+    auto log = logan_lite("policy.rule");
+    std::lock_guard<std::recursive_mutex> l(lock_);
     
-    int policy_num = cfgapi_obj_policy_match(new_proxy);
-    int verdict = cfgapi_obj_policy_action(policy_num);
+    int policy_num = policy_match(proxy);
+    int verdict = policy_action(policy_num);
     if(verdict == POLICY_ACTION_PASS) {
 
-        ProfileContent* pc  = cfgapi_obj_policy_profile_content(policy_num);
-        ProfileDetection* pd = cfgapi_obj_policy_profile_detection(policy_num);
-        ProfileTls* pt = cfgapi_obj_policy_profile_tls(policy_num);
-        ProfileAuth* pa = cfgapi_obj_policy_profile_auth(policy_num);
-        ProfileAlgDns* p_alg_dns  = cfgapi_obj_policy_profile_alg_dns(policy_num);
-        
-        
-        const char* pc_name = "none";
-        const char* pd_name = "none";
-        const char* pt_name = "none";
-        const char* pa_name = "none";
-        
+        ProfileContent *pc = policy_prof_content(policy_num);
+        ProfileDetection *pd = policy_prof_detection(policy_num);
+        ProfileTls *pt = policy_prof_tls(policy_num);
+        ProfileAuth *pa = policy_prof_auth(policy_num);
+        ProfileAlgDns *p_alg_dns = policy_prof_alg_dns(policy_num);
+
+
+        const char *pc_name = "none";
+        const char *pd_name = "none";
+        const char *pt_name = "none";
+        const char *pa_name = "none";
+
         //Algs will be list of single letter abreviations
         // DNS alg: D
-        std::string algs_name = ""; 
-        
+        std::string algs_name;
+
         /* Processing content profile */
-        if(pc)
-        if(cfgapi_obj_profile_content_apply(originator,new_proxy,pc)) {
-            pc_name = pc->prof_name.c_str();
+        if (pc) {
+            if (prof_content_apply(originator, proxy, pc)) {
+                pc_name = pc->prof_name.c_str();
+            }
         }
         
         
         /* Processing detection profile */
-        if(pd)
-        if(cfgapi_obj_profile_detect_apply(originator,new_proxy,pd)) {
-            pd_name = pd->prof_name.c_str();
-        }        
+        if (pd) {
+            if(prof_detect_apply(originator, proxy, pd)) {
+                pd_name = pd->prof_name.c_str();
+            }
+        }
         
         /* Processing TLS profile*/
-        if(pt)
-        if(cfgapi_obj_profile_tls_apply(originator,new_proxy,pt)) {
-            pt_name = pt->prof_name.c_str();
-        }        
+        if (pt) {
+            if(prof_tls_apply(originator, proxy, pt)) {
+                pt_name = pt->prof_name.c_str();
+            }
+        }
         
         /* Processing ALG : DNS*/
-        if(p_alg_dns)
-        if(cfgapi_obj_alg_dns_apply(originator,new_proxy,p_alg_dns)) {
-            algs_name += p_alg_dns->prof_name.c_str();
-        }        
+        if (p_alg_dns) {
+            if (prof_alg_dns_apply(originator, proxy, p_alg_dns)) {
+                algs_name += p_alg_dns->prof_name;
+            }
+        }
 
         
-        MitmProxy* mitm_proxy = static_cast<MitmProxy*>(new_proxy); 
-        AppHostCX* mitm_originator = static_cast<AppHostCX*>(originator);
+        auto* mitm_proxy = dynamic_cast<MitmProxy*>(proxy);
+        auto* mitm_originator = dynamic_cast<AppHostCX*>(originator);
         
         /* Processing Auth profile */
-        if(pa) {
+        if(pa && mitm_proxy) {
             // auth is applied on proxy
             mitm_proxy->opt_auth_authenticate = pa->authenticate;
             mitm_proxy->opt_auth_resolve = pa->resolve;
@@ -1685,66 +1657,68 @@ int cfgapi_obj_policy_apply(baseHostCX* originator, baseProxy* new_proxy) {
         // ALGS can operate only on MitmHostCX classes
 
         
-        INF_("Connection %s accepted: policy=%d cont=%s det=%s tls=%s auth=%s algs=%s",originator->full_name('L').c_str(),policy_num,pc_name,pd_name,pt_name,pa_name,algs_name.c_str());
+        _inf("Connection %s accepted: policy=%d cont=%s det=%s tls=%s auth=%s algs=%s",originator->full_name('L').c_str(),policy_num,pc_name,pd_name,pt_name,pa_name,algs_name.c_str());
         
     } else {
-        INF_("Connection %s denied: policy=%d",originator->full_name('L').c_str(),policy_num);
+        _inf("Connection %s denied: policy=%d",originator->full_name('L').c_str(),policy_num);
     }
     
     return policy_num;
 }
 
 
-bool cfgapi_obj_policy_apply_tls(int policy_num, baseCom* xcom) {
-    ProfileTls* pt = cfgapi_obj_policy_profile_tls(policy_num);
-    return cfgapi_obj_policy_apply_tls(pt,xcom);
+bool CfgFactory::policy_apply_tls (int policy_num, baseCom *xcom) {
+    ProfileTls* pt = policy_prof_tls(policy_num);
+    return policy_apply_tls(pt, xcom);
 }
 
-bool should_redirect_warning_port(ProfileTls* pt, SSLCom* com) {
+bool CfgFactory::should_redirect (ProfileTls *pt, SSLCom *com) {
+
+    auto log = logan_lite("policy.rule");
     
     bool ret = false;
     
-    DEB_("should_redirect_warning_port[%s]",com->hr());
+    _deb("should_redirect[%s]",com->hr());
     
     if(com && com->owner_cx()) {
         
         try {
             int num_port = std::stoi(com->owner_cx()->port());
-            DEB_("should_redirect_warning_port[%s]: owner port %d",com->hr(), num_port);
+            _deb("should_redirect[%s]: owner port %d",com->hr(), num_port);
             
             
             if(pt->redirect_warning_ports.ptr()) {
                 // we have port redirection list (which ports should be redirected/replaced for cert issue warning)
-                DEB_("should_redirect_warning_port[%s]: checking port list present",com->hr());
+                _deb("should_redirect[%s]: checking port list present",com->hr());
                 
                 auto it = pt->redirect_warning_ports.ptr()->find(num_port);
                 
                 if(it != pt->redirect_warning_ports.ptr()->end()) {
-                    DIA_("should_redirect_warning_port[%s]: port %d in redirect list",com->hr(),num_port);
+                    _dia("should_redirect[%s]: port %d allowed to be redirected if needed",com->hr(),num_port);
                     ret = true;
                 }
             }
             else {
                 // if we have list empty (uninitialized), we assume only 443 should be redirected
                 if(num_port == 443) {
-                    DEB_("should_redirect_warning_port[%s]: implicit 443 redirection allowed (no port list)",com->hr());
+                    _deb("should_redirect[%s]: implicit 443 redirection allowed (no port list)",com->hr());
                     ret = true;
                 }
             }
         }
-        catch(std::invalid_argument) {}
-        catch(std::out_of_range) {}
+        catch(std::invalid_argument& ) {}
+        catch(std::out_of_range& ) {}
     }
     
     return ret;
 }
 
-bool cfgapi_obj_policy_apply_tls(ProfileTls* pt, baseCom* xcom) {
+bool CfgFactory::policy_apply_tls (ProfileTls *pt, baseCom *xcom) {
 
     bool tls_applied = false;     
     
     if(pt != nullptr) {
-        SSLCom* sslcom = dynamic_cast<SSLCom*>(xcom);
+        auto* sslcom = dynamic_cast<SSLCom*>(xcom);
         if(sslcom != nullptr) {
             sslcom->opt_bypass = !pt->inspect;
             sslcom->opt_allow_unknown_issuer = pt->allow_untrusted_issuers;
@@ -1752,13 +1726,13 @@ bool cfgapi_obj_policy_apply_tls(ProfileTls* pt, baseCom* xcom) {
             sslcom->opt_allow_not_valid_cert = pt->allow_invalid_certs;
             sslcom->opt_allow_self_signed_cert = pt->allow_self_signed;
 
-            SSLCom* peer_sslcom = dynamic_cast<SSLCom*>(sslcom->peer());
+            auto* peer_sslcom = dynamic_cast<SSLCom*>(sslcom->peer());
 
             if( peer_sslcom &&
                     pt->failed_certcheck_replacement &&
-                    should_redirect_warning_port(pt,peer_sslcom)) {
+                    should_redirect(pt, peer_sslcom)) {
 
-                DEB_("cfgapi_obj_policy_apply_tls: applying profile, repl=%d, repl_ovrd=%d, repl_ovrd_tmo=%d, repl_ovrd_tmo_type=%d",
+                _deb("policy_apply_tls: applying profile, repl=%d, repl_ovrd=%d, repl_ovrd_tmo=%d, repl_ovrd_tmo_type=%d",
                      pt->failed_certcheck_replacement,
                      pt->failed_certcheck_override,
                      pt->failed_certcheck_override_timeout,
@@ -1785,7 +1759,7 @@ bool cfgapi_obj_policy_apply_tls(ProfileTls* pt, baseCom* xcom) {
             sslcom->opt_ocsp_stapling_mode = pt->ocsp_stapling_mode;
        
             if(pt->sni_filter_bypass.valid()) {
-                if(pt->sni_filter_bypass.ptr()->size() > 0) {
+                if( ! pt->sni_filter_bypass.ptr()->empty() ) {
                     sslcom->sni_filter_to_bypass().ref(pt->sni_filter_bypass);
                 }
             }
@@ -1800,67 +1774,66 @@ bool cfgapi_obj_policy_apply_tls(ProfileTls* pt, baseCom* xcom) {
 }
 
 
-void cfgapi_cleanup()
+void CfgFactory::cfgapi_cleanup()
 {
-  cfgapi_cleanup_obj_policy();
-  cfgapi_cleanup_obj_address();
-  cfgapi_cleanup_obj_port();
-  cfgapi_cleanup_obj_proto();
-  cfgapi_cleanup_obj_profile_content();
-  cfgapi_cleanup_obj_profile_detection();
-  cfgapi_cleanup_obj_profile_tls();
-  cfgapi_cleanup_obj_profile_auth();
-  cfgapi_cleanup_obj_profile_alg_dns();
+    cleanup_db_policy();
+    cleanup_db_address();
+    cleanup_db_port();
+    cleanup_db_proto();
+    cleanup_db_prof_content();
+    cleanup_db_prof_detection();
+    cleanup_db_prof_tls();
+    cleanup_db_prof_auth();
+    cleanup_db_prof_alg_dns();
 }
 
 
-void cfgapi_log_version(bool warn_delay)
+void CfgFactory::log_version (bool warn_delay)
 {
-    CRI_("Starting Smithproxy %s (socle %s)",SMITH_VERSION,SOCLE_VERSION);
+    _cri("Starting Smithproxy %s (socle %s)",SMITH_VERSION,SOCLE_VERSION);
     
     if(SOCLE_DEVEL || SMITH_DEVEL) {
-        WARS_("");
+        _war("");
         if(SOCLE_DEVEL) {
-            WAR_("Socle library version %s (dev)",SOCLE_VERSION);
+            _war("Socle library version %s (dev)",SOCLE_VERSION);
         }
 #ifdef SOCLE_MEM_PROFILE
-        WARS_("*** PERFORMANCE: Socle library has extra memory profiling enabled! ***");
+        _war("*** PERFORMANCE: Socle library has extra memory profiling enabled! ***");
 #endif
         if(SMITH_DEVEL) {
-            WAR_("Smithproxy version %s (dev)",SMITH_VERSION);
+            _war("Smithproxy version %s (dev)",SMITH_VERSION);
         }        
-        WARS_("");
+        _war("");
         
         if(warn_delay) {
-            WARS_("  ... start will continue in 3 sec.");
+            _war("  ... start will continue in 3 sec.");
             sleep(3);
         }
     }
 }
 
-int apply_index(std::string& what , const std::string& idx) {
-    DEB_("apply_index: what=%s idx=%s",what.c_str(),idx.c_str());
+int CfgFactory::apply_tenant_index(std::string& what, int& idx) {
+    _deb("apply_index: what=%s idx=%d",what.c_str(),idx);
     int port = std::stoi(what);
-    int index = std::stoi(idx);
-    what = std::to_string(port + index);
+    what = std::to_string(port + idx);
 
     return 0;
 }
 
 
-bool cfgapi_apply_tenant_config () {
+bool CfgFactory::apply_tenant_config () {
     int ret = 0;
 
-    if(cfg_tenant_index.size() > 0 && cfg_tenant_name.size() > 0) {
-        ret += apply_index(cfg_tcp_listen_port,cfg_tenant_index);
-        ret += apply_index(cfg_ssl_listen_port,cfg_tenant_index);
-        ret += apply_index(cfg_dtls_port,cfg_tenant_index);
-        ret += apply_index(cfg_udp_port,cfg_tenant_index);
-        ret += apply_index(cfg_socks_port,cfg_tenant_index);
-        ret += apply_index(cfgapi_identity_portal_port_http,cfg_tenant_index);
-        ret += apply_index(cfgapi_identity_portal_port_https,cfg_tenant_index);
+    if( (  tenant_index >= 0 ) && ( ! tenant_name.empty() ) ) {
+        ret += apply_tenant_index(listen_tcp_port, tenant_index);
+        ret += apply_tenant_index(listen_tls_port, tenant_index);
+        ret += apply_tenant_index(listen_dtls_port, tenant_index);
+        ret += apply_tenant_index(listen_udp_port, tenant_index);
+        ret += apply_tenant_index(listen_socks_port, tenant_index);
+        ret += apply_tenant_index(AuthFactory::get().portal_port_http, tenant_index);
+        ret += apply_tenant_index(AuthFactory::get().portal_port_https, tenant_index);
 
-        cli_port += std::stoi(cfg_tenant_index);
+        cli_port += tenant_index;
     }
 
     return (ret == 0);
