@@ -53,47 +53,41 @@
 #include <sys/resource.h>
 #include <execinfo.h>
 
-#include <logger.hpp>
+#include <log/logger.hpp>
 #include <daemon.hpp>
 #include <display.hpp>
 
 #define UNW_LOCAL_ONLY
 #include <libunwind.h>
 
-#ifdef __cplusplus
-extern "C" {
-#endif
 
-    
-std::string PID_FILE(PID_FILE_DEFAULT);
-
-void daemon_set_tenant(const std::string& name, const std::string& tenant_id) {
+void DaemonFactory::set_tenant(const std::string& name, const std::string& tenant_id) {
     PID_FILE = string_format("/var/run/%s.%s.pid",name.c_str(), tenant_id.c_str());
 }
 
-void daemonize(void) {
+void DaemonFactory::daemonize() {
         
     /* Our process ID and Session ID */
     pid_t pid, sid;
     
-    DIAS_("daemonize start");
+    _dia("daemonize start");
     
     struct stat st;
     if( stat(PID_FILE.c_str(),&st) == 0) {
-        ERRS_("There seems to be smithproxy already running in the system. Aborting.");
+        _err("There seems to be smithproxy already running in the system. Aborting.");
         exit(EXIT_FAILURE);
     }
     
     /* Fork off the parent process */
     pid = fork();
     if (pid < 0) {
-            FATS_("daemonize: failed to fork!");
+            _fat("daemonize: failed to fork!");
             exit(EXIT_FAILURE);
     }
     /* If we got a good PID, then
         we can exit the parent process. */
     if (pid > 0) {
-            DIAS_("daemonize: exiting from master");
+            _dia("daemonize: exiting from master");
             exit(EXIT_SUCCESS);
     }
 
@@ -106,7 +100,7 @@ void daemonize(void) {
     sid = setsid();
     if (sid < 0) {
             /* Log the failure */
-            FATS_("daemonize: failed to setsid!");
+            _fat("daemonize: failed to setsid!");
             exit(EXIT_FAILURE);
     }
     
@@ -115,7 +109,7 @@ void daemonize(void) {
     /* Change the current working directory */
     if ((chdir("/")) < 0) {
             /* Log the failure */
-            FATS_("daemonize: failed to chdir to '/'!");
+            _fat("daemonize: failed to chdir to '/'!");
             exit(EXIT_FAILURE);
     }
     
@@ -127,37 +121,37 @@ void daemonize(void) {
     close(STDOUT_FILENO);
     close(STDERR_FILENO);
         
-    DIAS_("daemonize: finished");
+    _dia("daemonize: finished");
 }
 
-void daemon_write_pidfile() {
+void DaemonFactory::write_pidfile() {
     FILE* pf = fopen(PID_FILE.c_str(),"w");
     fprintf(pf,"%d",getpid());
     fclose(pf);
 }
 
-void daemon_unlink_pidfile() {
+void DaemonFactory::unlink_pidfile() {
     unlink(PID_FILE.c_str());
 }
 
-bool daemon_exists_pidfile() {
+bool DaemonFactory::exists_pidfile() {
     struct stat st;
     int result = stat(PID_FILE.c_str(), &st);
     return result == 0;
 }
 
-int daemon_get_limit_fd() {
+int DaemonFactory::get_limit_fd() {
     struct rlimit r;
     int ret = getrlimit(RLIMIT_NOFILE,&r);
     if(ret < 0) {
-        ERR_("daemon_get_limit_fd: cannot obtain fd limits: %s", string_error().c_str());
+        _err("get_limit_fd: cannot obtain fd limits: %s", string_error().c_str());
         return -1;
     }
 
     return r.rlim_cur;
 }
 
-void daemon_set_limit_fd(int max) {
+void DaemonFactory::set_limit_fd(int max) {
     int n = max;
     if(max == 0) 
         n = 100000;
@@ -169,29 +163,36 @@ void daemon_set_limit_fd(int max) {
     int ret = setrlimit(RLIMIT_NOFILE,&r);
 
     if(ret < 0) {
-        ERR_("daemon_set_limit_fd: cannot set fd limits: %s", string_error().c_str());
+        _err("set_limit_fd: cannot set fd limits: %s", string_error().c_str());
     }
 }
 
-void set_signal(unsigned int SIG, void (*sig_handler)(int)) {
+void DaemonFactory::set_signal(unsigned int SIG, void (*sig_handler)(int)) {
     struct sigaction act_segv;
     sigemptyset(&act_segv.sa_mask);
     act_segv.sa_flags = 0;
     
     if(sig_handler != nullptr)  act_segv.sa_handler = sig_handler;
     
-    sigaction( SIG, &act_segv, NULL);
+    sigaction( SIG, &act_segv, nullptr);
 }
 
-#define LOG_FILENAME_SZ 512
-volatile char crashlog_file[LOG_FILENAME_SZ];
 
-void set_crashlog(const char* file) {
+void DaemonFactory::set_crashlog(const char* file) {
     memset((void*)crashlog_file,0,LOG_FILENAME_SZ);
     strncpy((char*)crashlog_file,file,LOG_FILENAME_SZ-1);
 }
 
-static void uw_btrace_handler(int sig) {
+void writecrash(int fd, const char* msg, int len)  {
+   int w = 0; int rep = 0;
+   do {
+       int curw = ::write(fd, msg, len);
+       if(curw > 0) w+= w;
+       rep++;
+   } while(w < len && rep < 10);
+}
+
+void DaemonFactory::uw_btrace_handler(int sig) {
     thread_local unw_cursor_t cursor; 
     thread_local unw_context_t uc;
     thread_local unw_word_t ip;
@@ -200,14 +201,16 @@ static void uw_btrace_handler(int sig) {
     unw_getcontext(&uc);
     unw_init_local(&cursor, &uc);
 
+    DaemonFactory& df = DaemonFactory::instance();
+
     char buf_line[256];
     int chars = snprintf(buf_line,255," ======== Smithproxy exception handler (sig %d) =========\n",sig);
 
-    int CRLOG = open((const char*)crashlog_file,O_CREAT | O_WRONLY | O_TRUNC,S_IRUSR|S_IWUSR);
-    write(STDERR_FILENO,buf_line,chars);
-    write(CRLOG,buf_line,chars);
-    write(STDERR_FILENO,"Traceback:\n",11);
-    write(CRLOG,"Traceback:\n",11);
+    int CRLOG = open((const char*)df.crashlog_file, O_CREAT | O_WRONLY | O_TRUNC,S_IRUSR|S_IWUSR);
+    writecrash(STDERR_FILENO,buf_line,chars);
+    writecrash(CRLOG,buf_line,chars);
+    writecrash(STDERR_FILENO,"Traceback:\n",11);
+    writecrash(CRLOG,"Traceback:\n",11);
 
     while (unw_step(&cursor) > 0) {
         memset(buf_line,0,256);
@@ -219,22 +222,22 @@ static void uw_btrace_handler(int sig) {
         unw_get_reg(&cursor, UNW_REG_IP, &ip);
         unw_get_reg(&cursor, UNW_REG_SP, &sp);
         
-        int chars = snprintf(buf_line, 255, "ip = %lx, sp = %lx: (%s+0x%x) [%p]\n", (long) ip, (unsigned long) sp, buf_fun, (unsigned int) offset, (void*)ip);
+        chars = snprintf(buf_line, 255, "ip = %lx, sp = %lx: (%s+0x%x) [%p]\n", (long) ip, (unsigned long) sp, buf_fun, (unsigned int) offset, (void*)ip);
 
-        write(CRLOG,buf_line,chars);
-        write(STDERR_FILENO,buf_line,chars);
+        writecrash(CRLOG,buf_line,chars);
+        writecrash(STDERR_FILENO,buf_line,chars);
     }
-    
-    write(STDERR_FILENO," ===============================================\n",50);
-    write(CRLOG," ===============================================\n",50);
+
+    writecrash(STDERR_FILENO," ===============================================\n",50);
+    writecrash(CRLOG," ===============================================\n",50);
     close(CRLOG);
-    
-    daemon_unlink_pidfile();
+
+    df.unlink_pidfile();
     
     exit(-1);
 }
 
-void set_daemon_signals(void (*terminate_handler)(int),void (*reload_handler)(int)) {
+void DaemonFactory::set_daemon_signals(void (*terminate_handler)(int),void (*reload_handler)(int)) {
     // install signal handler, we do want to release the memory properly
         // signal handler installation
     
@@ -247,6 +250,12 @@ void set_daemon_signals(void (*terminate_handler)(int),void (*reload_handler)(in
     set_signal(SIGSEGV,uw_btrace_handler);
 }
 
-#ifdef __cplusplus
+std::string& DaemonFactory::class_name() const {
+    static std::string s("DaemonFactory");
+    return s;
 }
-#endif
+
+std::string DaemonFactory::hr() const {
+    return string_format("pidfile=(%s)", PID_FILE.c_str());
+}
+
