@@ -38,89 +38,77 @@
 
     """
 
-import traceback
-import time
-import mmap
-import os
-import sys
-import struct
-import socket
 import logging
+import struct
 
 from shmtable import ShmTable
-
 
 flog = logging.getLogger('bend')
 
 
 class TokenTable(ShmTable):
     def __init__(self):
-        ShmTable.__init__(self,576)
-        self.tokens = {}   # token => url
+        ShmTable.__init__(self, 576)
+        self.tokens = {}  # token => url
         self.used_tokens = []  # used throw here - delete when appropriate
-        self.active_queue = []    # add everything here. After size grows to some point, start
-                               # deleting also active unused yed tokens
-    
-        self.delete_used_threshold =   5   # 1 means immediately
+        self.active_queue = []  # add everything here. After size grows to some point, start
+        # deleting also active unused yed tokens
+
+        self.delete_used_threshold = 5  # 1 means immediately
         self.delete_active_threshold = 200  # mark oldest tokens above this margin as used
-    
+        self.normalize = False
+
     def on_new_table(self):
         ShmTable.on_new_table(self)
         self.tokens = {}
 
-    def on_new_entry(self,blob):
-        ShmTable.on_new_entry(self,blob)
-        t,u = struct.unpack('64s512s',blob)
-        
-        t_i = t.find('\x00',0,512)
-        u_i = u.find('\x00',0,512)
+    def on_new_entry(self, blob):
+        ShmTable.on_new_entry(self, blob)
+        t, u = struct.unpack('64s512s', blob)
+
+        t_i = t.find('\x00', 0, 512)
+        u_i = u.find('\x00', 0, 512)
         tt = t[:t_i]
         uu = u[:u_i]
-        
+
         flog.debug("TokenTable::on_new_entry: " + tt + ":" + uu)
         self.tokens[tt] = uu
         self.life_queue(tt)
-        
-    def toggle_used(self,token):
+
+    def toggle_used(self, token):
         self.used_tokens.append(token)
         if len(self.used_tokens) > self.delete_used_threshold:
-            
+
             # delete all used tokens from DB
             for t in self.used_tokens:
                 flog.debug("toggle_used: wiping used token " + t)
-                self.tokens.pop(t,None)
-                
+                self.tokens.pop(t, None)
+
             self.used_tokens = []
             self.save(True)
-            
 
-    def life_queue(self,token):
+    def life_queue(self, token):
         self.active_queue.append(token)
-        
+
         while len(self.active_queue) > self.delete_active_threshold:
             oldest_token = self.active_queue[0]
             flog.debug("life_queue: too many active tokens, dropping oldest one " + oldest_token)
             self.toggle_used(oldest_token)
             self.active_queue = self.active_queue[1:]
-        
-    
+
     def save(self, inc_version=False):
         self.seek(0)
         self.clear()
-        self.write_header(inc_version,len(self.tokens.keys()))
-        
+        self.write_header(inc_version, len(self.tokens.keys()))
+
         write_cnt = 0
         for k in self.tokens.keys():
-              try:
-                  self.write(struct.pack("64s512s",k,self.tokens[k]))
-                  write_cnt = write_cnt + 1
-              except IndexError:
-                  continue
+            try:
+                self.write(struct.pack("64s512s", k, self.tokens[k]))
+                write_cnt = write_cnt + 1
+            except IndexError:
+                continue
 
         flog.debug("save: %d tokens written to table" % (write_cnt,))
 
-        
-        self.normalize = False # each dump effectively normalizes db
-        
-
-
+        self.normalize = False  # each dump effectively normalizes db
