@@ -123,7 +123,7 @@ std::pair<std::string,std::string> Inspector::split_fqdn_subdomain(std::string& 
 
 std::regex DNS_Inspector::wildcard = std::regex("[^.]+\\.(.*)$");
 
-bool DNS_Inspector::interested(AppHostCX* cx) {
+bool DNS_Inspector::interested(AppHostCX* cx) const {
     if(cx->com()->nonlocal_dst_port() == 53)
         return true;
     return false;
@@ -135,7 +135,7 @@ void DNS_Inspector::update(AppHostCX* cx) {
     
     
     duplexFlow& f = cx->flow();
-    DIA___("DNS_Inspector::update[%s]: stage %d start (flow size %d, last flow entry data length %d)",cx->c_name(),stage, f.flow().size(),f.flow().back().second->size());
+    _dia("DNS_Inspector::update[%s]: stage %d start (flow size %d, last flow entry data length %d)",cx->c_name(),stage, f.flow().size(),f.flow().back().second->size());
     
     /* INIT */
     
@@ -153,7 +153,7 @@ void DNS_Inspector::update(AppHostCX* cx) {
     
     DNS_Packet* ptr = nullptr;
     buffer *xbuf = cur_pos.second;
-    buffer buf = xbuf->view(0,xbuf->size());
+    buffer shallow_xbuf = xbuf->view(0, xbuf->size());
 
     // check if response is already available
     DNS_Response* cached_entry = nullptr;
@@ -162,89 +162,89 @@ void DNS_Inspector::update(AppHostCX* cx) {
     unsigned int red = 0;
     
     if(is_tcp) {
-        unsigned short data_size = ntohs(buf.get_at<unsigned short>(0));
-        if(buf.size() < data_size) {
-            DIA___("DNS_Inspector::update[%s]: not enough DNS data in TCP stream: expected %d, but having %d. Waiting to more.",cx->c_name(), data_size, buf.size());
+        unsigned short data_size = ntohs(shallow_xbuf.get_at<unsigned short>(0));
+        if(shallow_xbuf.size() < data_size) {
+            _dia("DNS_Inspector::update[%s]: not enough DNS data in TCP stream: expected %d, but having %d. Waiting to more.", cx->c_name(), data_size, shallow_xbuf.size());
             return;
         }
         red += 2;
     }
     
-    int mem_len = buf.size();
+    int mem_len = shallow_xbuf.size();
     switch(cur_pos.first)  {
         case 'r':
             stage = 0;
-            for(unsigned int it = 0; red < buf.size() && it < 10; it++) {
+            for(unsigned int it = 0; red < shallow_xbuf.size() && it < 10; it++) {
                 ptr = new DNS_Request();
-                buffer cur_buf = buf.view(red,buf.size()-red);
+                buffer cur_buf = shallow_xbuf.view(red, shallow_xbuf.size() - red);
                 int cur_red = ptr->load(&cur_buf);
                 
                 // because of non-standard return value from above load(), we need to adjust red bytes manually
                 if(cur_red == 0) { cur_red = cur_buf.size(); }
                 
-                DIA___("DNS_Inspector::update[%s]: red  %d, load returned %d", cx->c_name(), red, cur_red);
-                DEB___("DNS_Inspector::update[%s]: flow: %s", cx->c_name(), cx->flow().hr().c_str());
+                _dia("DNS_Inspector::update[%s]: red  %d, load returned %d", cx->c_name(), red, cur_red);
+                _deb("DNS_Inspector::update[%s]: flow: %s", cx->c_name(), cx->flow().hr().c_str());
                 
                 // on success write to requests_
                 if(cur_red >= 0) {
                     red += cur_red;
                     
                     if(requests_[ptr->id()] != nullptr) {
-                        INF___("DNS_Inspector::update[%s]: detected re-sent request",cx->c_name());
+                        _not("DNS_Inspector::update[%s]: detected re-sent request",cx->c_name());
                         delete requests_[ptr->id()];
                         requests_.erase(ptr->id());
                     }
                     
-                    DIA___("DNS_Inspector::update[%s]: adding key 0x%x red=%d, buffer_size=%d, ptr=0x%x",cx->c_name(),ptr->id(),red,cur_buf.size(),ptr);
+                    _dia("DNS_Inspector::update[%s]: adding key 0x%x red=%d, buffer_size=%d, ptr=0x%x",cx->c_name(),ptr->id(),red,cur_buf.size(),ptr);
                     requests_[ptr->id()] = (DNS_Request*)ptr;
                     
-                    DEB___("DNS_Inspector::update[%s]: this 0x%x, requests size %d",cx->c_name(),this, requests_.size());
+                    _deb("DNS_Inspector::update[%s]: this 0x%x, requests size %d",cx->c_name(),this, requests_.size());
                     
                     cx->idle_delay(30);
                 } else {
                     red = 0;
                     delete ptr;
                     ptr = (DNS_Packet*)0xCABA1A;
-                    ERR_("BUG CAUGHT: buffer:\n%s",hex_dump(cur_buf).c_str());
+                    _err("BUG CAUGHT: buffer:\n%s",hex_dump(cur_buf).c_str());
                 }
                 
                 // on failure or last data exit loop
                 if(cur_red <= 0) {
-                    DIA___("DNS_Inspector::update[%s]: finishing reading from buffers: red=%d, buffer_size=%d",cx->c_name(),red,cur_buf.size());
+                    _dia("DNS_Inspector::update[%s]: finishing reading from buffers: red=%d, buffer_size=%d",cx->c_name(),red,cur_buf.size());
                     break;
                 }
             }
             
             if(ptr == (DNS_Packet*)0xCABA1A) {
-	       ERRS_("BUG CAUGHT.");
-	       
-	       
-	       goto fail;
-	    }
+	            _err("BUG CAUGHT.");
+
+	            goto fail;
+	        }
             
             if(opt_cached_responses && ( ((DNS_Request*)ptr)->question_type_0() == A || ((DNS_Request*)ptr)->question_type_0() == AAAA ) ) {
-                inspect_dns_cache.lock();
-                cached_entry = inspect_dns_cache.get(ptr->question_str_0());
+                std::scoped_lock<std::recursive_mutex> l_(DNS::get_dns_lock());
+
+                cached_entry = DNS::get_dns_cache().get(ptr->question_str_0());
                 if(cached_entry != nullptr) {
-                    DIA___("DNS answer for %s is already in the cache",cached_entry->question_str_0().c_str());
+                    _dia("DNS answer for %s is already in the cache",cached_entry->question_str_0().c_str());
 
                     
                     if(cached_entry->cached_packet != nullptr) {
                         
                         // do TTL check
-                        DIAS___("cached entry TTL check");
+                        _dia("cached entry TTL check");
                         
                         time_t now = time(nullptr);
                         bool ttl_check = true;
                         
                         for(auto idx: cached_entry->answer_ttl_idx) {
                             uint32_t ttl = ntohl(cached_entry->cached_packet->get_at<uint32_t>(idx));
-                            DEB___("cached response ttl byte index %d value %d",idx,ttl);
+                            _deb("cached response ttl byte index %d value %d",idx,ttl);
                             if(now > ttl + cached_entry->loaded_at) {
-                                DEB___("  %ds -- expired", now - (ttl + cached_entry->loaded_at));
+                                _deb("  %ds -- expired", now - (ttl + cached_entry->loaded_at));
                                 ttl_check = false;
                             } else {
-                                DEB___("  %ds left to expiry", (ttl + cached_entry->loaded_at) - now);
+                                _deb("  %ds left to expiry", (ttl + cached_entry->loaded_at) - now);
                             }
                         }
                     
@@ -260,26 +260,25 @@ void DNS_Inspector::update(AppHostCX* cx) {
                             cached_response_ttl_idx = cached_entry->answer_ttl_idx;
                             cached_response_decrement = now - cached_entry->loaded_at;
                         
-                            DIAS___("cached entry TTL check: OK");
-                            DEB___("cached response prepared: size=%d, setting overwrite id=%d",cached_response->size(),cached_response_id);
+                            _dia("cached entry TTL check: OK");
+                            _deb("cached response prepared: size=%d, setting overwrite id=%d",cached_response->size(),cached_response_id);
                         } else {
-                            DIAS___("cached entry TTL check: failed");
+                            _dia("cached entry TTL check: failed");
                         }
 
                     }
                 } else {
-                    DIA___("DNS answer for %s is not in cache",ptr->question_str_0().c_str());
+                    _dia("DNS answer for %s is not in cache",ptr->question_str_0().c_str());
                 }
-                inspect_dns_cache.unlock();
             }
             
             break;
         case 'w':
             stage = 1;
-            for(unsigned int it = 0; red < buf.size() && it < 10; it++) {
+            for(unsigned int it = 0; red < shallow_xbuf.size() && it < 10; it++) {
                 ptr = new DNS_Response();
                 
-                buffer cur_buf = buf.view(red,buf.size()-red);
+                buffer cur_buf = shallow_xbuf.view(red, shallow_xbuf.size() - red);
                 int cur_red = ptr->load(&cur_buf);
                 
                 
@@ -295,37 +294,37 @@ void DNS_Inspector::update(AppHostCX* cx) {
                             ((DNS_Response*)ptr)->cached_packet->append(cur_buf.data(),cur_red);
                         }
                         
-                        DEB___("caching response packet: size=%d",((DNS_Response*)ptr)->cached_packet->size())
+                        _deb("caching response packet: size=%d",((DNS_Response*)ptr)->cached_packet->size());
                     }
                     
                     mem_pos += cur_red;
                     red = cur_red;
                     
-                    DIA___("DNS_Inspector::update[%s]: loaded new response (at %d size %d out of %d)",cx->c_name(),red,mem_pos,mem_len);
+                    _dia("DNS_Inspector::update[%s]: loaded new response (at %d size %d out of %d)",cx->c_name(),red,mem_pos,mem_len);
                     if (!validate_response((DNS_Response*)ptr)) {
                         // invalid, delete
 
                         cx->writebuf()->clear();
                         cx->error(true);
-                        WAR___("DNS inspection: cannot find corresponding DNS request id 0x%x: dropping connection.",ptr->id());
+                        _war("DNS inspection: cannot find corresponding DNS request id 0x%x: dropping connection.",ptr->id());
                         delete ptr;
                     }
                     else {
                         // DNS response is valid
                         responses_ ++;
 
-                        DIA___("DNS_Inspector::update[%s]: valid response",cx->c_name());
+                        _dia("DNS_Inspector::update[%s]: valid response",cx->c_name());
 
                         if(store((DNS_Response*)ptr)) {
                             stored_ = true;
                             // DNS response is interesting (A record present) - we stored it , ptr is VALID
-                            DIA___("DNS_Inspector::update[%s]: contains interesting info, stored",cx->c_name());
+                            _dia("DNS_Inspector::update[%s]: contains interesting info, stored",cx->c_name());
                             
                         } else {
                             delete ptr;
                             ptr = nullptr;
                             
-                            DIA___("DNS_Inspector::update[%s]: no interesting info there, deleted",cx->c_name());
+                            _dia("DNS_Inspector::update[%s]: no interesting info there, deleted",cx->c_name());
                         }
                         
                         if(is_tcp)
@@ -346,7 +345,7 @@ void DNS_Inspector::update(AppHostCX* cx) {
     
     fail:
     
-    DIA___("DNS_Inspector::update[%s]: stage %d end (flow size %d)",cx->c_name(),stage, f.flow().size());
+    _dia("DNS_Inspector::update[%s]: stage %d end (flow size %d)",cx->c_name(),stage, f.flow().size());
 }
 
 
@@ -357,35 +356,41 @@ bool DNS_Inspector::store(DNS_Response* ptr) {
 
     std::string ip = ptr->answer_str().c_str();
     if(ip.size() > 0) {
-        INF___("DNS inspection: %s is at%s",ptr->question_str_0().c_str(),ip.c_str()); //ip is already prepended with " "
+        _not("DNS inspection: %s is at%s",ptr->question_str_0().c_str(),ip.c_str()); //ip is already prepended with " "
     }
     else {
-        DIA___("DNS inspection: non-A response for %s",ptr->question_str_0().c_str());
+        _dia("DNS inspection: non-A response for %s",ptr->question_str_0().c_str());
         is_a_record = false;
     }
-    DIA___("DNS response: %s",ptr->to_string().c_str());
+    _dia("DNS response: %s",ptr->to_string().c_str());
 
 
     if(is_a_record) {
         std::string question = ptr->question_str_0();
-        
-        inspect_dns_cache.lock();
-        inspect_dns_cache.set(question,ptr);
-        DIA___("DNS_Inspector::update: %s added to cache (%d elements of max %d)",ptr->question_str_0().c_str(),inspect_dns_cache.cache().size(), inspect_dns_cache.max_size());
-        inspect_dns_cache.unlock();
+
+        {
+            std::scoped_lock<std::recursive_mutex> l_(DNS::get_dns_lock());
+            DNS::get_dns_cache().set(question, ptr);
+        }
+            _dia("DNS_Inspector::update: %s added to cache (%d elements of max %d)", ptr->question_str_0().c_str(),
+                   DNS::get_dns_cache().cache().size(), DNS::get_dns_cache().max_size());
+
         
         std::pair<std::string,std::string> dom_pair = split_fqdn_subdomain(question);
-        DEB___("topdomain = %s, subdomain = %s",dom_pair.first.c_str(), dom_pair.second.c_str());    
+        _deb("topdomain = %s, subdomain = %s",dom_pair.first.c_str(), dom_pair.second.c_str());    
         
         if(dom_pair.first.size() > 0 && dom_pair.second.size() > 0) {
-            domain_cache.lock();
-            auto subdom_cache = domain_cache.get(dom_pair.first);
+
+            std::scoped_lock<std::recursive_mutex> ll_(DNS::get_domain_lock());
+
+            auto subdom_cache = DNS::get_domain_cache().get(dom_pair.first);
             if(subdom_cache != nullptr) {
-                subdom_cache->lock();
+
+                std::lock_guard<std::recursive_mutex> sl_(subdom_cache->getlock());
                 
-                DIA___("Top domain cache entry found for domain %s",dom_pair.first.c_str());
+                _dia("Top domain cache entry found for domain %s",dom_pair.first.c_str());
                 if(subdom_cache->get(dom_pair.second) != nullptr) {
-                    DIA___("Sub domain cache entry found for subdomain %s",dom_pair.second.c_str());
+                    _dia("Sub domain cache entry found for subdomain %s",dom_pair.second.c_str());
                 }
                 
                 
@@ -394,23 +399,20 @@ bool DNS_Inspector::store(DNS_Response* ptr) {
                         std::string  s =  subdomain.first;
                         expiring_int* i = subdomain.second;
                         
-                        DEB___("Sub domain cache list: entry %s",s.c_str());
+                        _deb("Sub domain cache list: entry %s",s.c_str());
                     }
                 }
                 
                 subdom_cache->set(dom_pair.second,new expiring_int(1,28000));
-                subdom_cache->unlock();
-            
             }
             
             else {
-                DIA___("Top domain cache entry NOT found for domain %s",dom_pair.first.c_str());
-                domain_cache_entry_t* subdom_cache = new domain_cache_entry_t(string_format("DNS cache for %s",dom_pair.first.c_str()).c_str(),100,true);
-                subdom_cache->set(dom_pair.second,new expiring_int(1,28000));
+                _dia("Top domain cache entry NOT found for domain %s",dom_pair.first.c_str());
+                auto* subdom_cache = DNS::make_domain_entry(dom_pair.first);
+                      subdom_cache -> set(dom_pair.second, new expiring_int(1, DNS::sub_ttl));
                 
-                domain_cache.set(dom_pair.first,subdom_cache);
+                DNS::get_domain_cache().set(dom_pair.first, subdom_cache);
             }
-            domain_cache.unlock();
         }
     }    
     
@@ -422,19 +424,19 @@ bool DNS_Inspector::validate_response(DNS_Response* ptr) {
     unsigned int id = ptr->id();
     DNS_Request* req = find_request(id);
     if(req) {
-        DIA___("DNS_Inspector::validate_response: request 0x%x found",id);
+        _dia("DNS_Inspector::validate_response: request 0x%x found",id);
         return true;
       
     } else {
-        DIA___("DNS_Inspector::validate_response: request 0x%x not found",id);
-        ERR___("validating DNS response for %s failed.",ptr->to_string().c_str());
+        _dia("DNS_Inspector::validate_response: request 0x%x not found",id);
+        _err("validating DNS response for %s failed.",ptr->to_string().c_str());
         return false; // FIXME: for debug
     }
     
     return false;
 }
 
-std::string DNS_Inspector::to_string(int verbosity) {
+std::string DNS_Inspector::to_string(int verbosity) const {
     std::string r = Inspector::to_string()+"\n  ";
     
     r += string_format("tcp: %d requests: %d valid responses: %d stored: %d",is_tcp,requests_.size(),responses_,stored_);
@@ -446,17 +448,17 @@ void Inspector::apply_verdict(AppHostCX* cx) {
 }
 
 void DNS_Inspector::apply_verdict(AppHostCX* cx) {
-    DEBS___("DNS_Inspector::apply_verdict called");
+    _deb("DNS_Inspector::apply_verdict called");
     
     //TODO: dirty, make more generic
     if(cached_response != nullptr) {
-        DEB___("DNS_Inspector::apply_verdict: mangling response id=%d",cached_response_id);
+        _deb("DNS_Inspector::apply_verdict: mangling response id=%d",cached_response_id);
         *((uint16_t*)cached_response->data()) = htons(cached_response_id);
         
         for(auto i: cached_response_ttl_idx) {
             uint32_t orig_ttl = ntohl(cached_response->get_at<uint32_t>(i));
             uint32_t new_ttl = orig_ttl - cached_response_decrement;
-            DEB___("DNS_Inspector::apply_verdict: mangling original ttl %d to %d at index %d",orig_ttl,new_ttl,i);
+            _deb("DNS_Inspector::apply_verdict: mangling original ttl %d to %d at index %d",orig_ttl,new_ttl,i);
             
             uint8_t* ptr = cached_response->data();
             uint32_t* ptr_ttl  = (uint32_t*)&ptr[i];
@@ -465,12 +467,12 @@ void DNS_Inspector::apply_verdict(AppHostCX* cx) {
         }
         
         if(! is_tcp) {
-            DEBS___("udp encapsulation"); 
+            _deb("udp encapsulation"); 
             cx->to_write(cached_response->data(), cached_response->size());
             int w = cx->write();
-            DIA___("DNS_Inspector::apply_verdict: %d bytes written of cached response size %d",w,cached_response->size());
+            _dia("DNS_Inspector::apply_verdict: %d bytes written of cached response size %d",w,cached_response->size());
         } else {
-            DEBS___("tcp encapsulation");
+            _deb("tcp encapsulation");
             uint16_t* ptr = (uint16_t*)cached_response->data();
             uint16_t len = htons(cached_response->size());
             buffer b;
@@ -479,11 +481,11 @@ void DNS_Inspector::apply_verdict(AppHostCX* cx) {
             cx->to_write(b);
             int w = cx->write();
             
-            DIA___("DNS_Inspector::apply_verdict: %d bytes written of cached response size %d",w,b.size());
+            _dia("DNS_Inspector::apply_verdict: %d bytes written of cached response size %d",w,b.size());
         }
         
     } else {
         // what to do now?
-        ERRS___("cannot send cached response, original reply not found.");
+        _err("cannot send cached response, original reply not found.");
     }
 }
