@@ -40,17 +40,16 @@
 #include <regex>
 #include <cstdlib>
 #include <ctime>
-#include <time.h>
 
 #include <mitmproxy.hpp>
 #include <mitmhost.hpp>
-#include <logger.hpp>
+#include <log/logger.hpp>
 #include <cfgapi.hpp>
-#include <cfgapi_auth.hpp>
 #include <sockshostcx.hpp>
 #include <uxcom.hpp>
 #include <staticcontent.hpp>
 #include <filterproxy.hpp>
+#include <authfactory.hpp>
 
 #include <algorithm>
 #include <ctime>
@@ -59,18 +58,12 @@
 DEFINE_LOGGING(MitmProxy);
 
 
-unsigned int MitmProxy::half_timeout = 30;
-
-socle::meter MitmProxy::total_mtr_up;
-socle::meter MitmProxy::total_mtr_down;
-
-ptr_cache<std::string,whitelist_verify_entry_t> MitmProxy::whitelist_verify("whitelist - verify",500,true,whitelist_verify_entry_t::is_expired);
-
 MitmProxy::MitmProxy(baseCom* c): baseProxy(c), sobject() {
 
     // FIXME: testing filter - get back to it later!
     //add_filter("test",new TestFilter(this,5));
 
+    log = logan::attach(this, "proxy");
 }
 
 void MitmProxy::toggle_tlog() {
@@ -78,27 +71,30 @@ void MitmProxy::toggle_tlog() {
     // create traffic logger if it doesn't exist
     if(tlog_ == nullptr) {
 
-        tlog_ = new socle::trafLog(this,cfg_traflog_dir.c_str(),cfg_traflog_file_pref.c_str(),cfg_traflog_file_suff.c_str());
+        tlog_ = new socle::trafLog( this,
+                                    CfgFactory::get().traflog_dir.c_str(),
+                                    CfgFactory::get().traflog_file_prefix.c_str(),
+                                    CfgFactory::get().traflog_file_suffix.c_str());
     }
     
     // check if we have there status file
     if(tlog_) {
-        std::string data_dir = cfg_traflog_dir;
+        std::string data_dir = CfgFactory::get().traflog_dir;
 
         data_dir += "/disabled";
         
-        struct stat st;
+        struct stat st{0};
         int result = stat(data_dir.c_str(), &st);
         bool present = (result == 0);
         
         if(present) {
-            if(tlog()->status() == true) {
-                WARS___("capture disabled by disabled-file");
+            if(tlog()->status()) {
+                _war("capture disabled by disabled-file");
             }
             tlog()->status(false);
         } else {
-            if(tlog()->status() == false) {
-                WARS___("capture re-enabled from previous disabled-file state");
+            if(! tlog()->status()) {
+                _war("capture re-enabled from previous disabled-file state");
             }            
             tlog()->status(true);
         }
@@ -109,47 +105,41 @@ void MitmProxy::toggle_tlog() {
 MitmProxy::~MitmProxy() {
     
     if(write_payload()) {
-        DEBS___("MitmProxy::destructor: syncing writer");
+        _deb("MitmProxy::destructor: syncing writer");
 
-        for(typename std::vector<baseHostCX*>::iterator j = this->left_sockets.begin(); j != this->left_sockets.end(); ++j) {
-            auto cx = (*j);
-            if(cx->log().size()) {
-                if(tlog()) tlog()->write('L', cx->log());
-                cx->log() = "";
+        for(auto* cx: ls()) {
+            if(! cx->comlog().empty()) {
+                if(tlog()) tlog()->write('L', cx->comlog());
+                cx->comlog().clear();
             }
         }               
         
-        for(typename std::vector<baseHostCX*>::iterator j = this->right_sockets.begin(); j != this->right_sockets.end(); ++j) {
-            auto cx = (*j);
-            if(cx->log().size()) {
-                if(tlog()) tlog()->write('R', cx->log());
-                cx->log() = "";
+        for(auto* cx: rs()) {
+            if(! cx->comlog().empty()) {
+                if(tlog()) tlog()->write('R', cx->comlog());
+                cx->comlog().clear();
             }
         }         
         
         if(tlog()) tlog()->left_write("Connection stop\n");
     }
     
-    if(content_rule_ != nullptr) {
-      delete content_rule_;
-    }
-        
+    delete content_rule_;
     delete tlog_;
-    
-    if(identity_ != nullptr) { delete identity_; }
+    delete identity_;
     
     // delete all filters
-    for (auto p: filters_) {
+    for (auto const& p: filters_) {
         delete p.second;
     }
 }
 
-std::string MitmProxy::to_string(int verbosity) { 
+std::string MitmProxy::to_string(int verbosity) const {
     std::stringstream r;
     r <<  "MitmProxy:" + baseProxy::to_string(verbosity);
     
     if(verbosity >= INF) {
-        r << string_format(" policy: %d ",matched_policy());
+        r << string_format(" policy: %d ", matched_policy());
         
         if(identity_resolved()) {
             r << string_format("identity: %s ",identity_->username().c_str());
@@ -165,7 +155,7 @@ std::string MitmProxy::to_string(int verbosity) {
             PolicyRule* p = nullptr;
             
             if(matched_policy() >= 0) {
-                p = cfgapi_obj_policy.at(matched_policy());
+                p = CfgFactory::get().db_policy.at(matched_policy());
             }
             
             r << string_format("\n    PolicyRule Id: 0x%x",p);
@@ -184,7 +174,7 @@ std::string MitmProxy::to_string(int verbosity) {
 void MitmProxy::identity_resolved(bool b) {
     identity_resolved_ = b;
 }
-bool MitmProxy::identity_resolved() {
+bool MitmProxy::identity_resolved() const {
     return identity_resolved_;
 }
 
@@ -196,32 +186,52 @@ bool MitmProxy::apply_id_policies(baseHostCX* cx) {
     if(cx->com()) {
         af = cx->com()->l3_proto();
     }
-    std::string str_af = inet_family_str(af); 
-    
-    if(af == AF_INET || af == 0) cfgapi_identity_ip_lock.lock();
-    if(af == AF_INET6) cfgapi_identity_ip6_lock.lock();    
-
-    // use common base pointer, so we can use all IdentityInfo types
+    std::string str_af = inet_family_str(af);
     IdentityInfoBase* id_ptr = nullptr;
-    
+
+    bool found = false;
+    std::vector<std::string> group_vec;
+
     if(af == AF_INET || af == 0) {
-        auto ip = auth_ip_map.find(cx->host());
-        if (ip != auth_ip_map.end()) {
+        std::scoped_lock<std::recursive_mutex> l_(AuthFactory::get_ip4_lock());
+
+        // use common base pointer, so we can use all IdentityInfo types
+
+        auto ip = AuthFactory::get_ip4_map().find(cx->host());
+        if (ip != AuthFactory::get_ip4_map().end()) {
             id_ptr = &(*ip).second;
+
+            if(id_ptr) {
+                found = true;
+                for (auto const &my_id: id_ptr->groups_vec) {
+                    group_vec.push_back(my_id);
+                }
+            }
         }
     }
     else if(af == AF_INET6) {
-        auto ip = auth_ip6_map.find(cx->host());
-        if (ip != auth_ip6_map.end()) {
+        std::scoped_lock<std::recursive_mutex> l_(AuthFactory::get_ip6_lock());
+        // use common base pointer, so we can use all IdentityInfo types
+
+        auto ip = AuthFactory::get_ip6_map().find(cx->host());
+        if (ip != AuthFactory::get_ip6_map().end()) {
             id_ptr = &(*ip).second;
+
+            if(id_ptr) {
+                found = true;
+                for(auto const& my_id: id_ptr->groups_vec) {
+                    group_vec.push_back(my_id);
+                }
+            }
         }
     }
-    
+
+
     ProfileSubAuth* final_profile = nullptr;
     
-    if( id_ptr != nullptr) {
-        DIA___("apply_id_policies: matched policy: %d",matched_policy());        
-        PolicyRule* policy = cfgapi_obj_policy.at(matched_policy());
+    if( found ) {
+        _dia("apply_id_policies: matched policy: %d",matched_policy());        
+        PolicyRule* policy = CfgFactory::get().db_policy.at(matched_policy());
         
         ProfileAuth* auth_policy = policy->profile_auth;
 
@@ -231,12 +241,12 @@ bool MitmProxy::apply_id_policies(baseHostCX* cx) {
                 ProfileSubAuth* sub_prof = sub;
                 std::string sub_name = sub->name;
                 
-                DIA___("apply_id_policies: checking identity policy for: %s", sub_name.c_str());
+                _dia("apply_id_policies: checking identity policy for: %s", sub_name.c_str());
                 
-                for(auto my_id: id_ptr->groups_vec) {
-                    DIA___("apply_id_policies: identity in policy: %s, match-test real user group '%s'",sub_prof->name.c_str(), my_id.c_str());
+                for(auto const& my_id: group_vec) {
+                    _dia("apply_id_policies: identity in policy: %s, match-test real user group '%s'",sub_prof->name.c_str(), my_id.c_str());
                     if(sub_prof->name == my_id) {
-                        DIAS___("apply_id_policies: .. matched.");
+                        _dia("apply_id_policies: .. matched.");
                         final_profile = sub_prof;
                         break;
                     }
@@ -253,47 +263,43 @@ bool MitmProxy::apply_id_policies(baseHostCX* cx) {
             const char* pc_name = "none";
             const char* pd_name = "none";
             const char* pt_name = "none";
-            std::string algs = "";
+            std::string algs;
             
-            DIA___("apply_id_policies: assigning sub-profile %s",final_profile->name.c_str());
+            _dia("apply_id_policies: assigning sub-profile %s",final_profile->name.c_str());
             if(final_profile->profile_content != nullptr) {
-                if (cfgapi_obj_profile_content_apply(cx,this,final_profile->profile_content)) {
+                if (CfgFactory::get().prof_content_apply(cx, this, final_profile->profile_content)) {
                     pc_name = final_profile->profile_content->prof_name.c_str();
-                    DIA___("apply_id_policies: assigning content sub-profile %s",final_profile->profile_content->prof_name.c_str());
+                    _dia("apply_id_policies: assigning content sub-profile %s",final_profile->profile_content->prof_name.c_str());
                 }
             }
             if(final_profile->profile_detection != nullptr) {
-                if (cfgapi_obj_profile_detect_apply(cx,this,final_profile->profile_detection)) {
+                if (CfgFactory::get().prof_detect_apply(cx, this, final_profile->profile_detection)) {
                     pd_name = final_profile->profile_detection->prof_name.c_str();
-                    DIA___("apply_id_policies: assigning detection sub-profile %s",final_profile->profile_detection->prof_name.c_str());
+                    _dia("apply_id_policies: assigning detection sub-profile %s",final_profile->profile_detection->prof_name.c_str());
                 }
             }
             if(final_profile->profile_tls != nullptr) {
-                if(cfgapi_obj_profile_tls_apply(cx,this,final_profile->profile_tls)) {
+                if(CfgFactory::get().prof_tls_apply(cx, this, final_profile->profile_tls)) {
                     pt_name = final_profile->profile_tls->prof_name.c_str();
-                    DIA___("apply_id_policies: assigning tls sub-profile %s",final_profile->profile_tls->prof_name.c_str());
+                    _dia("apply_id_policies: assigning tls sub-profile %s",final_profile->profile_tls->prof_name.c_str());
                 }
             }
             if(final_profile->profile_alg_dns != nullptr) {
-                if(cfgapi_obj_alg_dns_apply(cx,this,final_profile->profile_alg_dns)) {
+                if(CfgFactory::get().prof_alg_dns_apply(cx, this, final_profile->profile_alg_dns)) {
                     algs += final_profile->profile_alg_dns->prof_name + " ";
-                    DIA___("apply_id_policies: assigning tls sub-profile %s",final_profile->profile_tls->prof_name.c_str());
+                    _dia("apply_id_policies: assigning tls sub-profile %s",final_profile->profile_tls->prof_name.c_str());
                 }
             }
             
             // end of custom sub-profiles
-            INF___("Connection %s: identity-based sub-profile: name=%s cont=%s det=%s tls=%s algs=%s",cx->full_name('L').c_str(),final_profile->name.c_str(),
+            _inf("Connection %s: identity-based sub-profile: name=%s cont=%s det=%s tls=%s algs=%s",cx->full_name('L').c_str(),final_profile->name.c_str(),
                             pc_name, pd_name, pt_name, algs.c_str()
                             );
         }
-        
-        if(af == AF_INET || af == 0) cfgapi_identity_ip_lock.unlock();
-        if(af == AF_INET6) cfgapi_identity_ip6_lock.unlock();  
+
         return (final_profile != nullptr);
     } 
 
-    if(af == AF_INET || af == 0) cfgapi_identity_ip_lock.unlock();
-    if(af == AF_INET6) cfgapi_identity_ip6_lock.unlock();      
     return false;
 }
 
@@ -320,43 +326,46 @@ bool MitmProxy::resolve_identity(baseHostCX* cx,bool insert_guest=false) {
     
     bool valid_ip_auth = false;
     
-    DIA___("identity check[%s]: source: %s",str_af.c_str(), cx->host().c_str());
-    
-    if(af == AF_INET) cfgapi_auth_shm_ip_table_refresh();
-    if(af == AF_INET6) cfgapi_auth_shm_ip6_table_refresh();
-    
-    
-    cfgapi_identity_ip_lock.lock();
+    _dia("identity check[%s]: source: %s",str_af.c_str(), cx->host().c_str());
+
     shm_logon_info_base* id_ptr = nullptr;
     
     if(af == AF_INET || af == 0) {
-        DEB___("identity check[%s]: table size: %d",str_af.c_str(), auth_ip_map.size());
-        auto ip = auth_ip_map.find(cx->host());
-        if (ip != auth_ip_map.end()) {
+
+        std::scoped_lock<std::recursive_mutex> l_(AuthFactory::get_ip4_lock());
+        AuthFactory::get().shm_ip4_table_refresh();
+
+        _deb("identity check[%s]: table size: %d",str_af.c_str(), AuthFactory::get_ip4_map().size());
+        auto ip = AuthFactory::get_ip4_map().find(cx->host());
+        if (ip != AuthFactory::get_ip4_map().end()) {
             shm_logon_info& li = (*ip).second.last_logon_info;
-            id_ptr = &li;
+            id_ptr = li.clone();
         } else {
-            if (insert_guest == true) {
+            if (insert_guest) {
                 id_ptr = new shm_logon_info(cx->host().c_str(),"guest","guest+guests+guests_ipv4");
             }
         }
     }
     else if(af == AF_INET6) {
         /* maintain in sync with previous if block */
-        DEB___("identity check[%s]: table size: %d",str_af.c_str(), auth_ip6_map.size());
-        auto ip = auth_ip6_map.find(cx->host());
-        if (ip != auth_ip6_map.end()) {
+
+        std::scoped_lock<std::recursive_mutex> l_(AuthFactory::get_ip6_lock());
+        AuthFactory::get().shm_ip6_table_refresh();
+
+        _deb("identity check[%s]: table size: %d",str_af.c_str(), AuthFactory::get_ip6_map().size());
+        auto ip = AuthFactory::get_ip6_map().find(cx->host());
+        if (ip != AuthFactory::get_ip6_map().end()) {
             shm_logon_info6& li = (*ip).second.last_logon_info;
-            id_ptr = &li;
+            id_ptr = li.clone();
         } else {
-            if (insert_guest == true) {
+            if (insert_guest) {
                 id_ptr = new shm_logon_info6(cx->host().c_str(),"guest","guest+guests+guests_ipv6");
             }
         }    
     }
 
     if(id_ptr != nullptr) {
-        DIA___("identity found for %s %s: user: %s groups: %s",str_af.c_str(),cx->host().c_str(),id_ptr->username().c_str(), id_ptr->groups().c_str());
+        _dia("identity found for %s %s: user: %s groups: %s",str_af.c_str(),cx->host().c_str(),id_ptr->username().c_str(), id_ptr->groups().c_str());
 
         // if update_auth_ip_map fails, identity is no longer valid!
         
@@ -372,12 +381,16 @@ bool MitmProxy::resolve_identity(baseHostCX* cx,bool insert_guest=false) {
         // apply specific identity-based profile. 'li' is still valid, since we still hold the lock
         // get ptr to identity_info
 
-        DIA___("resolve_identity[%s]: about to call apply_id_policies, group: %s",str_af.c_str(), id_ptr->groups().c_str());
+        _dia("resolve_identity[%s]: about to call apply_id_policies, group: %s",str_af.c_str(), id_ptr->groups().c_str());
         apply_id_policies(cx);
+
+
+        // id_ptr is either a clone, or new guest id
+        delete id_ptr;
     }
-    
-    cfgapi_identity_ip_lock.unlock();
-    DEB___("identity check[%s]: return %d",str_af.c_str(), valid_ip_auth);
+
+
+    _deb("identity check[%s]: return %d",str_af.c_str(), valid_ip_auth);
     return valid_ip_auth;
 }
 
@@ -392,28 +405,30 @@ bool MitmProxy::update_auth_ipX_map(baseHostCX* cx) {
     }
     std::string str_af = inet_family_str(af);    
     
-    if(af == AF_INET || af == 0) cfgapi_identity_ip_lock.lock();
-    if(af == AF_INET6) cfgapi_identity_ip6_lock.lock();
 
-    DEB___("update_auth_ip_map: start for %s %s",str_af.c_str(), cx->host().c_str());
+    _deb("update_auth_ip_map: start for %s %s",str_af.c_str(), cx->host().c_str());
     
     IdentityInfoBase* id_ptr = nullptr;
     
     if(af == AF_INET || af == 0) {
-        auto ip = auth_ip_map.find(cx->host());
-        if (ip != auth_ip_map.end()) {
+        std::scoped_lock<std::recursive_mutex> l_(AuthFactory::get_ip4_lock());
+
+        auto ip = AuthFactory::get_ip4_map().find(cx->host());
+        if (ip != AuthFactory::get_ip4_map().end()) {
             id_ptr = &(*ip).second;
         }
     }
     else if(af == AF_INET6) {
-        auto ip = auth_ip6_map.find(cx->host());
-        if (ip != auth_ip6_map.end()) {
+        std::scoped_lock<std::recursive_mutex> l_(AuthFactory::get_ip6_lock());
+
+        auto ip = AuthFactory::get_ip6_map().find(cx->host());
+        if (ip != AuthFactory::get_ip6_map().end()) {
             id_ptr = &(*ip).second;
         }
     }
     
     if(id_ptr != nullptr) {
-        DIA___("update_auth_ip_map: user %s from %s %s (groups: %s)",id_ptr->username.c_str(), str_af.c_str(), cx->host().c_str(), id_ptr->groups.c_str());
+        _dia("update_auth_ip_map: user %s from %s %s (groups: %s)",id_ptr->username.c_str(), str_af.c_str(), cx->host().c_str(), id_ptr->groups.c_str());
 
         id_ptr->last_seen_policy = matched_policy();
         
@@ -421,25 +436,28 @@ bool MitmProxy::update_auth_ipX_map(baseHostCX* cx) {
             id_ptr->touch();
             ret = true;
         } else {
-            INF___("identity timeout: user %s from %s %s (groups: %s)",id_ptr->username.c_str(), str_af.c_str(), cx->host().c_str(), id_ptr->groups.c_str());
+            _inf("identity timeout: user %s from %s %s (groups: %s)",id_ptr->username.c_str(), str_af.c_str(), cx->host().c_str(), id_ptr->groups.c_str());
             
             // erase internal ip map entry
-            if(af == AF_INET || af == 0) cfgapi_ip_auth_remove(cx->host());
-            if(af == AF_INET6) cfgapi_ip6_auth_remove(cx->host());
+            if(af == AF_INET || af == 0) {
+                std::scoped_lock<std::recursive_mutex> l_(AuthFactory::get_ip4_lock());
+                AuthFactory::get().ip4_remove(cx->host());
+            }
+            else if(af == AF_INET6) {
+                std::scoped_lock<std::recursive_mutex> l_(AuthFactory::get_ip6_lock());
+                AuthFactory::get().ip6_remove(cx->host());
+            }
         }
     }
-
-    if(af == AF_INET || af == 0) cfgapi_identity_ip_lock.unlock();
-    if(af == AF_INET6) cfgapi_identity_ip6_lock.unlock();
     
-    DEB___("update_auth_ip_map: finished for %s %s, result %d",str_af.c_str(), cx->host().c_str(),ret);
+    _deb("update_auth_ip_map: finished for %s %s, result %d",str_af.c_str(), cx->host().c_str(),ret);
     return ret;
 }
 
 
-void MitmProxy::add_filter(std::string name, FilterProxy* fp) {
+void MitmProxy::add_filter(std::string const& name, FilterProxy* fp) {
 
-    filters_.push_back(std::pair<std::string,FilterProxy*>(name,fp));
+    filters_.emplace_back(std::pair<std::string,FilterProxy*>(name,fp));
     
     for(auto s: fp->ls()) {
         com()->set_monitor(s->socket());
@@ -455,11 +473,11 @@ void MitmProxy::add_filter(std::string name, FilterProxy* fp) {
 
 int MitmProxy::handle_sockets_once(baseCom* xcom) {
     
-    for(auto filter_pair: filters_) {
-        std::string& filter_name = filter_pair.first;
+    for(auto const& filter_pair: filters_) {
+        std::string const& filter_name = filter_pair.first;
         baseProxy* filter_proxy = filter_pair.second;
         
-        DEB___("MitmProxy::handle_sockets_once: running filter %s", filter_name.c_str());
+        _deb("MitmProxy::handle_sockets_once: running filter %s", filter_name.c_str());
         filter_proxy->handle_sockets_once(xcom);
     }
     
@@ -490,7 +508,7 @@ bool MitmProxy::handle_authentication(MitmHostCX* mh)
         resolve_identity(mh);
         
         if(!identity_resolved()) {        
-            DEBS___("identity check: unknown");
+            _deb("identity check: unknown");
             
             if(opt_auth_authenticate) {
                 if(mh->replacement_type() == MitmHostCX::REPLACETYPE_HTTP) {
@@ -503,15 +521,15 @@ bool MitmProxy::handle_authentication(MitmHostCX* mh)
                     // wait, if header won't come in some time, kill the proxy
                     if(mh->meter_read_bytes > 200) {
                         // we cannot use replacements and identity is not resolved... what we can do. Shutdown.
-                        EXTS___("not enough data received to ensure right replacement-aware protocol.");
-                        dead(true);
+                        _ext("not enough data received to ensure right replacement-aware protocol.");
+                        state().dead(true);
                     }
                 }
             }
         } else {
             if(auth_block_identity) {
                 if(mh->replacement_type() == MitmHostCX::REPLACETYPE_HTTP) {
-                    DIAS___("MitmProxy::on_left_bytes: we should block it");
+                    _dia("MitmProxy::on_left_bytes: we should block it");
                     mh->replacement_flag(MitmHostCX::REPLACE_BLOCK);
                     redirected = true;
                     handle_replacement_auth(mh);
@@ -527,40 +545,70 @@ bool MitmProxy::handle_com_response_ssl(MitmHostCX* mh)
 {
     bool redirected = false;
     
-    SSLCom* scom = dynamic_cast<SSLCom*>(mh->peercom());
+    auto* scom = dynamic_cast<SSLCom*>(mh->peercom());
     if(scom && scom->opt_failed_certcheck_replacement) {
-        if(scom->verify_status != SSLCom::VERIFY_OK) {
+        if(!(
+            scom->verify_get() == SSLCom::VERIFY_OK
+             ||
+            scom->verify_get() == ( SSLCom::VERIFY_OK | SSLCom::CLIENT_CERT_RQ )
+            )) {
             
             bool whitelist_found = false;
             
             //look for whitelisted entry
             std::string key = whitelist_make_key(mh);
-            if(key.size() > 0 && key != "?") {
-                whitelist_verify.lock();
-                whitelist_verify_entry_t* wh = whitelist_verify.get(key);
-                DIA___("whitelist_verify[%s]: %s",key.c_str(), wh ? "found" : "not found" );
+            if( ( !key.empty() ) && key != "?") {
+                std::lock_guard<std::recursive_mutex> l_(whitelist_verify().getlock());
+
+                whitelist_verify_entry_t* wh = whitelist_verify().get(key);
+                _dia("whitelist_verify[%s]: %s",key.c_str(), wh ? "found" : "not found" );
 
                 // !!! wh might be already invalid here, unlocked !!!
                 if(wh != nullptr) {
                     whitelist_found = true;
                     if (scom->opt_failed_certcheck_override_timeout_type == 1) {
-                        wh->expired_at = ::time(nullptr) + scom->opt_failed_certcheck_override_timeout;
-                        DIA___("whitelist_verify[%s]: timeout reset to %d",key.c_str(), scom->opt_failed_certcheck_override_timeout );
+                        wh->expired_at() = ::time(nullptr) + scom->opt_failed_certcheck_override_timeout;
+                        _dia("whitelist_verify[%s]: timeout reset to %d",key.c_str(), scom->opt_failed_certcheck_override_timeout );
                     }
                 }
-
-                whitelist_verify.unlock();
             }
             
             
             if(!whitelist_found) {
-                DIAS___("relaxed cert-check: peer sslcom verify not OK, not in whitelist");
+                _dia("relaxed cert-check: peer sslcom verify not OK, not in whitelist");
 
                 if(mh->replacement_type() == MitmHostCX::REPLACETYPE_NONE) {
-                    DIAS___(" -> replacement: none - letting go");
+
+                    // _dia(" -> replacement: none - letting go");
+                    // ok - it might happen signature is not yet triggered (and replacement message should happened).
+                    // there are 2 options:
+                    // a) block the connection
+                    // b) reset the connection
+
+                    // certainly there is NO F***ING WAY to letting it go.
+
+                    bool nice_redirect = false;
+
+                    if(first_right()) {
+                        if ("443" == first_right()->port()) {
+                            nice_redirect = true;
+                        }
+                    }
+
+                    if(! nice_redirect) {
+                        _war("certificate not OK on unknown protocol and port - dropping proxy");
+                        state().dead(true);
+                    }
+                    else {
+                        _war("certificate not OK on known port - assuming http");
+                        mh->replacement_type(MitmHostCX::REPLACETYPE_HTTP);
+                        mh->replacement_flag(MitmHostCX::REPLACE_BLOCK);
+                        redirected = true;
+                        handle_replacement_ssl(mh);
+                    }
                 }
                 else if(mh->replacement_type() == MitmHostCX::REPLACETYPE_HTTP) {
-                    DIAS___(" -> replacement: HTTP - redirecting");
+                    _dia(" -> replacement: HTTP - redirecting");
                     mh->replacement_flag(MitmHostCX::REPLACE_BLOCK);
                     redirected = true;
                     handle_replacement_ssl(mh);
@@ -568,25 +616,24 @@ bool MitmProxy::handle_com_response_ssl(MitmHostCX* mh)
                 } 
                 else if(scom->verify_check(SSLCom::CLIENT_CERT_RQ) && scom->opt_client_cert_action > 0) {
 
-                    DIA___(" -> client-cert request:  opt_client_cert_action=%d", scom->opt_client_cert_action);
+                    _dia(" -> client-cert request:  opt_client_cert_action=%d", scom->opt_client_cert_action);
 
                     if(scom->opt_client_cert_action >= 2) {
                         //we should not block
-                        DIAS___(" -> client-cert request: auto-whitelist");
+                        _dia(" -> client-cert request: auto-whitelist");
 
-                        whitelist_verify.lock();
+                        std::lock_guard<std::recursive_mutex> l_(whitelist_verify().getlock());
                         
                         whitelist_verify_entry v;
-                        whitelist_verify.set(key,new whitelist_verify_entry_t(v,scom->opt_failed_certcheck_override_timeout));
-                        whitelist_verify.unlock();
+                        whitelist_verify().set(key,new whitelist_verify_entry_t(v,scom->opt_failed_certcheck_override_timeout));
                     } else {
-                        DIAS___(" -> client-cert request: none");
+                        _dia(" -> client-cert request: none");
                     }
 
                 }
                 else {
-                    DIAS___(" -> replacement unknown: killing proxy");
-                    dead(true);
+                    _dia(" -> replacement unknown: killing proxy");
+                    state().dead(true);
                 }
             }
         }
@@ -598,7 +645,7 @@ bool MitmProxy::handle_com_response_ssl(MitmHostCX* mh)
 bool MitmProxy::handle_cached_response(MitmHostCX* mh) {
     
     if(mh->inspection_verdict() == Inspector::CACHED) {
-        DIAS___("cached content: not proxying");
+        _dia("cached content: not proxying");
         return true;
     }
     
@@ -611,9 +658,9 @@ void MitmProxy::on_left_bytes(baseHostCX* cx) {
         
         toggle_tlog();
         
-        if(cx->log().size()) {
-            if(tlog()) tlog()->write('L', cx->log());
-            cx->log() = "";
+        if(! cx->comlog().empty()) {
+            if(tlog()) tlog()->write('L', cx->comlog());
+            cx->comlog().clear();
         }
         
         if(tlog()) tlog()->left_write(cx->to_read());
@@ -622,7 +669,7 @@ void MitmProxy::on_left_bytes(baseHostCX* cx) {
 
     bool redirected = false;
     
-    MitmHostCX* mh = dynamic_cast<MitmHostCX*>(cx);
+    auto* mh = dynamic_cast<MitmHostCX*>(cx);
 
     if(mh != nullptr) {
 
@@ -633,7 +680,9 @@ void MitmProxy::on_left_bytes(baseHostCX* cx) {
         redirected = handle_com_response_ssl(mh);
         
         
-        if(handle_cached_response(mh) == true) { return; }
+        if(handle_cached_response(mh)) {
+            return;
+        }
     }
     
     
@@ -644,10 +693,10 @@ void MitmProxy::on_left_bytes(baseHostCX* cx) {
             if(content_rule() != nullptr) {
                 buffer b = content_replace_apply(cx->to_read());
                 j->to_write(b);
-                DIA___("mitmproxy::on_left_bytes: original %d bytes replaced with %d bytes",cx->to_read().size(),b.size())
+                _dia("mitmproxy::on_left_bytes: original %d bytes replaced with %d bytes",cx->to_read().size(),b.size());
             } else {
                 j->to_write(cx->to_read());
-                DIA___("mitmproxy::on_left_bytes: %d copied",cx->to_read().size())
+                _dia("mitmproxy::on_left_bytes: %d copied",cx->to_read().size());
             }
         } else {
         
@@ -661,10 +710,10 @@ void MitmProxy::on_left_bytes(baseHostCX* cx) {
             if(content_rule() != nullptr) {
                 buffer b = content_replace_apply(cx->to_read());
                 j->to_write(b);
-                DIA___("mitmproxy::on_left_bytes: original %d bytes replaced with %d bytes into delayed",cx->to_read().size(),b.size())
+                _dia("mitmproxy::on_left_bytes: original %d bytes replaced with %d bytes into delayed",cx->to_read().size(),b.size());
             } else {	  
                 j->to_write(cx->to_read());
-                DIA___("mitmproxy::on_left_bytes: %d copied to delayed",cx->to_read().size())
+                _dia("mitmproxy::on_left_bytes: %d copied to delayed",cx->to_read().size());
             }
         } else {
         
@@ -674,7 +723,7 @@ void MitmProxy::on_left_bytes(baseHostCX* cx) {
     }    
 
     //update meters
-    total_mtr_up.update(cx->to_read().size());
+    total_mtr_up().update(cx->to_read().size());
     mtr_up.update(cx->to_read().size());
 }
 
@@ -684,9 +733,9 @@ void MitmProxy::on_right_bytes(baseHostCX* cx) {
 
         toggle_tlog();
         
-        if(cx->log().size()) {
-            if(tlog()) tlog()->write('R',cx->log());
-            cx->log() = "";
+        if(! cx->comlog().empty()) {
+            if(tlog()) tlog()->write('R',cx->comlog());
+            cx->comlog().clear();
         }
         
         if(tlog()) tlog()->right_write(cx->to_read());
@@ -698,10 +747,10 @@ void MitmProxy::on_right_bytes(baseHostCX* cx) {
         if(content_rule() != nullptr) {
             buffer b = content_replace_apply(cx->to_read());
             j->to_write(b);
-            DIA___("mitmproxy::on_right_bytes: original %d bytes replaced with %d bytes",cx->to_read().size(),b.size())
+            _dia("mitmproxy::on_right_bytes: original %d bytes replaced with %d bytes",cx->to_read().size(),b.size());
         } else {      
             j->to_write(cx->to_read());
-            DIA___("mitmproxy::on_right_bytes: %d copied",cx->to_read().size())
+            _dia("mitmproxy::on_right_bytes: %d copied",cx->to_read().size());
         }
     }
     for(auto j: left_delayed_accepts) {
@@ -709,15 +758,15 @@ void MitmProxy::on_right_bytes(baseHostCX* cx) {
         if(content_rule() != nullptr) {
             buffer b = content_replace_apply(cx->to_read());
             j->to_write(b);
-            DIA___("mitmproxy::on_right_bytes: original %d bytes replaced with %d bytes into delayed",cx->to_read().size(),b.size())
+            _dia("mitmproxy::on_right_bytes: original %d bytes replaced with %d bytes into delayed",cx->to_read().size(),b.size());
         } else {      
             j->to_write(cx->to_read()); 
-            DIA___("mitmproxy::on_right_bytes: %d copied to delayed",cx->to_read().size())
+            _dia("mitmproxy::on_right_bytes: %d copied to delayed",cx->to_read().size());
         }
     }
 
     // update meters
-    total_mtr_down.update(cx->to_read().size());
+    total_mtr_down().update(cx->to_read().size());
     mtr_down.update(cx->to_read().size());
 }
 
@@ -725,25 +774,25 @@ void MitmProxy::on_right_bytes(baseHostCX* cx) {
 void MitmProxy::__debug_zero_connections(baseHostCX* cx) {
 
     if(cx->meter_write_count == 0 && cx->meter_write_bytes == 0 ) {
-        SSLCom* c = dynamic_cast<SSLCom*>(cx->com());
-        if(c) {
-            c->log_profiling_stats(iINF);
+        auto* xcom = dynamic_cast<SSLCom*>(cx->com());
+        if(xcom) {
+            xcom->log_profiling_stats(iINF);
             int p = 0; 
             int s = cx->socket();
             if(s == 0) s = cx->closed_socket();
             if(s != 0) {
                 buffer b(1024);
                 p = cx->com()->peek(s,b.data(),b.capacity(),0);
-                INF___("        cx peek size %d",p);
+                _inf("        cx peek size %d",p);
             }
             
         }
         
         if(cx->peer()) {
-            SSLCom* c = dynamic_cast<SSLCom*>(cx->peer()->com());
-            if(c) {
-                c->log_profiling_stats(iINF);
-                INF___("        peer transferred bytes: up=%d/%dB dw=%d/%dB",cx->peer()->meter_read_count,cx->peer()->meter_read_bytes,
+            auto* xcom_peer = dynamic_cast<SSLCom*>(cx->peer()->com());
+            if(xcom_peer) {
+                xcom_peer->log_profiling_stats(iINF);
+                _inf("        peer transferred bytes: up=%d/%dB dw=%d/%dB",cx->peer()->meter_read_count,cx->peer()->meter_read_bytes,
                                                                 cx->peer()->meter_write_count, cx->peer()->meter_write_bytes);
                 int p = 0; 
                 int s = cx->peer()->socket();
@@ -751,7 +800,7 @@ void MitmProxy::__debug_zero_connections(baseHostCX* cx) {
                 if(s != 0) {
                     buffer b(1024);
                     p = cx->peer()->com()->peek(s,b.data(),b.capacity(),0);
-                    INF___("        peer peek size %d",p);
+                    _inf("        peer peek size %d",p);
                 }                
             }
             
@@ -766,25 +815,25 @@ void MitmProxy::on_half_close(baseHostCX* cx) {
         if(half_holdtimer > 0) {
             
             // we count timer already!
-            int expiry = half_holdtimer + half_timeout - ::time(nullptr);
+            long expiry = half_holdtimer + half_timeout() - ::time(nullptr);
             
             if(expiry > 0) {
-                EXT___("half-closed: live peer with pending data: keeping up for %ds",expiry);
+                _ext("half-closed: live peer with pending data: keeping up for %ls",expiry);
             } else {
-                DIA___("half-closed: timer's up (%d). closing prematurely.",expiry);
-                this->dead(true);
+                _dia("half-closed: timer's up (%l). closing prematurely.",expiry);
+                state().dead(true);
             }
             
             
         } else {
-            DIA___("half-closed: live peer with pending data: keeping up for %ds",half_timeout);
+            _dia("half-closed: live peer with pending data: keeping up for %ds",half_timeout);
             half_holdtimer = ::time(nullptr);
         }
         
     } else {
         // if peer doesn't exist or peercom doesn't exit, mark proxy dead -- noone to speak to
-        DIAS___("half-closed: peer with pending write-data is dead.")
-        this->dead(true);
+        _dia("half-closed: peer with pending write-data is dead.");
+        state().dead(true);
     }
 }
 
@@ -792,88 +841,104 @@ void MitmProxy::on_left_error(baseHostCX* cx) {
 
     if(cx == nullptr) return;
     
-    if(this->dead()) return;  // don't process errors twice
+    if(state().dead()) return;  // don't process errors twice
 
     
-    DEB___("on_left_error[%s]: proxy marked dead",(this->error_on_read ? "read" : "write"));
-    DUMS___(to_string().c_str());
+    _deb("on_left_error[%s]: proxy marked dead",(state().error_on_read ? "read" : "write"));
+    _dum(to_string().c_str());
     
     if(write_payload()) {
         toggle_tlog();
-        if(tlog()) tlog()->left_write("Client side connection closed: " + cx->name() + "\n");
+        if(tlog()) {
+            tlog()->left_write("Client side connection closed: " + cx->name() + "\n");
+            if(! replacement_msg.empty()) {
+                tlog()->left_write(cx->name() + "   dropped by proxy:" + replacement_msg + "\n");
+            }
+        }
     }
     
     if(opt_auth_resolve)
         resolve_identity(cx);
 
     std::string flags = "L";
-    MitmHostCX* mh = dynamic_cast<MitmHostCX*>(cx);
+    auto* mh = dynamic_cast<MitmHostCX*>(cx);
     if (mh != nullptr && mh->inspection_verdict() == Inspector::CACHED) flags+="C";
 
-    std::string detail;
+    std::stringstream detail;
     
     if(cx->peercom()) {
-        SSLMitmCom* sc = dynamic_cast<SSLMitmCom*>(cx->peercom());
+        auto* sc = dynamic_cast<SSLMitmCom*>(cx->peercom());
         if(sc) {
-            detail += string_format("sni=%s ",sc->get_peer_sni().c_str());
+            detail << string_format("sni=%s ",sc->get_peer_sni().c_str());
         }
     }
     if(mh && mh->application_data) {
         
-        app_HttpRequest* http = dynamic_cast<app_HttpRequest*>(mh->application_data);
+        auto* http = dynamic_cast<app_HttpRequest*>(mh->application_data);
         if(http) {
-            detail += string_format("app=%s%s ",http->proto.c_str(),http->host.c_str()); 
+            detail << "app=" << http->proto << http->host << " ";
         }
         else {
-            detail += string_format("app=%s ",mh->application_data->hr().c_str()); 
+            detail << "app=" << mh->application_data->hr() << " ";
         }
     }
     
-    detail += string_format("user=%s up=%d/%dB dw=%d/%dB flags=%s+%s",
+    detail << string_format("user=%s up=%d/%dB dw=%d/%dB flags=%s+%s",
                                         (identity_resolved() ? identity()->username().c_str() : ""),
                                             cx->meter_read_count,cx->meter_read_bytes,
                                                                 cx->meter_write_count, cx->meter_write_bytes,
                                                                             flags.c_str(),
                                                                             com()->full_flags_str().c_str()
             );
-    
-    if(cx->peer() && cx->peer()->writebuf()->size() == 0) {
-        std::string msg = string_format("Connection from %s closed: %s",cx->full_name('L').c_str(),detail.c_str());
-        INFS_(msg.c_str());
-        if(LEV_(DEB)) __debug_zero_connections(cx);
+
+    if(cx->peer() && cx->peer()->writebuf()->empty()) {
+        std::stringstream msg;
+        msg << "Connection from " << cx->full_name('L') << " closed: " << detail.str();
+        if(! replacement_msg.empty() ) {
+            msg << ", dropped: " << replacement_msg;
+            _inf("%s", msg.str().c_str()); // log to generic logger
+        }
+        _inf("%s", msg.str().c_str());
+        if(*log.level() > DEB) __debug_zero_connections(cx);
         
-        this->dead(true); 
+        state().dead(true);
     } else {
         on_half_close(cx);
-        if(this->dead()) {
+        if(state().dead()) {
             // status dead is new, since we check dead status at the begining
-            std::string msg = string_format("Connection from %s left half-closed: %s",cx->full_name('L').c_str(),detail.c_str());
-            INFS___(msg);
+            std::stringstream msg;
+            msg << "Connection from " << cx->full_name('L') << " left half-closed: " << detail.str();
+            _inf("%s", msg.str().c_str());
         }
     }
     
     if(cx) {
-        cfgapi_ipX_auth_inc_counters(cx);
+        AuthFactory::get().ipX_inc_counters(cx);
     }
 }
 
 void MitmProxy::on_right_error(baseHostCX* cx)
 {
-    if(this->dead()) return;  // don't process errors twice
+    if(state().dead()) return;  // don't process errors twice
     
-    DEB___("on_right_error[%s]: proxy marked dead",(this->error_on_read ? "read" : "write"));
+    _deb("on_right_error[%s]: proxy marked dead",(state().error_on_read ? "read" : "write"));
     
     if(write_payload()) {
         toggle_tlog();
-        if(tlog()) tlog()->right_write("Server side connection closed: " + cx->name() + "\n");
+        if(tlog()) {
+            tlog()->right_write("Server side connection closed: " + cx->name() + "\n");
+            if(! replacement_msg.empty()) {
+                tlog()->right_write(cx->name() + "   dropped by proxy:" + replacement_msg + "\n");
+            }
+        }
     }
     
-//         INF___("Created new proxy 0x%08x from %s:%s to %s:%d",new_proxy,f,f_p, t,t_p );
+//         _inf("Created new proxy 0x%08x from %s:%s to %s:%d",new_proxy,f,f_p, t,t_p );
 
 
     std::string flags = "R";
-    std::string comflags = "";
-    MitmHostCX* mh_peer = dynamic_cast<MitmHostCX*>(cx->peer());
+    std::string comflags;
+    auto* mh_peer = dynamic_cast<MitmHostCX*>(cx->peer());
     if (mh_peer != nullptr) {
         if(mh_peer->inspection_verdict() == Inspector::CACHED) flags+="C";
         if(mh_peer->com() != nullptr)
@@ -881,21 +946,21 @@ void MitmProxy::on_right_error(baseHostCX* cx)
     }
 
     
-    std::string detail;
-    SSLMitmCom* sc = dynamic_cast<SSLMitmCom*>(cx->com());
+    std::stringstream detail;
+    auto* sc = dynamic_cast<SSLMitmCom*>(cx->com());
     if(sc) {
-        detail += string_format("sni=%s ",sc->get_peer_sni().c_str());
+        detail << "sni= " << sc->get_peer_sni();
     }
     if(mh_peer && mh_peer->application_data) {
-        app_HttpRequest* http = dynamic_cast<app_HttpRequest*>(mh_peer->application_data);
+        auto* http = dynamic_cast<app_HttpRequest*>(mh_peer->application_data);
         if(http) {
-            detail += string_format("app=%s%s ",http->proto.c_str(),http->host.c_str()); 
+            detail << "app=" << http->proto << http->host << " ";
         }
         else {
-            detail += string_format("app=%s ",mh_peer->application_data->hr().c_str()); 
+            detail << "app=" << mh_peer->application_data->hr() << " ";
         }
     }
-    detail += string_format("user=%s up=%d/%dB dw=%d/%dB flags=%s+%s",
+    detail << string_format("user=%s up=%d/%dB dw=%d/%dB flags=%s+%s",
                                         (identity_resolved() ? identity()->username().c_str() : ""),
                                             cx->meter_read_count,cx->meter_read_bytes,
                                                                 cx->meter_write_count, cx->meter_write_bytes,
@@ -904,24 +969,31 @@ void MitmProxy::on_right_error(baseHostCX* cx)
             );
     
     
-    if(cx->peer() && cx->peer()->writebuf()->size() == 0) {
-        std::string msg = string_format("Connection from %s closed: %s",cx->full_name('R').c_str(),detail.c_str());
-        INFS_(msg.c_str());
-        if(LEV_(DEB)) __debug_zero_connections(cx);
+    if( cx->peer() && cx->peer()->writebuf()->empty() ) {
+        std::stringstream msg;
+        msg << "Connection from " << cx->full_name('R') << " closed: " << detail.str().c_str();
+        if(! replacement_msg.empty() ) {
+            msg << ", dropped: " << replacement_msg;
+            _inf("%s", msg.str().c_str()); // log to generic logger
+        }
+        _inf("%s", msg.str().c_str());
+
+        if(*log.level() > DEB) __debug_zero_connections(cx);
         
-        this->dead(true); 
+        state().dead(true);
     } else {
 
         on_half_close(cx);
-        if(this->dead()) {
+        if(state().dead()) {
             // status dead is new, since we check dead status at the begining
-            std::string msg = string_format("Connection from %s right half-closed: %s",cx->full_name('R').c_str(),detail.c_str());
-            INFS___(msg);
+            std::stringstream msg;
+            msg << "Connection from " << cx->full_name('R') << " right half-closed: " << detail.str().c_str();
+            _inf("%s", msg.str().c_str());
         }
     } 
     
     if(cx->peer()) {
-        cfgapi_ipX_auth_inc_counters(cx->peer());        
+        AuthFactory::get().ipX_inc_counters(cx->peer());
     }
 }
 
@@ -937,58 +1009,56 @@ void MitmProxy::handle_replacement_auth(MitmHostCX* cx) {
   
     
     std::string repl;
-    std::string repl_port = cfgapi_identity_portal_port_http;
+    std::string repl_port = AuthFactory::get().portal_port_http;
     std::string repl_proto = "http";
-    int 	redir_hint = 0;
-    
+
     if(cx->application_data->is_ssl) {
         repl_proto = "https";
-        repl_port = cfgapi_identity_portal_port_https;
+        repl_port =AuthFactory::get().portal_port_https;
     }    
     
-    std::string block_pre("<h2 class=\"fg-red\">Page has been blocked</h2><p>Access has been blocked by smithproxy.</p>\
-    <p>To check your user privileges go to status page<p><p> <form action=\"");
+    std::string block_pre("<h2 class=\"fg-red\">Page has been blocked</h2><p>Access has been blocked by smithproxy.</p>"
+                          "<p>To check your user privileges go to status page<p><p> <form action=\"");
+
     std::string block_post("\"><input type=\"submit\" value=\"User Info\" class=\"btn-red\"></form>");
     
-    //cx->host().c_str()
-    
     if (cx->replacement_flag() == MitmHostCX::REPLACE_REDIRECT) {
-        //srand(time(nullptr) % ((unsigned long)cx));
-        //redir_hint = rand();
 
-        cfgapi_identity_token_lock.lock();
-        auto id_token = cfgapi_identity_token_cache.find(cx->host());
+        std::scoped_lock<std::recursive_mutex> l_(AuthFactory::get_token_lock());
+        auto id_token = AuthFactory::get_token_map().find(cx->host());
         
-        if(id_token != cfgapi_identity_token_cache.end()) {
-            INF___("found a cached token for %s",cx->host().c_str());
+        if(id_token != AuthFactory::get_token_map().end()) {
+            _inf("found a cached token for %s",cx->host().c_str());
             std::pair<unsigned int,std::string>& cache_entry = (*id_token).second;
             
             unsigned int now      = time(nullptr);
             unsigned int token_ts = cache_entry.first;
             std::string& token_tk = cache_entry.second;
             
-            if(now - token_ts < cfgapi_identity_token_timeout) {
-                INF___("MitmProxy::handle_replacement_auth: cached token %s for request: %s",token_tk.c_str(),cx->application_data->hr().c_str());
+            if(now - token_ts < AuthFactory::get().token_timeout) {
+                _inf("MitmProxy::handle_replacement_auth: cached token %s for request: %s",token_tk.c_str(),cx->application_data->hr().c_str());
                 
                 if(cx->com()) {
                     if(cx->com()->l3_proto() == AF_INET) {
-                        repl = redir_pre + repl_proto + "://"+cfgapi_identity_portal_address+":"+repl_port+"/cgi-bin/auth.py?token=" + token_tk + redir_suf;
+                        repl = redir_pre + repl_proto + "://"+AuthFactory::get().portal_address+":"+repl_port+"/cgi-bin/auth.py?token=" + token_tk + redir_suf;
                     } else if(cx->com()->l3_proto() == AF_INET6) {
-                        repl = redir_pre + repl_proto + "://"+cfgapi_identity_portal_address6+":"+repl_port+"/cgi-bin/auth.py?token=" + token_tk + redir_suf;
+                        repl = redir_pre + repl_proto + "://"+AuthFactory::get().portal_address6+":"+repl_port+"/cgi-bin/auth.py?token=" + token_tk + redir_suf;
                     } 
                 } 
                 
-                if(repl.size() == 0) {
+                if(repl.empty()) {
                     // default to IPv4 address
-                    repl = redir_pre + repl_proto + "://"+cfgapi_identity_portal_address+":"+repl_port+"/cgi-bin/auth.py?token=" + token_tk + redir_suf;
+                    repl = redir_pre + repl_proto + "://"+AuthFactory::get().portal_address+":"+repl_port+"/cgi-bin/auth.py?token=" + token_tk + redir_suf;
                 }
                 
-                repl = global_staticconent->render_server_response(repl);
+                repl = html()->render_server_response(repl);
                 
                 cx->to_write((unsigned char*)repl.c_str(),repl.size());
                 cx->close_after_write(true);
+
+                replacement_msg += "(auth: known token)";
             } else {
-                INF___("MitmProxy::handle_replacement_auth: expired token %s for request: %s",token_tk.c_str(),cx->application_data->hr().c_str());
+                _inf("MitmProxy::handle_replacement_auth: expired token %s for request: %s",token_tk.c_str(),cx->application_data->hr().c_str());
                 goto new_token;
             }
         } else {
@@ -997,101 +1067,219 @@ void MitmProxy::handle_replacement_auth(MitmHostCX* cx) {
             
             std::string token_text = cx->application_data->original_request();
           
-            for(auto i: cfgapi_obj_policy_profile_auth( cx->matched_policy())->sub_policies) {
-                DIA___("MitmProxy::handle_replacement_auth: token: requesting identity %s",i->name.c_str());
+            for(auto i: CfgFactory::get().policy_prof_auth(cx->matched_policy())->sub_policies) {
+                _dia("MitmProxy::handle_replacement_auth: token: requesting identity %s",i->name.c_str());
                 token_text  += " |" + i->name;
             }
             shm_logon_token tok = shm_logon_token(token_text.c_str());
             
-            INF___("MitmProxy::handle_replacement_auth: new auth token %s for request: %s",tok.token().c_str(),cx->application_data->hr().c_str());
+            _inf("MitmProxy::handle_replacement_auth: new auth token %s for request: %s",tok.token().c_str(),cx->application_data->hr().c_str());
             
             if(cx->com()) {
                 if(cx->com()->l3_proto() == AF_INET) {
-                    repl = redir_pre + repl_proto + "://"+cfgapi_identity_portal_address+":"+repl_port+"/cgi-bin/auth.py?token=" + tok.token() + redir_suf;
+                    repl = redir_pre + repl_proto + "://"+AuthFactory::get().portal_address+":"+repl_port+"/cgi-bin/auth.py?token=" + tok.token() + redir_suf;
                 } else if(cx->com()->l3_proto() == AF_INET6) {
-                    repl = redir_pre + repl_proto + "://"+cfgapi_identity_portal_address6+":"+repl_port+"/cgi-bin/auth.py?token=" + tok.token() + redir_suf;
+                    repl = redir_pre + repl_proto + "://"+AuthFactory::get().portal_address6+":"+repl_port+"/cgi-bin/auth.py?token=" + tok.token() + redir_suf;
                 } 
             } 
             
-            if(repl.size() == 0) {
+            if(repl.empty()) {
                 // default to IPv4 address
-                INFS___("XXX: fallback to IPv4");
-                repl = redir_pre + repl_proto + "://"+cfgapi_identity_portal_address+":"+repl_port+"/cgi-bin/auth.py?token=" + tok.token() + redir_suf;
+                _inf("XXX: fallback to IPv4");
+                repl = redir_pre + repl_proto + "://"+AuthFactory::get().portal_address+":"+repl_port+"/cgi-bin/auth.py?token=" + tok.token() + redir_suf;
             }
             
-            repl = global_staticconent->render_server_response(repl);
+            repl = html()->render_server_response(repl);
             
             cx->to_write((unsigned char*)repl.c_str(),repl.size());
             cx->close_after_write(true);
+            replacement_msg += "(auth: new token)";
+
+            AuthFactory::get().shm_token_table_refresh();
+
+            AuthFactory::get().shm_token_map_.entries().push_back(tok);
+            AuthFactory::get().shm_token_map_.acquire();
+            AuthFactory::get().shm_token_map_.save(true);
+            AuthFactory::get().shm_token_map_.release();
             
-            cfgapi_auth_shm_token_table_refresh();
-            
-            auth_shm_token_map.entries().push_back(tok);
-            auth_shm_token_map.acquire();
-            auth_shm_token_map.save(true);
-            auth_shm_token_map.release();
-            
-            DIAS___("MitmProxy::handle_replacement_auth: token table updated");
-            cfgapi_identity_token_cache[cx->host()] = std::pair<unsigned int,std::string>(time(nullptr),tok.token());
+            _dia("MitmProxy::handle_replacement_auth: token table updated");
+            AuthFactory::get_token_map()[cx->host()] = std::pair<unsigned int,std::string>(time(nullptr),tok.token());
         }
-        
-        cfgapi_identity_token_lock.unlock();
     } else
     if (cx->replacement_flag() == MitmHostCX::REPLACE_BLOCK) {
 
-        DIAS___("MitmProxy::handle_replacement_auth: instructed to replace block");
-        repl = block_pre + repl_proto + "://"+cfgapi_identity_portal_address+":"+repl_port + "/cgi-bin/auth.py?a=z" + block_post;
+        _dia("MitmProxy::handle_replacement_auth: instructed to replace block");
+        repl = block_pre + repl_proto + "://"+AuthFactory::get().portal_address+":"+repl_port + "/cgi-bin/auth.py?a=z" + block_post;
         
         std::string cap  = "Page Blocked";
-        std::string meta = "";
-        repl = global_staticconent->render_msg_html_page(cap,meta, repl,"500px");
-        repl = global_staticconent->render_server_response(repl);
+        std::string meta;
+        repl = html()->render_msg_html_page(cap,meta, repl,"700px");
+        repl = html()->render_server_response(repl);
         
         cx->to_write((unsigned char*)repl.c_str(),repl.size());
         cx->close_after_write(true);
 
+        replacement_msg += "(auth: blocked)";
+
     } else
     if (cx->replacement_flag() == MitmHostCX::REPLACE_NONE) {
-        DIAS___("MitmProxy::handle_replacement_auth: asked to handle NONE. No-op.");
+        _dia("MitmProxy::handle_replacement_auth: asked to handle NONE. No-op.");
     } 
 }
 
+std::string verify_flag_string(int code) {
+    switch(code) {
+        case SSLCom::VERIFY_OK:
+            return "Certificate verification successful";
+        case SSLCom::SELF_SIGNED:
+            return "Target certificate is self-signed";
+        case SSLCom::SELF_SIGNED_CHAIN:
+            return "Server certificate's chain contains self-signed, untrusted CA certificate";
+        case SSLCom::UNKNOWN_ISSUER:
+            return "Server certificate is issued by untrusted certificate authority";
+        case SSLCom::CLIENT_CERT_RQ:
+            return "Server is asking client for a certificate";
+        case SSLCom::REVOKED:
+            return "Server's certificate is REVOKED";
+        case SSLCom::HOSTNAME_FAILED:
+            return "Client application asked for SNI server is not offering";
+        default:
+            return "";
+    }
+}
+
+void MitmProxy::set_replacement_msg_ssl(SSLCom* scom) {
+    if(scom && scom->verify_get() != SSLCom::VERIFY_OK) {
+
+        if(scom->verify_check(SSLCom::SELF_SIGNED)) {
+            replacement_msg += "(ssl:" + verify_flag_string(SSLCom::SELF_SIGNED) + ")";
+        }
+        if(scom->verify_check(SSLCom::SELF_SIGNED_CHAIN)) {
+            replacement_msg += "(ssl:" + verify_flag_string(SSLCom::SELF_SIGNED_CHAIN) + ")";
+        }
+        if(scom->verify_check(SSLCom::UNKNOWN_ISSUER)) {
+            replacement_msg += "(ssl:" + verify_flag_string(SSLCom::UNKNOWN_ISSUER) + ")";
+        }
+        if(scom->verify_check(SSLCom::CLIENT_CERT_RQ)) {
+            replacement_msg += "(ssl:" + verify_flag_string(SSLCom::CLIENT_CERT_RQ) + ")";
+        }
+        if(scom->verify_check(SSLCom::REVOKED)) {
+            replacement_msg += "(ssl:" + verify_flag_string(SSLCom::REVOKED) + ")";
+        }
+        if(scom->verify_check(SSLCom::HOSTNAME_FAILED)) {
+            replacement_msg += "(ssl:" + verify_flag_string(SSLCom::HOSTNAME_FAILED) + ")";
+        }
+    }
+}
+
+std::string MitmProxy::replacement_ssl_verify_detail(SSLCom* scom) {
+
+    std::stringstream ss;
+    if(scom) {
+        if (scom->verify_get() != SSLCom::VERIFY_OK) {
+            bool is_set = false;
+
+            if (scom->verify_check(SSLCom::SELF_SIGNED)) {
+                ss << "<p><h3 class=\"fg-red\">Reason:</h3> " << verify_flag_string(SSLCom::SELF_SIGNED) << ".</p>";
+                is_set = true;
+            }
+            if (scom->verify_check(SSLCom::SELF_SIGNED_CHAIN)) {
+                ss << "<p><h3 class=\"fg-red\">Reason:</h3> " << verify_flag_string(SSLCom::SELF_SIGNED_CHAIN)
+                   << ".</p>";
+                is_set = true;
+            }
+            if (scom->verify_check(SSLCom::UNKNOWN_ISSUER)) {
+                ss << "<p><h class=\"fg-red\"3>Reason:</h3>" << verify_flag_string(SSLCom::UNKNOWN_ISSUER) << ".</p>";
+                is_set = true;
+            }
+            if (scom->verify_check(SSLCom::CLIENT_CERT_RQ)) {
+                ss << "<p><h3 class=\"fg-red\">Reason:</h3>" << verify_flag_string(SSLCom::CLIENT_CERT_RQ) << ".<p>";
+                is_set = true;
+            }
+            if (scom->verify_check(SSLCom::REVOKED)) {
+                ss << "<p><h3 class=\"fg-red\">Reason:</h3>" << verify_flag_string(SSLCom::REVOKED) << ". "
+                                                                                                       "This is a serious issue, it's highly recommended to not continue "
+                                                                                                       "to this page.</p>";
+                is_set = true;
+            }
+            if (scom->verify_check(SSLCom::HOSTNAME_FAILED)) {
+                ss << "<p><h3 class=\"fg-red\">Reason:</h3>" << verify_flag_string(SSLCom::HOSTNAME_FAILED) << ".</p>";
+                is_set = true;
+            }
+
+
+            if (!is_set) {
+                ss << string_format(
+                        "<p><h3 class=\"fg-red\">Reason:</h3>Oops, no detailed problem description (code: 0x%x)</p>",
+                        scom->verify_get());
+            }
+        } else {
+            ss << string_format(
+                    "<p><h3 class=\"fg-red\">Reason:</h3>Oops, no detailed problem description (code: 0x%x)</p>",
+                    scom->verify_get());
+        }
+    }
+
+    return ss.str();
+}
+
+
+std::string MitmProxy::replacement_ssl_page(SSLCom* scom, app_HttpRequest* app_request, std::string const& more_info) {
+    std::string repl;
+
+    std::string block_target_info = "<p><h3 class=\"fg-red\">Requested site:</h3>" + app_request->proto + app_request->host + "</p>";
+
+    std::string block_additinal_info = replacement_ssl_verify_detail(scom) + more_info;
+
+    std::string cap = "TLS security warning";
+    std::string meta;
+    std::string war_img = html()->render_noargs("html_img_warning");
+    std::string msg = string_format("<h2 class=\"fg-red\">%s TLS security warning</h2>%s",war_img.c_str(),(block_target_info + block_additinal_info).c_str());
+    repl = html()->render_msg_html_page(cap, meta, msg,"700px");
+    repl = html()->render_server_response(repl);
+
+    return repl;
+}
 
 void MitmProxy::handle_replacement_ssl(MitmHostCX* cx) {
     
     std::string repl;
     
-    SSLCom* scom = dynamic_cast<SSLCom*>(cx->peercom());
+    auto* scom = dynamic_cast<SSLCom*>(cx->peercom());
     if(!scom) {
         std::string error("<html><head></head><body><p>Internal error</p><p>com object is not ssl-type</p></body></html>");
-        error = global_staticconent->render_server_response(error);
+        error = html()->render_server_response(error);
         
         cx->to_write((unsigned char*)error.c_str(),error.size());
-        cx->close_after_write(true);  
-        
-        ERRS___("cannot handle replacement for TLS, com is not SSLCom");
+        cx->close_after_write(true);
+        set_replacement_msg_ssl(scom);
+
+        _err("cannot handle replacement for TLS, com is not SSLCom");
         
         return;
     }
     
     std::string block_additinal_info;
-    std::string block_override_pre = "<form action=\"/SM/IT/HP/RO/XY";
-    
-    std::string key = whitelist_make_key(cx);
-    if(cx->peer()) {
-        block_override_pre += "/override/target=" + key;//cx->peer()->host() + "#" + cx->peer()->port() + "&";
-    }
     std::string block_override;
-    std::string block_target_info;
+
+    if(scom->opt_failed_certcheck_override)   {
+        std::string block_override_pre = "<form action=\"/SM/IT/HP/RO/XY";
+
+        std::string key = whitelist_make_key(cx);
+        if(cx->peer()) {
+            block_override_pre += "/override/target=" + key;//cx->peer()->host() + "#" + cx->peer()->port() + "&";
+        }
+        std::string block_target_info;
+        block_override = block_override_pre + R"(orig_url=/"><input type="submit" value="Override" class="btn-red"></form>)";
+    }
     
-    
-    app_HttpRequest* app_request = dynamic_cast<app_HttpRequest*>(cx->application_data);
+    auto* app_request = dynamic_cast<app_HttpRequest*>(cx->application_data);
     if(app_request != nullptr) {
         
-//         INF___(" --- request: %s",app_request->request().c_str());
-//         INF___(" ---     uri: %s",app_request->uri.c_str());
-//         INF___(" --- origuri: %s",app_request->original_request().c_str());
-//         INF___(" --- referer: %s",app_request->referer.c_str());
+//         _inf(" --- request: %s",app_request->request().c_str());
+//         _inf(" ---     uri: %s",app_request->uri.c_str());
+//         _inf(" --- origuri: %s",app_request->original_request().c_str());
+//         _inf(" --- referer: %s",app_request->referer.c_str());
         
         
         if(app_request->request().find("/SM/IT/HP/RO/XY/override") != std::string::npos) {
@@ -1102,42 +1290,49 @@ void MitmProxy::handle_replacement_ssl(MitmHostCX* cx) {
                         
             if(scom->opt_failed_certcheck_override) {
             
-                DIA___("ssl_override: ph4 - asked for verify override for %s", whitelist_make_key(cx).c_str());
+                _dia("ssl_override: ph4 - asked for verify override for %s", whitelist_make_key(cx).c_str());
                 
                 std::string orig_url = "about:blank";
                 
                 //we require orig_url is the last argument!!!
-                unsigned int a = app_request->request().find("orig_url");
+                auto a = app_request->request().find("orig_url");
                 if(a != std::string::npos) {
                     //len of "orig_url=" is 9
                     orig_url = app_request->request().substr(a+9);
                 }
                 
                 
-                std::string override_applied = string_format("<html><head><meta http-equiv=\"Refresh\" content=\"0; url=%s\"></head><body><!-- applied, redirecting back to %s --></body></html>",
+                std::string override_applied = string_format(
+                        "<html><head><meta http-equiv=\"Refresh\" content=\"0; url=%s\"></head><body><!-- applied, redirecting back to %s --></body></html>",
                                                             orig_url.c_str(),orig_url.c_str());
 
                 whitelist_verify_entry v;
 
-                whitelist_verify.lock();
-                whitelist_verify.set(key,new whitelist_verify_entry_t(v,scom->opt_failed_certcheck_override_timeout));
-                whitelist_verify.unlock();
+                {
+                    std::lock_guard<std::recursive_mutex> l_(whitelist_verify().getlock());
+                    whitelist_verify().set(whitelist_make_key(cx),
+                                         new whitelist_verify_entry_t(v, scom->opt_failed_certcheck_override_timeout));
+                }
                 
-                override_applied = global_staticconent->render_server_response(override_applied);
+                override_applied = html()->render_server_response(override_applied);
                 
                 cx->to_write((unsigned char*)override_applied.c_str(),override_applied.size());
                 cx->close_after_write(true);
+                set_replacement_msg_ssl(scom);
+                replacement_msg += "(ssl: override)";
                 
-                WAR___("Connection from %s: SSL override activated for %s",cx->full_name('L').c_str(), app_request->request().c_str());
+                _war("Connection from %s: SSL override activated for %s",cx->full_name('L').c_str(), app_request->request().c_str());
                 
                 return;
                 
             } else {
                 // override is not enabled, but client somehow reached this (attack?)
                 std::string error("<html><head></head><body><p>Failed to override</p><p>Action is denied.</p></body></html>");
-                error = global_staticconent->render_server_response(error);
+                error = html()->render_server_response(error);
                 cx->to_write((unsigned char*)error.c_str(),error.size());
-                cx->close_after_write(true);                  
+                cx->close_after_write(true);
+                set_replacement_msg_ssl(scom);
+                replacement_msg += "(ssl: override disabled)";
                 
                 return;
             }
@@ -1148,90 +1343,67 @@ void MitmProxy::handle_replacement_ssl(MitmHostCX* cx) {
             // PHASE III.
             // display warning and button which will trigger override
         
-            DIA___("ssl_override: ph3 - warning replacement for %s", whitelist_make_key(cx).c_str());
+            _dia("ssl_override: ph3 - warning replacement for %s", whitelist_make_key(cx).c_str());
             
-            block_target_info = "<p><h3 class=\"fg-red\">Requested site:</h3>" + app_request->proto + app_request->host + "</p>";
-            block_override = string_format("orig_url=%s\"><input type=\"submit\" value=\"Override\" class=\"btn-red\"></form>","/");
+            repl = replacement_ssl_page(scom, app_request, block_override);
 
-            if(scom->verify_get() > 0) {
-                bool is_set = false;
-                
-                if(scom->verify_check(SSLCom::SELF_SIGNED)) {
-                        block_additinal_info += "<p><h3 class=\"fg-red\">Reason:</h3>Target certificate is self-signed.</p>"; is_set = true;
-                }
-                if(scom->verify_check(SSLCom::SELF_SIGNED_CHAIN)) {
-                        block_additinal_info += "<p><h3 class=\"fg-red\">Reason:</h3>Server certificate's chain contains self-signed, untrusted CA certificate.</p>"; is_set = true;
-                }
-                if(scom->verify_check(SSLCom::UNKNOWN_ISSUER)) {
-                        block_additinal_info += "<p><h class=\"fg-red\"3>Reason:</h3>Server certificate is issued by untrusted certificate authority.</p>"; is_set = true;
-                }
-                if(scom->verify_check(SSLCom::CLIENT_CERT_RQ)) {
-                        block_additinal_info += "<p><h3 class=\"fg-red\">Reason:</h3>Server is asking for a client certificate.<p>"; is_set = true;
-                }
-                if(scom->verify_check(SSLCom::REVOKED)) {
-                        block_additinal_info += "<p><h3 class=\"fg-red\">Reason:</h3>Server's certificate is REVOKED. This is a serious issue, it's highly recommended to not continue to this page.</p>"; is_set = true;
-                }                
-                if(scom->verify_check(SSLCom::HOSTNAME_FAILED)) {
-                        block_additinal_info += "<p><h3 class=\"fg-red\">Reason:</h3>Client application asked for server (sni) server is not offering.</p>"; is_set = true;
-                }                
-
-                
-                if(!is_set) {
-                        block_additinal_info += string_format("<p><h3 class=\"fg-red\">Reason:</h3>Oops, no detailed problem description (code: 0x%04x)</p>",scom->verify_get());
-                }
-            } else {
-                block_additinal_info += string_format("<p><h3 class=\"fg-red\">Reason:</h3>Oops, no detailed problem description (code: 0x%04x)</p>",scom->verify_get());
-            }
-            
-            if(scom->opt_failed_certcheck_override)  block_additinal_info += block_override_pre + block_override;
-            
-            DIAS___("MitmProxy::handle_replacement_ssl: instructed to replace block");
-            
-            std::string cap = "TLS security warning";
-            std::string meta;
-            std::string war_img = global_staticconent->render_noargs("html_img_warning");
-            std::string msg = string_format("<h2 class=\"fg-red\">%s TLS security warning</h2>%s",war_img.c_str(),(block_target_info + block_additinal_info).c_str());
-            repl = global_staticconent->render_msg_html_page(cap, meta, msg,"500px");
-            repl = global_staticconent->render_server_response(repl);
-            
             cx->to_write((unsigned char*)repl.c_str(),repl.size());
+            set_replacement_msg_ssl(scom);
             cx->close_after_write(true);
         } else 
         if(app_request->uri == "/"){
             // PHASE II
             // redir to warning message
             
-            DIA___("ssl_override: ph2 - redir to warning replacement for  %s", whitelist_make_key(cx).c_str());
+            _dia("ssl_override: ph2 - redir to warning replacement for  %s", whitelist_make_key(cx).c_str());
             
             std::string repl = "<html><head><meta http-equiv=\"Refresh\" content=\"0; url=/SM/IT/HP/RO/XY/warning?q=1\"></head><body></body></html>";
-            repl = global_staticconent->render_server_response(repl);
-            cx->to_write((unsigned char*)repl.c_str(),repl.size());
-            cx->close_after_write(true);            
+            repl = html()->render_server_response(repl);
+            cx->to_write(repl);
+            cx->close_after_write(true);
+            set_replacement_msg_ssl(scom);
         }   
         else {
             // PHASE I
             // redirecting to / -- for example some subpages would be displayed incorrectly
             
-            DIA___("ssl_override: ph1 - redir to / for %s", whitelist_make_key(cx).c_str());
+            _dia("ssl_override: ph1 - redir to / for %s", whitelist_make_key(cx).c_str());
             
             std::string redir_pre("<html><head><script>top.location.href=\"");
             std::string redir_suf("\";</script></head><body></body></html>");  
             
             //std::string repl = "<html><head><meta http-equiv=\"Refresh\" content=\"0; url=/\"></head><body></body></html>";            
             std::string repl = redir_pre + "/" + redir_suf;   
-            repl = global_staticconent->render_server_response(repl);
+            repl = html()->render_server_response(repl);
             
             cx->to_write((unsigned char*)repl.c_str(),repl.size());
-            cx->close_after_write(true);            
+            cx->close_after_write(true);
+            set_replacement_msg_ssl(scom);
         }
     }
-    
+
+    else {
+        _dia("ssl_override: enforced ph1 - redir to / for %s", whitelist_make_key(cx).c_str());
+        _inf("readbuf: \n%s", hex_dump(cx->readbuf(), 4).c_str());
+
+        std::string redir_pre("<html><head><script>top.location.href=\"");
+        std::string redir_suf("\";</script></head><body></body></html>");
+
+        //std::string repl = "<html><head><meta http-equiv=\"Refresh\" content=\"0; url=/\"></head><body></body></html>";
+        std::string repl = redir_pre + "/" + redir_suf;
+        repl = html()->render_server_response(repl);
+
+        cx->to_write((unsigned char*)repl.c_str(),repl.size());
+        cx->close_after_write(true);
+        set_replacement_msg_ssl(scom);
+        replacement_msg += "(ssl: enforced)";
+    }
 }
 
 void MitmProxy::init_content_replace() {
     
     if(content_rule_ != nullptr) {
-        DIAS___("MitmProxy::init_content_replace: deleting old replace rules");
+        _dia("MitmProxy::init_content_replace: deleting old replace rules");
         delete content_rule_;
     }
     
@@ -1243,8 +1415,7 @@ buffer MitmProxy::content_replace_apply(buffer b) {
     std::string result = data;
     
     int stage = 0;
-    for(auto i = content_rule()->begin(); i != content_rule()->end(); ++i) {
-        ProfileContentRule& profile = (*i);
+    for(auto& profile: *content_rule()) {
         
         try {
             std::regex re_match(profile.match.c_str());
@@ -1254,7 +1425,7 @@ buffer MitmProxy::content_replace_apply(buffer b) {
 
                 // unfortunately std::regex_replace doesn't return if it really replaced something
                 // ... which is no problem if we don't care. But in case we want to replace only 
-                // .... nth occurence, we have to do extra search to check (requiring one extra regex match).
+                // .... nth occurrence, we have to do extra search to check (requiring one extra regex match).
                 std::smatch sm;
                 bool is_there = std::regex_search(result,sm,re_match);
 
@@ -1268,7 +1439,7 @@ buffer MitmProxy::content_replace_apply(buffer b) {
                             result = std::regex_replace(result, re_match, repl);              
                         }
                         profile.replace_each_counter_ = 0;
-                        DIA___("Replacing bytes[stage %d]: n-th counter hit",stage);
+                        _dia("Replacing bytes[stage %d]: n-th counter hit",stage);
                     }
                 }
                 
@@ -1280,10 +1451,10 @@ buffer MitmProxy::content_replace_apply(buffer b) {
                 }
             }
 
-            DIA___("Replacing bytes[stage %d]:",stage);
+            _dia("Replacing bytes[stage %d]:",stage);
         }
-        catch(std::regex_error e) {
-        NOT___("MitmProxy::content_replace_apply: failed to replace string: %s",e.what());
+        catch(std::regex_error const& e) {
+        _not("MitmProxy::content_replace_apply: failed to replace string: %s",e.what());
         }
         
         ++stage;
@@ -1296,14 +1467,14 @@ buffer MitmProxy::content_replace_apply(buffer b) {
     buffer ret_b;
     ret_b.append(result.c_str(),result.size());
     
-    DIA___("content rewritten: original %d bytes with new %d bytes.",b.size(),ret_b.size());
-    DUM___("Replacing bytes (%d):\n%s\n# with bytes(%d):\n%s",data.size(),hex_dump(b).c_str(),ret_b.size(),hex_dump(ret_b).c_str());
+    _dia("content rewritten: original %d bytes with new %d bytes.",b.size(),ret_b.size());
+    _dum("Replacing bytes (%d):\n%s\n# with bytes(%d):\n%s",data.size(),hex_dump(b).c_str(),ret_b.size(),hex_dump(ret_b).c_str());
     return ret_b;
 }
 
 void MitmProxy::tap() {
 
-    DIAS___("MitmProxy::tap: start");
+    _dia("MitmProxy::tap: start");
     
     for (auto cx: left_sockets) {
         com()->unset_monitor(cx->socket());
@@ -1317,7 +1488,7 @@ void MitmProxy::tap() {
 
 void MitmProxy::untap() {
 
-    DIAS___("MitmProxy::untap: start");
+    _dia("MitmProxy::untap: start");
 
     for (auto cx: left_sockets) {
         com()->set_monitor(cx->socket());
@@ -1332,11 +1503,10 @@ void MitmProxy::untap() {
 MitmHostCX* MitmProxy::first_left() {
     MitmHostCX* ret{};
     
-    if(ls().size()) {
+    if(! ls().empty()) {
         ret = dynamic_cast<MitmHostCX*>(ls().at(0));
     }
-    else 
-    if(lda().size()) {
+    else if(! lda().empty()) {
         ret = dynamic_cast<MitmHostCX*>(lda().at(0));
     }
         
@@ -1344,16 +1514,15 @@ MitmHostCX* MitmProxy::first_left() {
 }
 
 MitmHostCX* MitmProxy::first_right() {
-    MitmHostCX* ret{};
+    MitmHostCX* ret = nullptr;
     
-    if(rs().size()) {
+    if(! rs().empty()) {
         ret = dynamic_cast<MitmHostCX*>(rs().at(0));
     }
-    else 
-    if(rda().size()) {
-        ret = dynamic_cast<MitmHostCX*>(rda().at(0));
+    else if(! rda().empty()) {
+            ret = dynamic_cast<MitmHostCX*>(rda().at(0));
     }
-        
+
     return ret;
 }
 
@@ -1362,35 +1531,35 @@ MitmHostCX* MitmProxy::first_right() {
 bool MitmMasterProxy::ssl_autodetect = false;
 bool MitmMasterProxy::ssl_autodetect_harder = true;
 
-#define NEW_CX_PEEK_BUFFER_SZ  10
 bool MitmMasterProxy::detect_ssl_on_plain_socket(int s) {
     
     int ret = false;
+    constexpr unsigned int NEW_CX_PEEK_BUFFER_SZ = 10;
+    constexpr int time_increment = 2500; // 2.5ms
+    constexpr int time_max = time_increment*5;
 
-    int time_increment = 2500; // 2.5ms
-    int time_max = time_increment*5;
     int time_taken = 0;
     
     if (s > 0) {
         again:
-        
         char peek_buffer[NEW_CX_PEEK_BUFFER_SZ];
+
         int b = ::recv(s,peek_buffer,NEW_CX_PEEK_BUFFER_SZ,MSG_PEEK|MSG_DONTWAIT);
         
         if(b > 6) {
             if (peek_buffer[0] == 0x16 && peek_buffer[1] == 0x03 && ( peek_buffer[5] == 0x00 || peek_buffer[5] == 0x01 || peek_buffer[5] == 0x02 )) {
-                INF___("detect_ssl_on_plain_socket: SSL detected on socket %d",s);
+                _inf("detect_ssl_on_plain_socket: SSL detected on socket %d",s);
                 ret = true;
             }
         } else {
             if(ssl_autodetect_harder && time_taken < time_max) {
-                struct timespec t;
+                struct timespec t{0};
                 t.tv_sec = 0;
                 t.tv_nsec = time_increment;
                 
                 ::nanosleep(&t,nullptr);
                 time_taken += time_increment;
-                DIA___("detect_ssl_on_plain_socket: SSL strict detection on socket %d: dalayed by %dnsec",s,time_increment);
+                _dia("detect_ssl_on_plain_socket: SSL strict detection on socket %d: dalayed by %dnsec",s,time_increment);
                 goto again;
             }
         }
@@ -1401,19 +1570,18 @@ bool MitmMasterProxy::detect_ssl_on_plain_socket(int s) {
 
 baseHostCX* MitmMasterProxy::new_cx(int s) {
     
-    DEBS___("MitmMasterProxy::new_cx: new_cx start");
+    _deb("MitmMasterProxy::new_cx: new_cx start");
     
     bool is_ssl = false;
     bool is_ssl_port = false;
     
-    SSLCom* my_sslcom = dynamic_cast<SSLCom*>(com());
+    auto* my_sslcom = dynamic_cast<SSLCom*>(com());
     baseCom* c = nullptr;
     
     if(my_sslcom != nullptr) {
         is_ssl_port = true;
     }
-    else
-    if(ssl_autodetect) {
+    else if(ssl_autodetect) {
         // my com is NOT ssl-based, trigger auto-detect
 
         is_ssl = detect_ssl_on_plain_socket(s);
@@ -1425,13 +1593,13 @@ baseHostCX* MitmMasterProxy::new_cx(int s) {
         } 
     }
     
-    if(c == nullptr) {
+    if(! c) {
         c = com()->slave();
     }
     
     auto r = new MitmHostCX(c,s);
     if (is_ssl) {
-        INF___("Connection %s: SSL detected on unusual port.",r->c_name());
+        _inf("Connection %s: SSL detected on unusual port.",r->c_name());
         r->is_ssl = true;
         r->is_ssl_port = is_ssl_port;
     }
@@ -1439,7 +1607,7 @@ baseHostCX* MitmMasterProxy::new_cx(int s) {
         r->is_ssl = true;
     }
     
-    DEB___("Pausing new connection %s",r->c_name());
+    _deb("Pausing new connection %s",r->c_name());
     r->waiting_for_peercom(true);
     return r; 
 }
@@ -1448,7 +1616,7 @@ void MitmMasterProxy::on_left_new(baseHostCX* just_accepted_cx) {
     // to create child proxy and attach this cx to it.
 
     if(! just_accepted_cx->com()->nonlocal_dst_resolved()) {
-        ERRS___("Was not possible to resolve original destination!");
+        _err("Was not possible to resolve original destination!");
         just_accepted_cx->shutdown();
         delete just_accepted_cx;
     } 
@@ -1458,14 +1626,14 @@ void MitmMasterProxy::on_left_new(baseHostCX* just_accepted_cx) {
         just_accepted_cx->name();
         just_accepted_cx->com()->resolve_socket_src(just_accepted_cx->socket(),&h,&p);
 
-        MitmProxy* new_proxy = new MitmProxy(just_accepted_cx->com()->slave());
+        auto* new_proxy = new MitmProxy(just_accepted_cx->com()->slave());
         
         // let's add this just_accepted_cx into new_proxy
         if(just_accepted_cx->read_waiting_for_peercom()) {
-            DEBS___("MitmMasterProxy::on_left_new: ldaadd the new waiting_for_peercom cx");
+            _deb("MitmMasterProxy::on_left_new: ldaadd the new waiting_for_peercom cx");
             new_proxy->ldaadd(just_accepted_cx);
         } else{
-            DEBS___("MitmMasterProxy::on_left_new: ladd the new cx (unpaused)");
+            _deb("MitmMasterProxy::on_left_new: ladd the new cx (unpaused)");
             new_proxy->ladd(just_accepted_cx);
         }
         
@@ -1476,31 +1644,31 @@ void MitmMasterProxy::on_left_new(baseHostCX* just_accepted_cx) {
         short unsigned int target_port = just_accepted_cx->com()->nonlocal_dst_port();
         short unsigned int orig_target_port;
 
-        if( target_host == cfgapi_tenant_magic_ip) {
+        if( target_host == CfgFactory::get().tenant_magic_ip) {
             
             orig_target_host = target_host;
             orig_target_port = target_port;
             
             if(target_port == 65000 || target_port == 143) {
                 // bend broker magic IP
-                target_port = 65000 + cfgapi_tenant_index;
+                target_port = 65000 + CfgFactory::get().tenant_index;
             }
             else if(target_port != 443) {
                 // auth portal https magic IP
-                target_port = std::stoi(cfgapi_identity_portal_port_http);
+                target_port = std::stoi(AuthFactory::get().portal_port_http);
             } else {
                 // auth portal plaintext magic IP
-                target_port = std::stoi(cfgapi_identity_portal_port_https);
+                target_port = std::stoi(AuthFactory::get().portal_port_https);
             }
             target_host = "127.0.0.1";
             
-            DIA___("Connection from %s to %s:%d: traffic redirected from magic IP to %s:%d",just_accepted_cx->c_name(), 
+            _dia("Connection from %s to %s:%d: traffic redirected from magic IP to %s:%d",just_accepted_cx->c_name(), 
                  orig_target_host.c_str(), orig_target_port,
                  target_host.c_str(), target_port);
             matched_vip = true;
         }
         
-        MitmHostCX *target_cx = new MitmHostCX(just_accepted_cx->com()->slave(), target_host.c_str(), 
+        auto *target_cx = new MitmHostCX(just_accepted_cx->com()->slave(), target_host.c_str(),
                                             string_format("%d",target_port).c_str()
                                             );
         
@@ -1517,11 +1685,11 @@ void MitmMasterProxy::on_left_new(baseHostCX* just_accepted_cx) {
         bool delete_proxy = false;
         
         // apply policy and get result
-        int policy_num = cfgapi_obj_policy_apply(just_accepted_cx,new_proxy);
+        int policy_num = CfgFactory::get().policy_apply(just_accepted_cx, new_proxy);
 
         // bypass ssl com to VIP
         if(matched_vip) {
-            SSLCom* scom = dynamic_cast<SSLCom*>(just_accepted_cx->com());
+            auto* scom = dynamic_cast<SSLCom*>(just_accepted_cx->com());
             if(scom != nullptr) {
                 scom->opt_bypass = true;
             }
@@ -1555,20 +1723,20 @@ void MitmMasterProxy::on_left_new(baseHostCX* just_accepted_cx) {
                     // reload table and check timeouts each 5 seconds 
                     time_t now = time(nullptr);
                     if(now > auth_table_refreshed + 5) {
-                        cfgapi_auth_shm_ip_table_refresh();
-                        cfgapi_auth_shm_ip6_table_refresh();
+                        AuthFactory::get().shm_ip4_table_refresh();
+                        AuthFactory::get().shm_ip6_table_refresh();
                         auth_table_refreshed = now;
                         
                         //one day this can run in separate thread to not slow down session setup rate
-                        cfgapi_ip_auth_timeout_check();
-                        cfgapi_ip6_auth_timeout_check();
+                        AuthFactory::get().ip4_timeout_check();
+                        AuthFactory::get().ip6_timeout_check();
                     }
                     
                     
                     if(!res) {
                         if(target_port != 80 && target_port != 443){
                             delete_proxy = true;
-                            INF___("Dropping non-replaceable connection %s due to unknown source IP",just_accepted_cx->c_name());
+                            _inf("Dropping non-replaceable connection %s due to unknown source IP",just_accepted_cx->c_name());
                             goto end;
                         }
                     } else {
@@ -1582,43 +1750,58 @@ void MitmMasterProxy::on_left_new(baseHostCX* just_accepted_cx) {
                         std::string str_af = inet_family_str(af);
                         
                         
-                        cfgapi_identity_ip_lock.lock();
-                        
                         // use common base pointer, so we can use all IdentityInfo types
+
                         IdentityInfoBase* id_ptr = nullptr;
+                        bool found = false;
+                        std::vector<std::string> group_vec;
                         
                         if(af == AF_INET || af == 0) {
-                            auto ip = auth_ip_map.find(just_accepted_cx->host());
-                            if (ip != auth_ip_map.end()) {
+
+                            std::scoped_lock<std::recursive_mutex> l_(AuthFactory::get_ip4_lock());
+
+                            auto ip = AuthFactory::get_ip4_map().find(just_accepted_cx->host());
+                            if (ip != AuthFactory::get_ip4_map().end()) {
                                 id_ptr = &(*ip).second;
+                                found = true;
+                                if(id_ptr)
+                                    group_vec = id_ptr->groups_vec;
                             }
                         }
                         else if(af == AF_INET6) {
-                            auto ip = auth_ip6_map.find(just_accepted_cx->host());
-                            if (ip != auth_ip6_map.end()) {
+
+                            std::scoped_lock<std::recursive_mutex> l_(AuthFactory::get_ip6_lock());
+
+                            auto ip = AuthFactory::get_ip6_map().find(just_accepted_cx->host());
+                            if (ip != AuthFactory::get_ip6_map().end()) {
                                 id_ptr = &(*ip).second;
+                                found = true;
+                                if(id_ptr)
+                                    group_vec = id_ptr->groups_vec;
                             }
                         }
                         
-                        if(id_ptr != nullptr) {
+                        if( found ) {
                             //std::string groups = id_ptr->last_logon_info.groups();
                             
-                            if(cfgapi_obj_policy_profile_auth(policy_num) != nullptr)
-                            for ( auto i: cfgapi_obj_policy_profile_auth(policy_num)->sub_policies) {
-                                for(auto x: id_ptr->groups_vec) {
-                                    DEB___("Connection identities: ip identity '%s' against policy '%s'",x.c_str(),i->name.c_str());
-                                    if(x == i->name) {
-                                        DIA___("Connection identities: ip identity '%s' matches policy '%s'",x.c_str(),i->name.c_str());
-                                        bad_auth = false;
+                            if(CfgFactory::get().policy_prof_auth(policy_num) != nullptr) {
+                                for ( auto i: CfgFactory::get().policy_prof_auth(policy_num)->sub_policies) {
+                                    for(auto const& x: group_vec) {
+                                        _deb("Connection identities: ip identity '%s' against policy '%s'",x.c_str(),i->name.c_str());
+                                        if(x == i->name) {
+                                            _dia("Connection identities: ip identity '%s' matches policy '%s'",x.c_str(),i->name.c_str());
+                                            bad_auth = false;
+                                        }
                                     }
                                 }
                             }
+
                             if(bad_auth) {
                                 if(target_port != 80 && target_port != 443) {
-                                    INF___("Dropping non-replaceable connection %s due to non-matching identity",just_accepted_cx->c_name());
+                                    _inf("Dropping non-replaceable connection %s due to non-matching identity",just_accepted_cx->c_name());
                                 }
                                 else {
-                                    INF___("Connection %s with non-matching identity (to be replaced)",just_accepted_cx->c_name());
+                                    _inf("Connection %s with non-matching identity (to be replaced)",just_accepted_cx->c_name());
                                     // set bad_auth true, because despite authentication failed, it could be replaced (we can let user know 
                                     // he is not allowed to proceed
                                     bad_auth = false;
@@ -1626,7 +1809,6 @@ void MitmMasterProxy::on_left_new(baseHostCX* just_accepted_cx) {
                                 }
                             }
                         }
-                        cfgapi_identity_ip_lock.unlock();
                         
                         if(bad_auth) {
                             delete_proxy = true;
@@ -1636,23 +1818,23 @@ void MitmMasterProxy::on_left_new(baseHostCX* just_accepted_cx) {
                 }
                 
                 // setup NAT
-                if(cfgapi_obj_policy.at(policy_num)->nat == POLICY_NAT_NONE && ! matched_vip) {
+                if(CfgFactory::get().db_policy.at(policy_num)->nat == POLICY_NAT_NONE && ! matched_vip) {
                     target_cx->com()->nonlocal_src(true);
                     target_cx->com()->nonlocal_src_host() = h;
                     target_cx->com()->nonlocal_src_port() = std::stoi(p);               
                 }
                 
                 // finalize connection acceptance by adding new proxy to proxies and connect
-                this->proxies().insert(new_proxy);
+                this->proxies().push_back(new_proxy);
                 
                 //FIXME: this is really ugly!! :) It's here since radd has been called before socket for target_cx was created.
-                int real_socket = target_cx->connect(false);
+                int real_socket = target_cx->connect();
                 com()->set_monitor(real_socket);
                 com()->set_poll_handler(real_socket,new_proxy);
                 
             } else {
                 delete_proxy = true;
-                NOT___("MitmMasterProxy::on_left_new: %s cannot be converted to MitmHostCx",just_accepted_cx->c_name());
+                _not("MitmMasterProxy::on_left_new: %s cannot be converted to MitmHostCx",just_accepted_cx->c_name());
             }
   
         } else {
@@ -1667,33 +1849,33 @@ void MitmMasterProxy::on_left_new(baseHostCX* just_accepted_cx) {
         
         
         if(delete_proxy) {
-            INF___("Dropping proxy %s",new_proxy->c_name());
+            _inf("Dropping proxy %s",new_proxy->c_name());
             delete new_proxy;
         }        
     }
     
-    DEBS___("MitmMasterProxy::on_left_new: finished");
+    _deb("MitmMasterProxy::on_left_new: finished");
 }
 
 int MitmMasterProxy::handle_sockets_once(baseCom* c) {
-    //T_DIAS___("slist",5,this->hr()+"\n===============\n");
+    //T__dia("slist",5,this->hr()+"\n===============\n");
     return ThreadedAcceptorProxy<MitmProxy>::handle_sockets_once(c);
 }
 
 
 void MitmUdpProxy::on_left_new(baseHostCX* just_accepted_cx)
 {
-    MitmProxy* new_proxy = new MitmProxy(com()->slave());
+    auto* new_proxy = new MitmProxy(com()->slave());
     // let's add this just_accepted_cx into new_proxy
     if(just_accepted_cx->read_waiting_for_peercom()) {
-        DEBS___("MitmMasterProxy::on_left_new: ldaadd the new waiting_for_peercom cx");
+        _deb("MitmMasterProxy::on_left_new: ldaadd the new waiting_for_peercom cx");
         new_proxy->ldaadd(just_accepted_cx);
     } else{
-        DEBS___("MitmMasterProxy::on_left_new: ladd the new cx (unpaused)");
+        _deb("MitmMasterProxy::on_left_new: ladd the new cx (unpaused)");
         new_proxy->ladd(just_accepted_cx);
     }
     
-    MitmHostCX *target_cx = new MitmHostCX(com()->slave(), just_accepted_cx->com()->nonlocal_dst_host().c_str(), 
+    auto *target_cx = new MitmHostCX(com()->slave(), just_accepted_cx->com()->nonlocal_dst_host().c_str(),
                                     string_format("%d",just_accepted_cx->com()->nonlocal_dst_port()).c_str()
                                     );
     
@@ -1704,7 +1886,7 @@ void MitmUdpProxy::on_left_new(baseHostCX* just_accepted_cx)
     just_accepted_cx->com()->resolve_socket_src(just_accepted_cx->socket(),&h,&p);
     target_cx->com()->l3_proto(just_accepted_cx->com()->l3_proto());
     
-    //DEB___("UDP proxy: src l3 = %s dst l3 = %s",inet_family_str(just_accepted_cx->com()->l3_proto()).c_str(), inet_family_str(target_cx->com()->l3_proto()).c_str());
+    //_deb("UDP proxy: src l3 = %s dst l3 = %s",inet_family_str(just_accepted_cx->com()->l3_proto()).c_str(), inet_family_str(target_cx->com()->l3_proto()).c_str());
     
     just_accepted_cx->peer(target_cx);
     target_cx->peer(just_accepted_cx);
@@ -1717,28 +1899,28 @@ void MitmUdpProxy::on_left_new(baseHostCX* just_accepted_cx)
     new_proxy->radd(target_cx);
 
     // apply policy and get result
-    int policy_num = cfgapi_obj_policy_apply(just_accepted_cx,new_proxy);
+    int policy_num = CfgFactory::get().policy_apply(just_accepted_cx, new_proxy);
     if(policy_num >= 0) {
-        this->proxies().insert(new_proxy);
+        this->proxies().push_back(new_proxy);
         
         ((MitmHostCX*)just_accepted_cx)->matched_policy(policy_num);
         target_cx->matched_policy(policy_num);
         new_proxy->matched_policy(policy_num);
         
-        if(cfgapi_obj_policy.at(policy_num)->nat == POLICY_NAT_NONE) {
+        if(CfgFactory::get().db_policy.at(policy_num)->nat == POLICY_NAT_NONE) {
             target_cx->com()->nonlocal_src(true);
             target_cx->com()->nonlocal_src_host() = h;
             target_cx->com()->nonlocal_src_port() = std::stoi(p);               
         }
 
-        int real_socket = target_cx->connect(false);
+        int real_socket = target_cx->connect();
         target_cx->rename();
         com()->set_monitor(real_socket);
         com()->set_poll_handler(real_socket,new_proxy);
     }
     
     new_proxy->name(new_proxy->to_string());
-    DEBS___("MitmUDPProxy::on_left_new: finished");    
+    _deb("MitmUDPProxy::on_left_new: finished");    
 }
 
 baseHostCX* MitmUdpProxy::MitmUdpProxy::new_cx(int s) {
