@@ -41,114 +41,122 @@
 
 DEFINE_LOGGING(PolicyRule);
 
-std::string PolicyRule::to_string(int verbosity) {
+std::string PolicyRule::to_string(int verbosity) const {
 
-    std::string from = "PolicyRule:";
+    std::stringstream from;
+    from << "PolicyRule:";
     
     switch(proto) {
         case 6:
-            from += " [tcp] ";
+            from << " [tcp] ";
             break;
         case 17:
-            from += " [udp] ";
+            from << " [udp] ";
             break;
         default:
-            from += string_format(" [%3d] ",proto);
+            from << string_format(" [%3d] ",proto);
     }
     
-    if(src_default) from += "*";
+    if(src_default) from << "*";
     for(auto it: src) {
         if(verbosity > iINF) {
-            from += string_format("(0x%x)",this);
+            from << string_format("(0x%x)",this);
         }
-        from += " ";
-        from += it->to_string();
+        from << it->to_string() << " ";
     }
-    from += ":";
-    if(src_ports_default) from += "*";
+    from << ":";
+    if(src_ports_default) from << "*";
     for(auto it: src_ports) {
-        from += string_format("(%d,%d) ",it.first,it.second);
+        if(it.first == 0 && it.second == 65535)
+            from << "(*) ";
+        else
+            from << string_format("(%d,%d) ",it.first,it.second);
     }
     
-    std::string to = "";
-    if(dst_default) to += "*";
+    std::stringstream to;
+
+    if(dst_default) to << "*";
     for(auto it: dst) {
-        to += " ";
-        to += it->to_string();
+        to << it->to_string() << " ";
     }
-    to += ":";
-    if(dst_ports_default) to+="*";
+    to << ":";
+    if(dst_ports_default) to << "*";
     for(auto it: dst_ports) {
-        to += string_format("(%d,%d) ",it.first,it.second);
+        if(it.first == 0 && it.second == 65535)
+            to << "(*) ";
+        else
+            to << string_format("(%d,%d) ",it.first,it.second);
     }
     
-    std::string out = from + " -> " + to + "= ";
+    std::stringstream out;
+    out << from.str() << "-> " << to.str() << "= ";
     
     switch(action) {
         case POLICY_ACTION_PASS:
-            out += "ACCEPT";
+            out << "ACCEPT";
             break;
         case POLICY_ACTION_DENY:
-            out += "REJECT";
+            out << "REJECT";
             break;
         default:
-            out += "???";
+            out << "???";
     }
     
     switch(nat) {
         case POLICY_NAT_NONE:
-            out += "(nonat)";
+            out << "(nonat)";
             break;
         case POLICY_NAT_AUTO:
-            out += "(iface)";
+            out << "(iface)";
             break;
         case POLICY_NAT_POOL:
-            out += "( pool)";
+            out << "( pool)";
             break;
         default:
-            out += "(  ?  )";
+            out << "(  ?  )";
             break;
     }
-    
-    out += " [" + std::to_string(cnt_matches) + "]";
+
+
+    if(verbosity > iINF)
+        out << " [" << std::to_string(cnt_matches) << "x]";
     
     if(verbosity > INF) {
-        out+=": ";
-        if(profile_auth) out += string_format("\n    auth=%s  (0x%x) ",profile_auth->prof_name.c_str(),profile_auth);
-        if(profile_tls) out += string_format("\n    tls=%s  (0x%x) ",profile_tls->prof_name.c_str(),profile_tls);
-        if(profile_detection) out += string_format("\n    det=%s  (0x%x) ",profile_detection->prof_name.c_str(),profile_detection);
-        if(profile_content) out += string_format("\n    cont=%s  (0x%x) ",profile_content->prof_name.c_str(),profile_content);
-        if(profile_alg_dns) out += string_format("\n    alg_dns=%s  (0x%x) ",profile_alg_dns->prof_name.c_str(),profile_alg_dns);
+        out << ": ";
+        if(profile_auth) out << string_format("\n    auth=%s  (0x%x) ",profile_auth->prof_name.c_str(),profile_auth);
+        if(profile_tls) out << string_format("\n    tls=%s  (0x%x) ",profile_tls->prof_name.c_str(),profile_tls);
+        if(profile_detection) out << string_format("\n    det=%s  (0x%x) ",profile_detection->prof_name.c_str(),profile_detection);
+        if(profile_content) out << string_format("\n    cont=%s  (0x%x) ",profile_content->prof_name.c_str(),profile_content);
+        if(profile_alg_dns) out << string_format("\n    alg_dns=%s  (0x%x) ",profile_alg_dns->prof_name.c_str(),profile_alg_dns);
     }
     
-    return out;
+    return out.str();
 }
 
 
 bool PolicyRule::match_addrgrp_cx(std::vector< AddressObject* >& sources, baseHostCX* cx) {
     bool match = false;
     
-    if(sources.size() == 0) {
+    if(sources.empty()) {
         match = true;
-//                 DIAS_("PolicyRule: matched ");
+//                 _dia("PolicyRule: matched ");
     } else {
         CIDR* l = cidr_from_str(cx->host().c_str());
-        for(std::vector<AddressObject*>::iterator j = sources.begin(); j != sources.end(); ++j ) {
-            AddressObject* comp = (*j);
+        for(auto* comp: sources) {
             
             if(comp->match(l)) {
-                if(LEV_(DIA)) {
+                if(*log.level() >= DIA) {
                     char* a = cidr_to_str(l);
-                    DIA_("PolicyRule::match_addrgrp_cx: comparing %s with rule %s: matched",a,comp->to_string().c_str());
-                    delete a;
+                    _dia("PolicyRule::match_addrgrp_cx: comparing %s with rule %s: matched", a, comp->to_string().c_str());
+                    free(a);
                 }
                 match = true;
                 break;
             } else {
-                if(LEV_(DIA)) {
+                if(*log.level() >= DIA) {
                     char* a = cidr_to_str(l);
-                    DIA_("PolicyRule::match_addrgrp_cx: comparing %s with rule %s: not matched",a,comp->to_string().c_str());
-                    delete[] a;
+                    _deb("PolicyRule::match_addrgrp_cx: comparing %s with rule %s: not matched", a, comp->to_string().c_str());
+                    free(a);
                 }
             }
         }
@@ -161,19 +169,19 @@ bool PolicyRule::match_addrgrp_cx(std::vector< AddressObject* >& sources, baseHo
 bool PolicyRule::match_rangegrp_cx(std::vector< range >& ranges, baseHostCX* cx) {
     bool match = false;
     
-    if(ranges.size() == 0) {
+    if(ranges.empty()) {
         match = true;
-//                 DIAS_("PolicyRule: matched ");
+//                 _dia("PolicyRule: matched ");
     } else {
         int p = std::stoi(cx->port());
-        for(std::vector<range>::iterator j = ranges.begin(); j != ranges.end(); ++j ) {
-            range& comp = (*j);
+        for(auto const& comp: ranges) {
+
             if((p >= comp.first) && (p <= comp.second)) {
-                DIA_("PolicyRule::match_rangergrp_cx: comparing %d with %s: matched",p,rangetos(comp).c_str());
+                _dia("PolicyRule::match_rangergrp_cx: comparing %d with %s: matched",p,rangetos(comp).c_str());
                 match = true;
                 break;
             } else {
-                DIA_("PolicyRule::match_rangergrp_cx: comapring %d with %s: not matched",p,rangetos(comp).c_str());
+                _deb("PolicyRule::match_rangergrp_cx: comapring %d with %s: not matched",p,rangetos(comp).c_str());
             }
         }
     }
@@ -185,16 +193,15 @@ bool PolicyRule::match_rangegrp_vecx(std::vector< range >& ranges, std::vector< 
     bool match = false;
     
     int idx = -1;
-    for(std::vector<baseHostCX*>::iterator i = vecx.begin(); i != vecx.end(); ++i ) {
+    for(auto cx: vecx) {
         ++idx;
-        baseHostCX* cx = (*i);
-        
-        match = match_rangegrp_cx(ranges,cx);
+
+        match = match_rangegrp_cx(ranges, cx);
         if(match) {
-            DIA_("PolicyRule::match_rangegrp_vecx: %s matched",cx->c_name());
+            _dia("PolicyRule::match_rangegrp_vecx: %s matched", cx->c_name());
             break;
         } else {
-            DIA_("PolicyRule::match_rangegrp_vecx: %s not matched",cx->c_name())
+            _deb("PolicyRule::match_rangegrp_vecx: %s not matched", cx->c_name());
         }
     }
     
@@ -206,16 +213,15 @@ bool PolicyRule::match_addrgrp_vecx(std::vector< AddressObject* >& sources, std:
     bool match = false;
     
     int idx = -1;
-    for(std::vector<baseHostCX*>::iterator i = vecx.begin(); i != vecx.end(); ++i ) {
+    for(auto cx: vecx) {
         ++idx;
-        baseHostCX* cx = (*i);
-        
+
         match = match_addrgrp_cx(sources,cx);
         if(match) {
-            DIA_("PolicyRule::match_addrgrp_vecx: %s matched",cx->c_name())
+            _dia("PolicyRule::match_addrgrp_vecx: %s matched",cx->c_name());
             break;
         } else {
-            DIA_("PolicyRule::match_addrgrp_vecx: %s not matched",cx->c_name())
+            _deb("PolicyRule::match_addrgrp_vecx: %s not matched",cx->c_name());
         }
     }
     
@@ -231,8 +237,7 @@ bool PolicyRule::match(baseProxy* p) {
     bool rpmatch = false;
     
     if(p != nullptr) {
-        DIAS_("PolicyRule::match");
-        
+
         lmatch = match_addrgrp_vecx(src,p->ls()) || match_addrgrp_vecx(src,p->lda());
         if(!lmatch) goto end;
 
@@ -244,20 +249,21 @@ bool PolicyRule::match(baseProxy* p) {
 
         rpmatch = match_rangegrp_vecx(dst_ports,p->rs()) || match_rangegrp_vecx(dst_ports,p->rda());
         if(!rpmatch) goto end;
-        
+
+        end:
+
+        if (lmatch && lpmatch && rmatch && rpmatch) {
+            _inf("PolicyRule::match %s OK", p->to_string(iINF).c_str());
+            cnt_matches++;
+
+            return true;
+
+        } else {
+            _dia("PolicyRule::match %s FAILED: %d:%d->%d:%d", p->to_string(iINF).c_str(), lmatch, lpmatch, rmatch, rpmatch);
+        }
+
     } else {
-        DIAS_("PolicyRule::match: p is nullptr");
-    }
-    
-    end:
-    
-    if (lmatch && lmatch && rmatch && rpmatch) {
-        DIAS_("PolicyRule::match ok");
-        cnt_matches++;
-        
-        return true;
-    } else {
-        DIA_("PolicyRule::match failed: %d:%d->%d:%d",lmatch,lpmatch,rmatch,rpmatch);
+        _err("PolicyRule::match: p is nullptr");
     }
 
     return false;
@@ -270,8 +276,17 @@ bool PolicyRule::match(std::vector<baseHostCX*>& l, std::vector<baseHostCX*>& r)
     bool rpmatch = false;
     
 
-    DIAS_("PolicyRule::match_lr");
-    
+    std::string ls("???");
+    std::string rs("???");
+
+    if(!l.empty()) {
+        ls = l[0]->to_string();
+    }
+    if(!r.empty()) {
+        rs = r[0]->to_string();
+    }
+
+
     lmatch = match_addrgrp_vecx(src,l);
     if(!lmatch) goto end;
 
@@ -285,27 +300,25 @@ bool PolicyRule::match(std::vector<baseHostCX*>& l, std::vector<baseHostCX*>& r)
     if(!rpmatch) goto end;
     
     
-    if(LEV >= DEB ) {
-        for(auto i: l) DUM_("PolicyRule::match_lr L: %s", i->to_string().c_str());
-        for(auto i: r) DUM_("PolicyRule::match_lr R: %s", i->to_string().c_str());
-        DEB_("PolicyRule::match_lr Success: %d:%d->%d:%d",lmatch,lpmatch,rmatch,rpmatch);
+    if(*log.level() >= DEB ) {
+        for(auto i: l) _dum("PolicyRule::match_lr L: %s", i->to_string().c_str());
+        for(auto i: r) _dum("PolicyRule::match_lr R: %s", i->to_string().c_str());
+        _deb("PolicyRule::match_lr Success: %d:%d->%d:%d", lmatch, lpmatch, rmatch, rpmatch);
     }
 
     end:
     
     
     if (lmatch && lpmatch && rmatch && rpmatch) {
-        DIAS_("PolicyRule::match_lr ok");
+        _inf("PolicyRule::match_lr %s <+> %s OK", ls.c_str(), rs.c_str());
         cnt_matches++;
         
         return true;
     } else {
-        DIA_("PolicyRule::match_lr failed: %d:%d->%d:%d",lmatch,lpmatch,rmatch,rpmatch);
+        _dia("PolicyRule::match_lr %s <+> %s FAILED: %d:%d->%d:%d", ls.c_str(), rs.c_str(), lmatch, lpmatch, rmatch, rpmatch);
     }
 
     return false;
 }
 
-PolicyRule::~PolicyRule() {
-}
 
