@@ -45,14 +45,14 @@
 
 #include <socle.hpp>
 
-#include <logger.hpp>
+#include <log/logger.hpp>
 #include <hostcx.hpp>
 #include <baseproxy.hpp>
 #include <masterproxy.hpp>
 #include <threadedacceptor.hpp>
 #include <display.hpp>
 
-#include <smithproxy.hpp>
+#include <main.hpp>
 #include <display.hpp>
 
 #include <libconfig.h++>
@@ -69,14 +69,20 @@
 #endif
 
 
-class SmithServerCX : public SmithProtoCX {
+class SmithServerCX : public SmithProtoCX, private LoganMate {
 public:
     SmithServerCX(baseCom* c, unsigned int s) : SmithProtoCX(c,s) {};
     SmithServerCX(baseCom* c, const char* h, const char* p) : SmithProtoCX(c,h,p) {};
-    virtual ~SmithServerCX() {};
-    
-    virtual void process_package(LTVEntry* e) {
-        DEB_("Package dump: \n%s",e->hr().c_str());
+    ~SmithServerCX() override = default;
+
+    logan_attached<SmithServerCX> log = logan_attached<SmithServerCX>(this, "com.smithd");
+    friend class logan_attached<SmithServerCX>;
+
+    std::string& class_name() const override {  static std::string s = "SmithServerCX"; return s; };
+    std::string hr() const override { return class_name(); }
+
+    void process_package(LTVEntry* e) override {
+        _deb("Package dump: \n%s",e->hr().c_str());
         
         
         LTVEntry* m = e->search({CL_REQTYPE});
@@ -84,7 +90,7 @@ public:
         
             LTVEntry* rating_rq = e->search({CL_PAYLOAD,RQT_RATEURL});
             if(rating_rq) {
-                INF_("found rating URI: %s",rating_rq->data_str().c_str());
+                _inf("found rating URI: %s",rating_rq->data_str().c_str());
                 
                 LTVEntry* resp = pkg_rating_response(111);
                 resp->pack();
@@ -94,21 +100,21 @@ public:
                 delete resp;
                 
             } else {
-                ERRS_("No payload or rating URL entry.");
+                _err("No payload or rating URL entry.");
                 LTVEntry* resp = pkg_error_response("nothing to rate");
                 resp->pack();
                 send(resp);
                 
                 delete resp;
             }
-            
+
         } else if (m && false) {
             /*another ID*/
         } else {
             if(m) {
-                ERR_("Unsupported request type %d", m->id());
+                _err("Unsupported request type %d", m->id());
             } else {
-                ERRS_("Request type unspecified, ignoring message.");
+                _err("Request type unspecified, ignoring message.");
             }
             
             LTVEntry* resp = pkg_error_response("unknown request");
@@ -119,32 +125,33 @@ public:
         }
     };
     
-    
-    LTVEntry* pkg_error_response(const char* text = nullptr) {
-        LTVEntry* e = new LTVEntry();
+    [[nodiscard]] // discarding leaks memory
+    static LTVEntry* pkg_error_response(const char* text = nullptr) {
+        auto* e = new LTVEntry();
         e->container(CL_INIT);
  
-        LTVEntry* err = new LTVEntry();
+        auto* err = new LTVEntry();
         err->set_num(255,LTVEntry::num,255);
         e->add(err);
         
         if(text != nullptr) {
-            LTVEntry* err2 = new LTVEntry();
+            auto* err2 = new LTVEntry();
             err2->set_str(254,LTVEntry::str,text);
             e->add(err2);
         }
         
         return e;
     }
-    
-    LTVEntry* pkg_rating_response(int category) {
-        LTVEntry* e = new LTVEntry();
+
+    [[nodiscard]] // discarding leaks memory
+    static LTVEntry* pkg_rating_response(int category) {
+        auto* e = new LTVEntry();
         e->container(CL_INIT);
         
-        LTVEntry* r = new LTVEntry();
+        auto* r = new LTVEntry();
         r->container(RQT_RATEURL);
         
-        LTVEntry* x = new LTVEntry();
+        auto* x = new LTVEntry();
         x->set_num(RATEURL_CAT,LTVEntry::num,category);
         r->add(x);
         
@@ -156,29 +163,38 @@ public:
 
 
 class SmithdProxy : public baseProxy {
+
+    logan_attached<SmithdProxy> log = logan_attached<SmithdProxy>(this, "com.smithd");
 public:
     SmithdProxy(baseCom* c) : baseProxy(c) {};
-    virtual ~SmithdProxy() {};
+    ~SmithdProxy() override = default;
 
-    virtual void on_left_error(baseHostCX*) {  dead(true); };
-    virtual void on_right_error(baseHostCX*) { dead(true); };    
+    std::string to_string(int verbosity=iINF) const override {
+        return "SmithdProxy|" + baseProxy::to_string(verbosity);
+    }
+
+    DECLARE_LOGGING(to_string);
+
+    void on_left_error(baseHostCX*)  override {  state().dead(true); };
+    void on_right_error(baseHostCX*) override { state().dead(true); };
     
-    virtual void on_left_bytes(baseHostCX* cx) {
-        INF_("Left %d bytes arrived to 0x%x",cx->readbuf()->size(), cx);
+    void on_left_bytes(baseHostCX* cx) override {
+        _inf("Left %d bytes arrived to 0x%x",cx->readbuf()->size(), cx);
     }
 };
-    
+
+DEFINE_LOGGING(SmithdProxy);
 
 class UxAcceptor : public ThreadedAcceptorProxy<SmithdProxy> {
 public:
     UxAcceptor(baseCom* c, int worker_id) : ThreadedAcceptorProxy<SmithdProxy>(c,worker_id) {};
     
-    virtual baseHostCX* new_cx(const char* h, const char* p) { return new SmithServerCX(com()->slave(),h,p); };
-    virtual baseHostCX* new_cx(int s) { return new SmithServerCX(com()->slave(), s); };
-    virtual void on_left_new(baseHostCX* cx) {
-        SmithdProxy* p = new SmithdProxy(com()->slave());
+    baseHostCX* new_cx(const char* h, const char* p) override { return new SmithServerCX(com()->slave(),h,p); };
+    baseHostCX* new_cx(int s) override { return new SmithServerCX(com()->slave(), s); };
+    void on_left_new(baseHostCX* cx) override {
+        auto* p = new SmithdProxy(com()->slave());
         p->ladd(cx);
-        this->proxies().insert(p);
+        this->proxies().push_back(p);
     }
 };
 
@@ -212,40 +228,63 @@ static loglevel  args_debug_flag = NON;
 bool config_file_check_only = false;
 
 
-void my_terminate (int param) {
-    
-    if (!cfg_daemonize)
-    printf("Terminating ...\n");
-    if (backend_proxy != nullptr) {
-        backend_proxy->dead(true);
+class SmithD {
+    logan_lite log_ = logan_lite("service");
+public:
+    SmithD() {}
+
+    static logan_lite& get_log() {
+        return instance().log_;
     }
 
-    cnt_terminate++;
-    if(cnt_terminate == 3) {
-        if (!cfg_daemonize)
-        printf("Failed to terminate gracefully. Next attempt will be enforced.\n");
+    static SmithD& instance() {
+        static SmithD d;
+        return d;
     }
-    if(cnt_terminate > 3) {
-        if (!cfg_daemonize)
-        printf("Enforced exit.\n");
-        abort();
-    }
-}
 
-void my_usr1 (int param) {
-    DIAS_("USR1 signal handler started");
-    NOTS_("reloading policies and its objects !!");
-    load_config(config_file,true);
-    DIAS_("USR1 signal handler finished");
-}
+    static void my_terminate (int param) {
+
+        auto& log = instance().get_log();
+
+        if (!cfg_daemonize)
+            printf("Terminating ...\n");
+        if (backend_proxy != nullptr) {
+            backend_proxy->state().dead(true);
+        }
+
+        cnt_terminate++;
+        if(cnt_terminate == 3) {
+            if (!cfg_daemonize)
+                printf("Failed to terminate gracefully. Next attempt will be enforced.\n");
+        }
+        if(cnt_terminate > 3) {
+            if (!cfg_daemonize)
+                printf("Enforced exit.\n");
+            abort();
+        }
+    }
+
+    static void my_usr1 (int param) {
+
+        auto& log = instance().get_log();
+
+        _dia("USR1 signal handler started");
+        _not("reloading policies and its objects !!");
+        load_config(config_file,true);
+        _dia("USR1 signal handler finished");
+    }
+
+};
+
+
 
 static struct option long_options[] =
     {
     /* These options set a flag. */
-    {"debug",   no_argument,       (int*)&args_debug_flag.level_, iDEB},
-    {"diagnose",   no_argument,       (int*)&args_debug_flag.level_, iDIA},
-    {"dump",   no_argument,       (int*)&args_debug_flag.level_, iDUM},
-    {"extreme",   no_argument,       (int*)&args_debug_flag.level_, iEXT},
+    {"debug",   no_argument,       (int*)&args_debug_flag.level_ref(), iDEB},
+    {"diagnose",   no_argument,       (int*)&args_debug_flag.level_ref(), iDIA},
+    {"dump",   no_argument,       (int*)&args_debug_flag.level_ref(), iDUM},
+    {"extreme",   no_argument,       (int*)&args_debug_flag.level_ref(), iEXT},
     
     {"config-file", required_argument, 0, 'c'},
     {"config-check-only",no_argument,0,'o'},
@@ -261,13 +300,17 @@ static struct option long_options[] =
 
 
 bool load_config(std::string& config_f, bool reload) {
+
     bool ret = true;
     std::lock_guard<std::recursive_mutex> l(merged_cfg_write_lock);
-    
+
+    auto this_daemon = DaemonFactory::instance();
+    auto& log = this_daemon.log;
+
     using namespace libconfig;
     Config cfgapi;
 
-    DIAS_("Reading config file");
+    _dia("Reading config file");
     
     // Read the file. If there is an error, report it and exit.
     try {
@@ -276,13 +319,13 @@ bool load_config(std::string& config_f, bool reload) {
         if (cfgapi.getRoot()["settings"].lookupValue("log_file",cfg_log_file)) {
             
             std::string log_target = cfg_log_file;
-            log_target = string_format(log_target,cfg_tenant_name.c_str());
+            log_target = string_format(log_target.c_str(),cfg_tenant_name.c_str());
             
             std::cout << "log target" << log_target << std::endl;
             
             // prepare custom crashlog file
             std::string crlog = log_target + ".crashlog.log";
-            set_crashlog(crlog.c_str());
+            this_daemon.set_crashlog(crlog.c_str());
             
             std::ofstream * o = new std::ofstream(log_target.c_str(),std::ios::app);
             get_logger()->targets(log_target,o);
@@ -300,33 +343,33 @@ bool load_config(std::string& config_f, bool reload) {
             }        
         }
 
-        cfgapi.getRoot()["settings"].lookupValue("log_level",cfg_log_level.level_);
+        cfgapi.getRoot()["settings"].lookupValue("log_level",cfg_log_level.level_ref());
     }
     catch(const FileIOException &fioex)
     {
-        ERR_("I/O error while reading config file: %s",config_f.c_str());
+        _err("I/O error while reading config file: %s",config_f.c_str());
         ret = false;   
     }
     catch(const ParseException &pex)
     {
-        ERR_("Parse error in %s at %s:%d - %s", config_f.c_str(), pex.getFile(), pex.getLine(), pex.getError());
+        _err("Parse error in %s at %s:%d - %s", config_f.c_str(), pex.getFile(), pex.getLine(), pex.getError());
         ret = false;
     }
 
     
     if(! ret ) {
-        FATS_("Unable to load config.");
+        _fat("Unable to load config.");
         return ret;
     }
     
     try {
-        cfgapi.getRoot()["settings"].lookupValue("log_level",cfg_log_level.level_);
+        cfgapi.getRoot()["settings"].lookupValue("log_level",cfg_log_level.level_ref());
         if(reload) {
         }
     }
     catch(const SettingNotFoundException &nfex) {
     
-        FAT_("Setting not found: %s",nfex.getPath());
+        _fat("Setting not found: %s",nfex.getPath());
         ret = false;
     }
     
@@ -334,7 +377,11 @@ bool load_config(std::string& config_f, bool reload) {
 }
 
 int smithd_apply_index(std::string& what , const std::string& idx) {
-    DEB_("apply_index: what=%s idx=%s",what.c_str(),idx.c_str());
+
+    auto this_daemon = DaemonFactory::instance();
+    auto& log = this_daemon.log;
+
+    _deb("apply_index: what=%s idx=%s",what.c_str(),idx.c_str());
     int port = std::stoi(what);
     int index = std::stoi(idx);
     what = std::to_string(port + index);
@@ -346,7 +393,7 @@ bool smithd_apply_tenant_config() {
     int ret = 0;
     
     if(cfg_tenant_index.size() > 0 && cfg_tenant_name.size() > 0) {
-        cfg_smithd_listen_port = string_format(cfg_smithd_listen_port,cfg_tenant_name.c_str());
+        cfg_smithd_listen_port = string_format(cfg_smithd_listen_port.c_str(),cfg_tenant_name.c_str());
     }
     
     
@@ -355,7 +402,10 @@ bool smithd_apply_tenant_config() {
 
 int main(int argc, char *argv[]) {
     
-    PID_FILE="/var/run/smithd.%s.pid";
+    DaemonFactory& this_daemon = DaemonFactory::instance();
+    auto& log = this_daemon.log;
+
+    this_daemon.PID_FILE="/var/run/smithd.%s.pid";
 
     config_file = "/etc/smithproxy/smithd.cfg";
     std::string config_file_tenant = "/etc/smithproxy/smithd.%s.cfg";
@@ -400,7 +450,7 @@ int main(int argc, char *argv[]) {
                 break;                
                 
             default:
-                ERR_("unknown option: '%c'",c);
+                std::cerr << "unknown option: " << c;
                 exit(1);                 
         }
     }
@@ -419,19 +469,19 @@ int main(int argc, char *argv[]) {
     std::cout << "tenant" << std::endl;
 
     if(cfg_tenant_index.size() > 0 && cfg_tenant_name.size() > 0) {
-        WAR_("Starting tenant: '%s', index %s",cfg_tenant_name.c_str(),cfg_tenant_index.c_str());
+        _war("Starting tenant: '%s', index %s",cfg_tenant_name.c_str(),cfg_tenant_index.c_str());
         std::cout << "tenant " << cfg_tenant_name.c_str() << "/" << cfg_tenant_index.c_str() << std::endl;
 
-        daemon_set_tenant("smithd",cfg_tenant_name);
+        this_daemon.set_tenant("smithd", cfg_tenant_name);
     } 
     else if (cfg_tenant_index.size() > 0 || cfg_tenant_name.size() > 0){
         
-        FATS_("You have to specify both options: --tenant-name AND --tenant-index");
+        _fat("You have to specify both options: --tenant-name AND --tenant-index");
         exit(-20);
     }
     else {
-        WARS_("Starting tenant: 0 (default)");
-        daemon_set_tenant("smithd","0"); 
+        _war("Starting tenant: 0 (default)");
+        this_daemon.set_tenant("smithd", "0");
     }
     
     
@@ -449,53 +499,49 @@ int main(int argc, char *argv[]) {
         
         struct stat s;
         if (stat(tenant_cfg.c_str(),&s) == 0) {
-            WAR_("Tenant config: %s",tenant_cfg.c_str());
+            _war("Tenant config: %s",tenant_cfg.c_str());
             config_file = tenant_cfg;
         } else {
-            WAR_("Tenant config %s not found. Using default.",tenant_cfg.c_str());
+            _war("Tenant config %s not found. Using default.",tenant_cfg.c_str());
         }
     }
     
-    WARS_(" ");
-    // set level to what's in the config
-    
+
     std::cout << "loading config from " << config_file << std::endl;
     
     if (!load_config(config_file)) {
         if(config_file_check_only) {
-            FATS_("Config check: error loading config file.");
+            _err("Config check: error loading config file.");
             exit(1);
         }
         else {
-            FATS_("Error loading config file on startup.");
+            _err("Error loading config file on startup.");
             std::cout << "error loading config" << std::endl;
             exit(1);
         }
     }
     
     if(!smithd_apply_tenant_config()) {
-        FATS_("Failed to apply tenant specific configuration!");
+        _fat("Failed to apply tenant specific configuration!");
         exit(2);
     }      
       
-      
-    WARS_(" ");
     // set level to what's in the config
     if (!load_config(config_file)) {
         if(custom_config_file); // make compiler happy
         
         if(config_file_check_only) {
-            FATS_("Config check: error loading config file.");
+            _fat("Config check: error loading config file.");
             exit(1);
         }
         else {
-            FATS_("Error loading config file on startup.");
+            _fat("Error loading config file on startup.");
             exit(1);
         }
     }
     
     if(config_file_check_only) {
-        DIAS_("Exiting, asked to check config file only.");
+        _dia("Exiting, asked to check config file only.");
         exit(0);
     }
 
@@ -506,24 +552,24 @@ int main(int argc, char *argv[]) {
     }
 #endif
   
-    if(daemon_exists_pidfile()) {
-        FATS_("There is PID file already in the system.");
-        FAT_("Please make sure smithd is not running, remove %s and try again.",PID_FILE.c_str());
+    if(this_daemon.exists_pidfile()) {
+        _fat("There is PID file already in the system.");
+        _fat("Please make sure smithd is not running, remove %s and try again.", this_daemon.PID_FILE.c_str());
         exit(-5);
     }
     
     if(cfg_daemonize) {
         if(get_logger()->targets().size() <= 0) {
-            FATS_("Cannot daemonize without logging to file.");
+            _fat("Cannot daemonize without logging to file.");
             exit(-5);
         }
         
         get_logger()->dup2_cout(false);
-        INFS_("entering daemon mode");
-        daemonize();
+        _inf("entering daemon mode");
+        this_daemon.daemonize();
     }
     // write out PID file
-    daemon_write_pidfile();
+    this_daemon.write_pidfile();
 
     //     atexit(__libc_freeres);    
     //     CRYPTO_malloc_debug_init();
@@ -534,31 +580,42 @@ int main(int argc, char *argv[]) {
 
     // no mercy here.
     unlink(cfg_smithd_listen_port.c_str());
-    backend_proxy = prepare_listener<UxProxy,UxCom>(cfg_smithd_listen_port,"ux-plain","/var/run/smithd.sock",cfg_smithd_workers);
-    
+    backend_proxy = nullptr;
+
+    try {
+        ServiceFactory::prepare_listener<UxProxy,UxCom>(cfg_smithd_listen_port,"ux-plain","/var/run/smithd.sock",cfg_smithd_workers);
+    } catch(socle::com_error const& e) {
+        _fat("Exception caught when creating listener: %s", e.what());
+        backend_proxy = nullptr;
+    }
+
     if(backend_proxy == nullptr && cfg_smithd_workers >= 0) {
         
-        FATS_("Failed to setup proxies. Bailing!");
+        _fat("Failed to setup proxies. Bailing!");
         exit(-1);
     }
 
-    set_daemon_signals(my_terminate,my_usr1);
+    this_daemon.set_daemon_signals(SmithD::my_terminate, SmithD::my_usr1);
     
     if(backend_proxy) {
-        INFS_("Starting smithd listener");
-        backend_thread = new std::thread([]() { 
-            set_daemon_signals(my_terminate,my_usr1);
-            DIA_("smithd: max file descriptors: %d",daemon_get_limit_fd());
+        _inf("Starting smithd listener");
+        backend_thread = new std::thread([]() {
+
+            auto this_daemon = DaemonFactory::instance();
+            auto& log = this_daemon.log;
+
+            this_daemon.set_daemon_signals(SmithD::my_terminate, SmithD::my_usr1);
+            _dia("smithd: max file descriptors: %d", this_daemon.get_limit_fd());
             
             backend_proxy->run(); 
-            DIAS_("smithd workers torn down."); 
+            _dia("smithd workers torn down.");
             backend_proxy->shutdown(); 
         } );
         pthread_setname_np(backend_thread->native_handle(),string_format("smithd_%s",cfg_tenant_index.c_str()).c_str());
     }
     
     
-    CRI_("Smithd backend (smithproxy %s, socle %s)  started",SMITH_VERSION,SOCLE_VERSION);
+    _cri("Smithd backend (smithproxy %s, socle %s)  started",SMITH_VERSION,SOCLE_VERSION);
     
     pthread_setname_np(pthread_self(),"smithd");
     
@@ -566,11 +623,10 @@ int main(int argc, char *argv[]) {
         backend_thread->join();
     }
 
-    if(backend_thread)
-        delete backend_thread;
+    delete backend_thread;
     
     // cleanup
-    daemon_unlink_pidfile();
+    this_daemon.unlink_pidfile();
     unlink(cfg_smithd_listen_port.c_str());
 }
 
