@@ -56,12 +56,17 @@ static std::string cfg_ux_socket = "/var/run/smithd.sock";
 
 
 
-class SmithClientCX : public SmithProtoCX {
+class SmithClientCX : public SmithProtoCX, private LoganMate {
 public:
     SmithClientCX(baseCom* c, unsigned int s) : SmithProtoCX(c,s) {};
     SmithClientCX(baseCom* c, const char* h, const char* p) : SmithProtoCX(c,h,p) {};
     virtual ~SmithClientCX() {};
-    
+
+    logan_attached<SmithClientCX> log = logan_attached<SmithClientCX>(this, "com.smithd");
+    friend class logan_attached<SmithClientCX>;
+
+    std::string& class_name() const override {  static std::string s = "SmithServerCX"; return s; };
+    std::string hr() const override { return class_name(); }
 
     //only used in test_url2
     sigslot::signal0<> sig_on_package;
@@ -70,13 +75,13 @@ public:
    
     virtual void process_package(LTVEntry* e) {
         
-	DEB_("Package dump: \n%s",e->hr().c_str());
+	_deb("Package dump: \n%s",e->hr().c_str());
         
         LTVEntry* m = e->search({1,1});
         if (m) {
-            DIA_("URL category: %d",m->data_int())
+            _dia("URL category: %d",m->data_int());
         } else {
-            ERR_("Unknown response: \n%s",e->hr().c_str());
+            _err("Unknown response: \n%s",e->hr().c_str());
         }
         
         //only used in test_url2
@@ -128,8 +133,8 @@ class SmithdProxy : public baseProxy {
         virtual baseHostCX* new_cx(const char* h, const char* p) { return new SmithClientCX(com(),h,p); };
         virtual baseHostCX* new_cx(int s) { return new SmithClientCX(com(), s); };
         
-        virtual void on_left_error(baseHostCX*) {  dead(true); };
-        virtual void on_right_error(baseHostCX*) { dead(true); };
+        virtual void on_left_error(baseHostCX*) {  state().dead(true); };
+        virtual void on_right_error(baseHostCX*) { state().dead(true); };
 };
 
 template <class COM, class CX, class PX>
@@ -151,8 +156,8 @@ class SimpleClient : public sigslot::has_slots<sigslot::multi_threaded_local> {
         inline CX* cx() { return cx_; }
         inline PX* px() { return px_; }
         
-        virtual int connect(bool b=false) { 
-            int r = cx()->connect(b); 
+        virtual int connect() {
+            int r = cx()->connect();
             if(r > 0) px()->ladd(cx()); 
             return r; 
         }
@@ -166,10 +171,12 @@ class SimpleClient : public sigslot::has_slots<sigslot::multi_threaded_local> {
 //only used in test_url2
 class PackageHandler :  public sigslot::has_slots<sigslot::multi_threaded_local> {
     public:
-        void on_package() { INFS_("MGR: package received (notify signal)"); };
+        logan_lite log = logan_lite("com.smithd");
+
+        void on_package() { _inf("MGR: package received (notify signal)"); };
         void on_package(SmithClientCX* cx, LTVEntry* pkg) { 
-            INFS_("MGR: package received (detail signal)");
-            DEB_("cx %s: data: \n%s",cx->c_name(), pkg->hr().c_str()); 
+            _inf("MGR: package received (detail signal)");
+            _deb("cx %s: data: \n%s",cx->c_name(), pkg->hr().c_str());
         };
 };
 
@@ -185,6 +192,7 @@ void test_url2(const char* url = nullptr) {
         // pack. Make buffer data from the object
         m->pack();
         client.cx()->send(m);
+        delete m; // coverity: 1408013
         
         // signal management
         PackageHandler mgr;
@@ -213,7 +221,7 @@ void test_url(const char* url = nullptr) {
     
     // since this is simple client, we will block. Normally it doesn't matter, run() would 
     // eventually take care of it
-    cx->connect(false);
+    cx->connect();
     
     // add cx to left side of proxy (there is no right side at all, we just need handle traffic
     // by proxy object.
@@ -228,7 +236,9 @@ void test_url(const char* url = nullptr) {
     
     // instruct cx to add data for sending
     cx->send(m);
-    
+
+    delete m; // coverity: 1407981
+
     // manage and run all connections until it's stopped. We stop in the client code,
     // where we receive response and close cx. Which will also close the proxy, which in turn terminates
     // run method.
@@ -238,9 +248,6 @@ void test_url(const char* url = nullptr) {
 
 int main(int argc, char *argv[]) {
     
-    get_logger()->dup2_cout(true);
-    get_logger()->level(INF);
-
     int loop_count = 1;
     
     // measure RTT for getting response
@@ -255,8 +262,16 @@ int main(int argc, char *argv[]) {
     
     for(int i = 0; i < loop_count; i++) {
     
-      ftime(&t_start);                                            
-      test_url2();
+      ftime(&t_start);
+
+      try {
+          test_url2();
+      } catch(socle::com_error const& e) {
+          std::cerr << "com error: " << e.what();
+          return 1;
+      }
+
+
       ftime(&t_current);
       
       t_diff = (int) (1000.0 * (t_current.time - t_start.time) + (t_current.millitm - t_start.millitm));
@@ -266,6 +281,7 @@ int main(int argc, char *argv[]) {
       if(t_diff < t_min || t_min == 0) t_min = t_diff;
       if(t_diff > t_max || t_max == 0) t_max = t_diff;
       
-      INF_(">> Server RTT: %dms  (avg=%.2fms min=%dms max=%dms)",t_diff,((float)t_sum)/t_cnt, t_min, t_max);
+      std::cout << string_format(">> Server RTT: %dms  (avg=%.2fms min=%dms max=%dms)",t_diff,
+              ((float)t_sum)/t_cnt, t_min, t_max);
     }
 }
