@@ -41,7 +41,7 @@
 
 #include <cfgapi.hpp>
 #include <sockshostcx.hpp>
-#include <logger.hpp>
+#include <log/logger.hpp>
 #include <dns.hpp>
 #include <smithdnsupd.hpp>
 
@@ -53,27 +53,27 @@ bool socksServerCX::global_async_dns = true;
 DEFINE_LOGGING(socksServerCX);
 
 socksServerCX::socksServerCX(baseCom* c, unsigned int s) : baseHostCX(c,s) {
-    state_ = INIT;
+    state_ = socks5_state::INIT;
 
     // copy setting from global/static variable - don't allow to change async
     // flag on the background during the object life
     async_dns = global_async_dns;
+    log = logan_attached<socksServerCX>(this, "com.socks");
 }
 
 socksServerCX::~socksServerCX() {
-
-    if(left)  { delete left; }
-    if(right) { delete right; }
+    delete left;
+    delete right;
 }
 
 
 int socksServerCX::process() {
     switch(state_) {
-        case INIT:
+        case socks5_state::INIT:
             return process_socks_hello();
-        case HELLO_SENT:
+        case socks5_state::HELLO_SENT:
             return 0; // we sent response to client hello, don't process anything
-        case WAIT_REQUEST:
+        case socks5_state::WAIT_REQUEST:
             return process_socks_request();
         default:
             break;
@@ -98,24 +98,24 @@ int socksServerCX::process_socks_hello() {
     
     // at this stage we have full client hello received
     if(version == 5) {
-        DIA_("socksServerCX::process_socks_init: version %d", version);
+        _dia("socksServerCX::process_socks_init: version %d", version);
         
         unsigned char server_hello[2];
         server_hello[0] = 5; // version
         server_hello[1] = 0; // no authentication
         
         writebuf()->append(server_hello,2);
-        state_ = HELLO_SENT;
-        state_ = WAIT_REQUEST;
+        state_ = socks5_state::HELLO_SENT;
+        state_ = socks5_state::WAIT_REQUEST;
         
         // flush all data, assuming 
         return b->size();
     }
     else if(version == 4) {
-        DIA_("socksServerCX::process_socks_init: version %d", version);
+        _dia("socksServerCX::process_socks_init: version %d", version);
         return process_socks_request();
     } else {
-        DIAS_("socksServerCX::process_socks_init: unsupported socks version");
+        _dia("socksServerCX::process_socks_init: unsupported socks version");
         error(true);
     }
 
@@ -127,11 +127,18 @@ bool socksServerCX::choose_server_ip(std::vector<std::string>& target_ips) {
     if(target_ips.empty())
         return false;
 
-    // for now use just first one (cleaned up from empty ones)
-    std::string t = target_ips.at(0);
+    uint64_t index = 0;
 
-    DIA_("choose_server_ip: chosen target: %s (out of %d)",t.c_str(),target_ips.size());
-    com()->nonlocal_dst_host() = t;
+    if(target_ips.size() > 1) {
+        //use some semi-random target
+        uint64_t baz = (uint64_t)this * (uint64_t)com() * time(nullptr);
+        index = baz % target_ips.size();
+    }
+
+    std::string target = target_ips.at(index);
+
+    _dia("choose_server_ip: chosen target: %s (index %d out of size %d)",target.c_str(), index, target_ips.size());
+    com()->nonlocal_dst_host() = target;
     com()->nonlocal_dst_resolved(true);
 
     return true;
@@ -146,17 +153,17 @@ bool socksServerCX::process_dns_response(DNS_Response* resp) {
     if (resp) {
         for (DNS_Answer &a: resp->answers()) {
             std::string a_ip = a.ip(false);
-            if (a_ip.size()) {
-                DIA_("process_dns_response: target candidate: %s", a_ip.c_str());
+            if (! a_ip.empty()) {
+                _dia("process_dns_response: target candidate: %s", a_ip.c_str());
                 target_ips.push_back(a_ip);
             }
         }
 
-        if (target_ips.size()) {
-            inspect_dns_cache.lock();
+        if (! target_ips.empty()) {
+            std::lock_guard<std::recursive_mutex> l_(DNS::get_dns_lock());
+
             DNS_Inspector di;
             del_resp = !di.store(resp);
-            inspect_dns_cache.unlock();
         }
 
         if (del_resp) {
@@ -166,10 +173,10 @@ bool socksServerCX::process_dns_response(DNS_Response* resp) {
 
     if(!target_ips.empty() && choose_server_ip(target_ips)) {
         setup_target();
-        DIAS_("process_dns_response: waiting for policy check");
+        _dia("process_dns_response: waiting for policy check");
 
     } else {
-        INFS___("process_dns_response: unable to find destination address for the request");
+        _dia("process_dns_response: unable to find destination address for the request");
         ret = false;
     }
 
@@ -178,27 +185,27 @@ bool socksServerCX::process_dns_response(DNS_Response* resp) {
 
 int socksServerCX::process_socks_request() {
 
-    socks5_request_error e = NONE;
+    socks5_request_error e = socks5_request_error::NONE;
     
     if(readbuf()->size() < 5) {
         return 0; // wait for more complete request
     }
-    
-    DIAS_("socksServerCX::process_socks_request");
 
-    if(state_ == DNS_QUERY_SENT) {
-        DIAS_("socksServerCX::process_socks_request: triggered when waiting for DNS response");
+    _dia("socksServerCX::process_socks_request");
+
+    if(state_ == socks5_state::DNS_QUERY_SENT) {
+        _dia("socksServerCX::process_socks_request: triggered when waiting for DNS response");
         return 0;
     }
 
-    DUM_("Request dump:\n%s",hex_dump(readbuf()->data(),readbuf()->size()).c_str());
+    _dum("Request dump:\n%s",hex_dump(readbuf()->data(),readbuf()->size()).c_str());
     
               version = readbuf()->get_at<unsigned char>(0);
     unsigned char cmd = readbuf()->get_at<unsigned char>(1);
     //@2 is reserved
     
     if(version < 4 or version > 5) {
-        e = UNSUPPORTED_VERSION;
+        e = socks5_request_error::UNSUPPORTED_VERSION;
         goto error;
     }
     
@@ -206,38 +213,38 @@ int socksServerCX::process_socks_request() {
         if(version == 5) {
             
             if(readbuf()->size() < 10) {
-                DIAS_("process_socks_request: socks5 request header too short");
+                _dia("process_socks_request: socks5 request header too short");
                 goto error;
             }
             
-            unsigned char atype   = readbuf()->get_at<unsigned char>(3);
+            socks5_atype atype   = static_cast<socks5_atype>(readbuf()->get_at<unsigned char>(3));
             
-            if(atype != IPV4 && atype != FQDN) {
-                e = UNSUPPORTED_ATYPE;
+            if(atype != socks5_atype::IPV4 && atype != socks5_atype::FQDN) {
+                e = socks5_request_error::UNSUPPORTED_ATYPE;
                 goto error;
             }
             
-            if(atype == FQDN) {
-                req_atype = FQDN;
-                state_ = REQ_RECEIVED;            
+            if(atype == socks5_atype::FQDN) {
+                req_atype = socks5_atype::FQDN;
+                state_ = socks5_state::REQ_RECEIVED;
                 
                 unsigned char fqdn_sz = readbuf()->get_at<unsigned char>(4);
                 if((unsigned int)fqdn_sz + 4 + 2 >= readbuf()->size()) {
-                    ERRS_("protocol error: request header out of boundary.");
+                    _err("protocol error: request header out of boundary.");
                     goto error;
                 }
-                
-                DIA_("socks5 protocol: fqdn size: %d",fqdn_sz);
+
+                _dia("socks5 protocol: fqdn size: %d",fqdn_sz);
                 std::string fqdn((const char*)&readbuf()->data()[5],fqdn_sz);
-                DIA_("socks5 protocol: fqdn requested: %s",fqdn.c_str());
+                _dia("socks5 protocol: fqdn requested: %s",fqdn.c_str());
                 req_str_addr = fqdn;
                 
                 req_port = ntohs(readbuf()->get_at<uint16_t>(5+fqdn_sz));
-                DIA_("socks5 protocol: port requested: %d",req_port);
+                _dia("socks5 protocol: port requested: %d",req_port);
 
                 com()->nonlocal_dst_port() = req_port;
                 com()->nonlocal_src(true);
-                DIA_("socksServerCX::process_socks_request: request (FQDN) for %s -> %s:%d",c_name(),com()->nonlocal_dst_host().c_str(),com()->nonlocal_dst_port());
+                _dia("socksServerCX::process_socks_request: request (FQDN) for %s -> %s:%d",c_name(),com()->nonlocal_dst_host().c_str(),com()->nonlocal_dst_port());
 
 
 
@@ -253,79 +260,79 @@ int socksServerCX::process_socks_request() {
                     target_ips.push_back(fqdn);
                 } else {
                     // really FQDN.
-                    
-                    inspect_dns_cache.lock();
-                    DNS_Response* dns_resp = inspect_dns_cache.get("A:"+fqdn);
+
+                    std::scoped_lock<std::recursive_mutex> l_(DNS::get_dns_lock());
+
+                    DNS_Response* dns_resp = DNS::get_dns_cache().get("A:"+fqdn);
                     if(dns_resp) {
-                        if (dns_resp->answers().size() > 0) {
+                        if ( ! dns_resp->answers().empty() ) {
                             int ttl = (dns_resp->loaded_at + dns_resp->answers().at(0).ttl_) - time(nullptr);                
                             if(ttl > 0) {
                                 for( DNS_Answer& a: dns_resp->answers() ) {
                                     std::string a_ip = a.ip(false);
-                                    if(a_ip.size()) {
-                                        DIA_("cache candidate: %s",a_ip.c_str());
+                                    if(! a_ip.empty() ) {
+                                        _dia("cache candidate: %s",a_ip.c_str());
                                         target_ips.push_back(a_ip);
                                     }
                                 }
                             }
                         }
                     }
-                    inspect_dns_cache.unlock();
                 }
 
 
                 // cache is not populated - send out query
-                if(target_ips.size() <= 0) {
+                if(target_ips.empty()) {
                     // no targets, send DNS query
                     
                     std::string nameserver = "8.8.8.8";
-                    if(cfgapi_obj_nameservers.size()) {
-                        nameserver = cfgapi_obj_nameservers.at(0);
+                    if(! CfgFactory::get().db_nameservers.empty()) {
+                        nameserver = CfgFactory::get().db_nameservers.at(0);
                     }
 
                     if(!async_dns) {
 
-                        DNS_Response *resp = resolve_dns_s(fqdn, A, nameserver);
+                        DNS_Response *resp = DNSFactory::get().resolve_dns_s(fqdn, A, nameserver);
 
                         process_dns_response(resp);
                         setup_target();
 
                     } else {
-                        int dns_sock = send_dns_request(fqdn, A, nameserver);
+                        int dns_sock = DNSFactory::get().send_dns_request(fqdn, A, nameserver);
                         if(dns_sock) {
-                            DIA___("dns request sent: %s", fqdn.c_str());
+                            _dia("dns request sent: %s", fqdn.c_str());
 
                             async_dns_socket.set(dns_sock, this, com());
                             async_dns_socket.opening();
 
                             com()->unset_monitor(socket());
 
-                            state_ = DNS_QUERY_SENT;
+                            state_ = socks5_state::DNS_QUERY_SENT;
                         } else {
-                            ERR___("failed to send dns request: %s", fqdn.c_str());
+                            _err("failed to send dns request: %s", fqdn.c_str());
                             error(true);
                         }
                     }
                 } else {
-                    if(!target_ips.empty() && choose_server_ip(target_ips)) {
+                    if(! target_ips.empty() && choose_server_ip(target_ips)) {
                         setup_target();
                     } else {
-                        INFS___("process_dns_response: unable to find destination address for the request");
+                        _err("process_dns_response: unable to find destination address for the request");
                         error(true);
                     }
                 }
             }
             else
-            if(atype == IPV4) {
-                req_atype = IPV4;
-                state_ = REQ_RECEIVED;
-                DIA_("socksServerCX::process_socks_request: request received, type %d", atype);
+            if(atype == socks5_atype::IPV4) {
+                req_atype = socks5_atype::IPV4;
+                state_ = socks5_state::REQ_RECEIVED;
+                _dia("socksServerCX::process_socks_request: request received, type %d", atype);
                 
                 uint32_t dst = readbuf()->get_at<uint32_t>(4);
                 req_port = ntohs(readbuf()->get_at<uint16_t>(8));
                 com()->nonlocal_dst_port() = req_port;
                 com()->nonlocal_src(true);
-                DIA_("socksServerCX::process_socks_request: request for %s -> %s:%d",c_name(),com()->nonlocal_dst_host().c_str(),com()->nonlocal_dst_port());
+                _dia("socksServerCX::process_socks_request: request for %s -> %s:%d",c_name(),com()->nonlocal_dst_host().c_str(),com()->nonlocal_dst_port());
 
                 
                 req_addr.s_addr=dst;
@@ -333,20 +340,20 @@ int socksServerCX::process_socks_request() {
                 com()->nonlocal_dst_host() = string_format("%s",inet_ntoa(req_addr));
                 com()->nonlocal_dst_port() = req_port;
                 com()->nonlocal_src(true);
-                DIA_("socksServerCX::process_socks_request: request (IPv4) for %s -> %s:%d",c_name(),com()->nonlocal_dst_host().c_str(),com()->nonlocal_dst_port());
+                _dia("socksServerCX::process_socks_request: request (IPv4) for %s -> %s:%d",c_name(),com()->nonlocal_dst_host().c_str(),com()->nonlocal_dst_port());
                 setup_target();
             }
         }
         
         else if (version == 4) {
             if(readbuf()->size() < 8) {
-                DIAS_("process_socks_request: socks4 request header too short");
+                _dia("process_socks_request: socks4 request header too short");
                 goto error;
             }            
             
-            req_atype = IPV4;
-            state_ = REQ_RECEIVED;
-            DIAS_("socksServerCX::process_socks_request: socks4 request received");
+            req_atype = socks5_atype::IPV4;
+            state_ = socks5_state::REQ_RECEIVED;
+            _dia("socksServerCX::process_socks_request: socks4 request received");
             
             req_port = ntohs(readbuf()->get_at<uint16_t>(2));
             uint32_t dst = readbuf()->get_at<uint32_t>(4);
@@ -357,12 +364,12 @@ int socksServerCX::process_socks_request() {
             com()->nonlocal_dst_host() = string_format("%s",inet_ntoa(req_addr));
             com()->nonlocal_dst_port() = req_port;
             com()->nonlocal_src(true);
-            DIA_("socksServerCX::process_socks_request: request (SOCKSv4) for %s -> %s:%d",c_name(),com()->nonlocal_dst_host().c_str(),com()->nonlocal_dst_port());
+            _dia("socksServerCX::process_socks_request: request (SOCKSv4) for %s -> %s:%d",c_name(),com()->nonlocal_dst_host().c_str(),com()->nonlocal_dst_port());
 
             setup_target();
         }
         else {
-            DIAS_("process_socks_request: unexpected socks version");
+            _dia("process_socks_request: unexpected socks version");
             goto error;
         }
         
@@ -370,7 +377,7 @@ int socksServerCX::process_socks_request() {
         return readbuf()->size();
 
     error:
-        DIA_("socksServerCX::process_socks_request: error %d",e);
+        _dia("socksServerCX::process_socks_request: error %d",e);
         error(true);
         return readbuf()->size();
 }
@@ -426,35 +433,32 @@ bool socksServerCX::setup_target() {
         
         
         left = n_cx;
-        DIA_("socksServerCX::setup_target: prepared left: %s",left->c_name());
+        _dia("socksServerCX::setup_target: prepared left: %s",left->c_name());
         right = target_cx;
-        DIA_("socksServerCX::setup_target: prepared right: %s",right->c_name());
+        _dia("socksServerCX::setup_target: prepared right: %s",right->c_name());
         
         // peers are now prepared for handover. Owning proxy will wipe this CX (it will be empty)
         // and if policy allows, left and right will be set (also in proxy owning this cx).
         
-        state_ = WAIT_POLICY;
+        state_ = socks5_state::WAIT_POLICY;
         read_waiting_for_peercom(true);
         
         return true;
 }
 
-bool socksServerCX::new_message() {
-    if(state_ == WAIT_POLICY && verdict_ == PENDING) {
+bool socksServerCX::new_message() const {
+    if(state_ == socks5_state::WAIT_POLICY && verdict_ == socks5_policy::PENDING) {
         return true;
     }
-    if(state_ == HANDOFF) {
-        return true;
-    }
-    
-    return false;
+    return state_ == socks5_state::HANDOFF;
+
 }
 
 void socksServerCX::verdict(socks5_policy p) {
     verdict_ = p;
-    state_ = POLICY_RECEIVED;
+    state_ = socks5_state::POLICY_RECEIVED;
     
-    if(verdict_ == ACCEPT || verdict_ == REJECT) {
+    if(verdict_ == socks5_policy::ACCEPT || verdict_ == socks5_policy::REJECT) {
         process_socks_reply();
     }
 }
@@ -466,17 +470,17 @@ int socksServerCX::process_socks_reply() {
         
         b[0] = 5;
         b[1] = 2; // denied
-        if(verdict_ == ACCEPT) b[1] = 0; //accept
+        if(verdict_ == socks5_policy::ACCEPT) b[1] = 0; //accept
         b[2] = 0;
-        b[3] = req_atype;
+        b[3] = static_cast<int>(req_atype);
         
         int cur = 4;
         
-        if(req_atype == IPV4) {
+        if(req_atype == socks5_atype::IPV4) {
             *((uint32_t*)&b[cur]) = req_addr.s_addr;
             cur += sizeof(uint32_t);
         }
-        else if(req_atype == FQDN) {
+        else if(req_atype == socks5_atype::FQDN) {
             
             b[cur] = (unsigned char)req_str_addr.size();
             cur++;
@@ -491,9 +495,9 @@ int socksServerCX::process_socks_reply() {
         cur += sizeof(uint16_t);
         
         writebuf()->append(b,cur);
-        state_ = REQRES_SENT;
-        
-        DUM_("socksServerCX::process_socks_reply: response dump:\n%s",hex_dump(b,cur).c_str());
+        state_ = socks5_state::REQRES_SENT;
+
+        _dum("socksServerCX::process_socks_reply: response dump:\n%s",hex_dump(b,cur).c_str());
 
         // response is about to be sent. In most cases client sends data on left,
         // but in case it's waiting ie. for banner, we must trigger proxy code to
@@ -516,13 +520,13 @@ int socksServerCX::process_socks_reply() {
         
         b[0] = 0;
         b[1] = 91; // denied
-        if(verdict_ == ACCEPT) b[1] = 90; //accept    
+        if(verdict_ == socks5_policy::ACCEPT) b[1] = 90; //accept
         
         *((uint16_t*)&b[2]) = htons(req_port);
         *((uint32_t*)&b[4]) = req_addr.s_addr;
         
         writebuf()->append(b,8);        
-        state_ = REQRES_SENT;
+        state_ = socks5_state::REQRES_SENT;
         return 8;
     }
     
@@ -530,12 +534,12 @@ int socksServerCX::process_socks_reply() {
 }
 
 void socksServerCX::pre_write() {
-    DEB_("socksServerCX::pre_write[%s]: writebuf=%d, readbuf=%d",c_name(),writebuf()->size(),readbuf()->size());
-    if(state_ == REQRES_SENT ) {
+    _deb("socksServerCX::pre_write[%s]: writebuf=%d, readbuf=%d",c_name(),writebuf()->size(),readbuf()->size());
+    if(state_ == socks5_state::REQRES_SENT ) {
         if(writebuf()->size() == 0) {
-            DIA_("socksServerCX::pre_write[%s]: all flushed, state change to HANDOFF: writebuf=%d, readbuf=%d",c_name(),writebuf()->size(),readbuf()->size());
+            _dia("socksServerCX::pre_write[%s]: all flushed, state change to HANDOFF: writebuf=%d, readbuf=%d",c_name(),writebuf()->size(),readbuf()->size());
             waiting_for_peercom(true);
-            state(HANDOFF);
+            state(socks5_state::HANDOFF);
         }
     }
 }
@@ -546,31 +550,28 @@ void socksServerCX::handle_event (baseCom *xcom) {
 
 
         if(com()->in_idleset(async_dns_socket.socket_)) {
-            INF___("handle_event: idling dns socket %d, closing", async_dns_socket.socket_);
+            _war("handle_event: idling dns socket %d, closing", async_dns_socket.socket_);
             error(true);
 
             return;
         }
 
         // timeout is zero - we won't wait
-        std::pair<DNS_Response *, int> rresp = recv_dns_response(async_dns_socket.socket_,0);
+        std::pair<DNS_Response *, int> rresp = DNSFactory::get().recv_dns_response(async_dns_socket.socket_,0);
         DNS_Response* resp = rresp.first;
         int red = rresp.second;
-        state_ = DNS_RESP_RECV;
+        state_ = socks5_state::DNS_RESP_RECV;
 
         if(red <= 0) {
-            INF___("handle_event: socket read returned %d",red);
+            _deb("handle_event: socket read returned %d",red);
             error(true);
-            if(resp) {
-                // unlikely I will get result on error, but one never knows
-                delete resp;
-            }
+            delete resp;
         } else {
-            INF___("handle_event: OK - socket read returned %d",red);
+            _deb("handle_event: OK - socket read returned %d",red);
             if(process_dns_response(resp)) {
-                INFS___("handle_event: OK, done");
+                _deb("handle_event: OK, done");
             } else {
-                ERRS___("handle_event: processing DNS response failed.");
+                _err("handle_event: processing DNS response failed.");
             }
         }
 
@@ -579,7 +580,7 @@ void socksServerCX::handle_event (baseCom *xcom) {
         async_dns_socket.closing();
 
     } else {
-        WARS___("handle_event: should not be here. Socket %d, async enabled: %d", async_dns_socket.socket_, async_dns);
+        _war("handle_event: should not be here. Socket %d, async enabled: %d", async_dns_socket.socket_, async_dns);
     }
 
 
