@@ -47,10 +47,10 @@
 #include <tcpcom.hpp>
 #include <sslmitmcom.hpp>
 
-typedef enum socks5_state_ { INIT=1, HELLO_SENT, WAIT_REQUEST, REQ_RECEIVED, WAIT_POLICY, POLICY_RECEIVED, REQRES_SENT, DNS_QUERY_SENT=15, DNS_RESP_RECV, HANDOFF=31 , ZOMBIE=255 } socks5_state;
-typedef enum socks5_request_error_ { NONE=0, UNSUPPORTED_VERSION, UNSUPPORTED_ATYPE } socks5_request_error;
-typedef enum socks5_atype_ { IPV4=1, FQDN=3, IPV6=4 } socks5_atype;
-typedef enum socks5_policy_ { PENDING, ACCEPT, REJECT } socks5_policy;
+typedef enum class socks5_state_ { INIT=1, HELLO_SENT, WAIT_REQUEST, REQ_RECEIVED, WAIT_POLICY, POLICY_RECEIVED, REQRES_SENT, DNS_QUERY_SENT=15, DNS_RESP_RECV, HANDOFF=31 , ZOMBIE=255 } socks5_state;
+typedef enum class socks5_request_error_ { NONE=0, UNSUPPORTED_VERSION, UNSUPPORTED_ATYPE } socks5_request_error;
+typedef enum class socks5_atype_ { IPV4=1, FQDN=3, IPV6=4 } socks5_atype;
+typedef enum class socks5_policy_ { PENDING, ACCEPT, REJECT } socks5_policy;
 
 typedef enum socks5_message_ { POLICY, UPGRADE } socks5_message;
 
@@ -72,7 +72,7 @@ public:
 
 class socksServerCX : public baseHostCX, public epoll_handler {
 public:
-    socksServerCX(baseCom* c, unsigned int s);  
+    socksServerCX(baseCom* c, unsigned int s);
     virtual ~socksServerCX();
 
 
@@ -83,11 +83,11 @@ public:
     virtual int process_socks_reply();
     virtual void pre_write();
     
-    virtual bool new_message();
+    bool new_message() const override;
     void verdict(socks5_policy);
     void state(socks5_state s) { state_ = s; };
     
-    socks5_policy verdict_ = PENDING;
+    socks5_policy verdict_ = socks5_policy::PENDING;
     socks5_state state_;
     
     //before handoff, prepare already new CX. 
@@ -118,6 +118,8 @@ private:
     // implement advanced logging
     DECLARE_C_NAME("sockHostCX");
     DECLARE_LOGGING(c_name);
+
+    logan_attached<socksServerCX> log;
 };
 
 #endif //_SOCKS5HOST_HPP_
