@@ -49,21 +49,39 @@
 class SocksProxy : public MitmProxy {
 public:
     explicit SocksProxy(baseCom*);
-    virtual ~SocksProxy();
-    virtual void on_left_message(baseHostCX* cx);
+    ~SocksProxy() override = default;
+    void on_left_message(baseHostCX* cx) override;
     
     virtual void socks5_handoff(socksServerCX* cx);
 
+    std::string to_string(int lev=iINF) const override { std::string r(string_format("SocksProxy[%s]", MitmProxy::to_string().c_str())); return r; };
+
+    DECLARE_C_NAME("SocksProxy");
+    DECLARE_LOGGING(to_string);
+
+private:
     time_t auth_table_refreshed = 0;
+
+    logan_attached<SocksProxy> log;
 };
 
 
 class MitmSocksProxy : public ThreadedAcceptorProxy<SocksProxy> {
 public:
 
-    MitmSocksProxy(baseCom* c, int worker_id) : ThreadedAcceptorProxy<SocksProxy>(c,worker_id) {};
-    virtual baseHostCX* new_cx(int s);
-    virtual void on_left_new(baseHostCX* just_accepted_cx);
+    MitmSocksProxy(baseCom* c, int worker_id) : ThreadedAcceptorProxy<SocksProxy>(c,worker_id) {
+        log = logan::attach(this, "masterproxy.socks");
+    };
+    baseHostCX* new_cx(int s) override;
+    void on_left_new(baseHostCX* just_accepted_cx) override;
+
+    std::string to_string(int lev=iINF) const override { static std::string r(string_format("MitmSocksProxy[%s]", baseProxy::to_string().c_str())); return r; };
+
+    DECLARE_C_NAME("MitmSocksProxy");
+    DECLARE_LOGGING(to_string);
+
+private:
+    logan_attached<MitmSocksProxy> log;
 };
 
 
